Revert to the old convention, where timecodes represent the start or each video frame and audio chunk.
This changed in CVS about a week ago, and it broke older recordings. Now it's changing back, which will break recordings made with the CVS version from the past.

The motivation behind reverting is to avoid a lot of complaints when 0.8 comes out. Now, it won't break recordings made under 0.7


git-svn-id: http://svn.mythtv.org/svn/trunk@442 7dbf422c-18fa-0310-86e9-fd20926502f2