Renchi Raju's new list type patch. av's playbackbox + epg rendering patches. Chris Morgan's playbackbox state machine patch.
Clean up the needevents stuff in RemoteFile, not necessary anymore.
Make the readaheadthread optional, and disable it in the playbackbox for more
of a speedup.
Try to disable dialogs when they're not on top.

Note: The 'watch recordings' screen needs beaten on by people to make sure everything's still working properly with the preview video playing.


git-svn-id: http://svn.mythtv.org/svn/trunk@2778 7dbf422c-18fa-0310-86e9-fd20926502f2