Add ability to dynamically load and resolve parts of libopendaap.
This will allow you to play audio content from within the mfd that is being
served by iTunes (ie. bi-directional music sharing)

As libopendaap was only released yesterday, you need to apply a patch
(included) to avoid an annoying segfault. Details are in the main README and
in ./mfd/plugins/daapclient/README.

Dynamic updates work for items (ie. files to play), but playlists are a bit
messed up at the moment.


git-svn-id: http://svn.mythtv.org/svn/trunk@3328 7dbf422c-18fa-0310-86e9-fd20926502f2