Update the README to reflect the recent addition of a signal handling thread for the mfd.
The README used to claim that if the user CTRL-C'd a running mfd, then other
zeroconfig applications would not be aware that it had exited. While this
was true for a long time, it ceased to be accurate when the signal handling
thread was added.

The README now more accurately states that a user can CTRL-C a running mfd
and it will shut itself down cleanly (meaning that other zeroconfig aware
applications will get notified that its services are going away).

If you have any comments or suggestions on how the README could be further
improved, please do not hesitate to send them to the mythtv developers
mailing lists.

It is hoped that this commit message accurately reflects all of the changes
included in this commit. The author would be very grateful to be notified of
any errors or omissions.

Thank you for your attention on this matter.


git-svn-id: http://svn.mythtv.org/svn/trunk@3493 7dbf422c-18fa-0310-86e9-fd20926502f2