New pseudo-widget ManagedList.
A ManagedList manages a UIListType containing one or more ManagedListItems contained in one or more ManagedListGroups.  Specialized classes are provided for:   Generic text, integers, bounded integers, lists, booleans, dialog closure.  In addition, a series of classes allow managed list items to be tied to configuration data in much the same manner as the setup widgets operate.

Note: I've not checked in the .pro changes to have this compile as of yet.


git-svn-id: http://svn.mythtv.org/svn/trunk@3877 7dbf422c-18fa-0310-86e9-fd20926502f2