Theme designers take note:  This update will prevent  users of your theme from acessing the advnaced recording options screen until you release an update.  The CVS versions of GANT and Blue have been updated as has Titivillus (at: http://www.jdonavan.net/myth)
With this commit the new "advanced" recording option screen will not be compiled and used.  This screen is themeable, and provides a friendlier face on the advanced recordings screen that the old Qt based one.

Modifed the grouping in the root group based on feedback from testers.
Modified list count, font sizes and element postioning in GANT and Blue based on feedback from testers.
List items that lead to other windows should now all be encased in brackets instead of using elipsises. (elipsi?)


git-svn-id: http://svn.mythtv.org/svn/trunk@3907 7dbf422c-18fa-0310-86e9-fd20926502f2