Patch from David H��rdeman to allow meta data display for files that have it when in file mode on the video tree.
Patch from Leo Weppelman (broken out from a patch by Xavier) that allows removeable disks to be displayed in the video tree.  The MediaMonitor needs a way to eject data CDs for this to be really useful IMO, since once you stick one in, it's in for good on my system...  I've got a ton of crap on my plate at the moment so if someone want to look at this that'd be great.


git-svn-id: http://svn.mythtv.org/svn/trunk@4206 7dbf422c-18fa-0310-86e9-fd20926502f2