Change new episodes only recording from a dup method to a dup in.  This allows you to match duplicates however you wish.  If a recording is NOT a repeat the scheduler will look in both current and old recordings for it.  This is how I originaly intended for this thing to work I just went about it the wrong way the frist time around.
Make opening a local ringbuffer a bit more fault tolerant.  When ever we open a ring buffer now we do  a quick check to see if the file handle can be read from.  If it can't we usleep a bit and try again, up to 5 times (currently).  This should have zero impact on folks who were not having problems transitioning from watching livetv to watching a recording.  For those that are having problems, this should take care of it, though you will have a noticable delay when transitioning before the system gets a good file handle.

I'm still trying to track down the root cause of this problem.  It's something specfic to certain machine configurations, though I'm not clear if it's filesystem, LVM, distro or kernel related (or possibly some combination of them all).


git-svn-id: http://svn.mythtv.org/svn/trunk@4450 7dbf422c-18fa-0310-86e9-fd20926502f2