Start of some useful changes to libmythui.
    Please note, a lot of these files are place holders for things I'm
actively working on. The main reason for the commit was to get this in so
that others will have some idea where I'm trying to get to (also, I need to
move the code tree to another box so that I can actually get some work done
on it during this week).

    In the midst of redesigning a tree object (MythUITree) which can get
passed a tree of content/menu items (will be MythTreeNode's) and be very
flexible about how it displays them (vertical view, horizontal view, folder
view, etc.). The MythTreeNode will replace GenericTree with something that
holds a void* pointer to the *thing* (ie. music track to play, video to
view). Also, inheriting MythTreeNode's is to be discouraged, as we want to
get away from having to dynamic_cast returned objects every time the user
navigates or selects anything. Instead, the base MythTreeNode will have an
extensive type enum, which, together with the void* pointer, will let us
just query an int value and then call an appropriate get() (ie.
getMusicMetadata(), getVideoMetadata(), etc.).

     Main stuff that in this commit that is working is a better xml parser
that handles debugging well. Oh and all MythUIType's can now draw a frame
around themselves to show where their boundaries are when Draw() is called.


git-svn-id: http://svn.mythtv.org/svn/trunk@4804 7dbf422c-18fa-0310-86e9-fd20926502f2