Title sorting for Watch/Delete Recordings, Set Priorities, and Previously Recorded pages. In English, titles should be sorted alphabetically excluding the prefix words The, A and An. These and the Program Finder are the only pages currently which display sorted lists of titles. Program Finder is already hard coded to treat "The" and "A" as exceptions for English only and would require much more work to convert.
While each of these pages have different code to suit different needs,
they each create hidden modified copies of the titles for sorting by
using the regular expression tr("^(The |A |An )") . Notice the
"tr(" which means that any person doing translations, programmer
or not, may translate this string ( "^(Das |de |la |a |Da )") for
whichever or how many ever prefix words are excluded when sorting
titles in their language. Translating this sting will then affect
the title sorting for each of these three pages. If a language
does not exclude any prefix words for sorting, simply translate
this string to be "".


git-svn-id: http://svn.mythtv.org/svn/trunk@4817 7dbf422c-18fa-0310-86e9-fd20926502f2