Tweak OSDListTreeType, OSDGenericTree, and OSDListBtnTypeItem to allow toggling of the checkbox on the OSD popup menu.
Add checkboxes to the OSD menu on the Commercial Auto-Skip, Aspect Ratio,
and Timestretch submenu items.  This allows the user to see what the
current value is set to for each of these items when they go into these
submenus.  The Auto-Skip and Timestretch OSD menus still close when an
item is selected, and the Aspect Ratio menu stays open.  When the user
picks a new aspect ratio, the checkbox for the old aspect is turned OFF
and the checkbox for the newly selected aspect is turned ON.


git-svn-id: http://svn.mythtv.org/svn/trunk@5502 7dbf422c-18fa-0310-86e9-fd20926502f2