The last of the files.
With checkin the new theme "Minimalist-wide" is now usable.  This theme takes advantage of the new wide screen mode to put more information on the screen.

It's not 100% complete, it needs tweaking and there's likely a window or two I missed.

Screenshots are available at http://www.jdonavan.net/myth/wide_mode


git-svn-id: http://svn.mythtv.org/svn/trunk@5526 7dbf422c-18fa-0310-86e9-fd20926502f2