Changed the keybindings code to print a warning, but still allow, multiple actions to be bound to the same key in a context.  The first action handled by the code wins and all others are ignored.
Added new, explicit actions for changing the rank in the program and
channelrecpriority screens.  By default, RANKINC and RANKDEC are bound
to the Right and Left keys, respectively.

Changed viewscheduled and program/channelrecpriority to exit on the
LEFT action like many other screens.  FYI, you won't see this behavior
on the *recpriority screens unless RANKDEC is bound to a different key
than LEFT.


git-svn-id: http://svn.mythtv.org/svn/trunk@5896 7dbf422c-18fa-0310-86e9-fd20926502f2