This change looks bigger than it is. It changes the API on DisplayRes a little bit. It adds a GetVideoModes() call which lists all the valid resolutions and refresh rates. In addition, the switch resolution methods allow a refresh rate. The GetVideoModes() method allows us to use a combo box when specifying output resolutions, reducing user error.
Also, GetDisplayRes() returns the DisplayRes singleton even if
the UseVideoModes setting is disabled. So the database needs to
be queried seperately for that setting.

Also, the new GetAspectRatio() method returns the true aspect ratio,
unless the user has a aspect overide set. The overide setting for
this is now the desired aspect ratio, not "Alternate Aspect Ratio",
allowing aspect ratios other than 4:3 and 16:9. Unfortunately any
existing overides are ignored and must be re-configured.

The public class DisplayResScreen replaces the private dim_t class,
adding refresh rate information.

util-x11.cpp contains GetNumberOfXineramaScreens() which returns
the number of XineramaScreens, or 0 if Xinerama is not supported.
This was in util.cpp, but Nigel suggested moving it because he
is removing OS specific things from util.cpp. Nigel also deserves
credit for getting DisplayResOSX working.

John Poet is still reporting a problem with resizing to HDTV sizes,
but his problem exists in this patch and current CVS. The resizing
to HDTV sizes in this patch works for me...


git-svn-id: http://svn.mythtv.org/svn/trunk@5947 7dbf422c-18fa-0310-86e9-fd20926502f2