This allows XvMC -> XVideo fallback to function when XvMC is present and mostly working, but still setup incorrectly. In particular with the release of Xorg 6.8.2, people are running into the problem of having /etc/X11/XvMCConfig pointing to the wrong XvMC library. This results in some XvMC calls succeeding while other calls fail.
This patch extends VOXv::GetBestSupportedCodec() so that it
actually opens a port and creates a XvMCContext. This context
creation fails if /etc/X11/XvMCConfig is set incorrectly. If
XvMC is detected but context creation fails, an error message
tells any interested user to check their DISPLAY variable.
If they are using libXvMCW, the message tells them to check
their /etc/X11/XvMCConfig file.


git-svn-id: http://svn.mythtv.org/svn/trunk@6257 7dbf422c-18fa-0310-86e9-fd20926502f2