Added CODEC_ID_MPEG2VIDEO_XVMC and CODEC_ID_MPEG2VIDEO_XVMC_VLD to the stream types parsed by the mpegvideo_parser in libs/libavcodec/parser.c. Also added a warning message if av_read_frame_internal() in libs/libavformat/utils.c can not find a parser for the current video.
These were parts of Terry Barnaby's XvMC-VLD channel change patch for MPEG-TS
streams. I did not add the unclean program exit on unrecognized stream in utils.c,
nor the changes the SIP parser. The SIP parser changes need to be reviewed by
someone familiar with DVB. It seems to be a hack to allow video and audio to
be decoded from the same type id (0x06), if there is one video stream and it
is listed before the audiostreams.


git-svn-id: http://svn.mythtv.org/svn/trunk@6260 7dbf422c-18fa-0310-86e9-fd20926502f2