This removes all but two exit's within libmyth and libmythtv. Please hold off on compiling cvs for the next 30 minutes, unless you really want to recompile everything using mythcontext.h twice... I have 3 updates coming on the heals of this one: one to update mythbrowser and mythdvd/mtd to check the result of gContext->Init(), one to remove the exit() in ThreadedFileWriter, and another to remove some exit() calls in the various myth programs.
This ads a centralized location for exit codes, exitcodes.h.
This is useful when we are examining the exit code, such as
when we do commercial flagging. For example, some error codes
get mistakenly considered as correct exits from commercial
flagging. A fix for this will be in the fourth exit() cleanup
commit.


git-svn-id: http://svn.mythtv.org/svn/trunk@6281 7dbf422c-18fa-0310-86e9-fd20926502f2