It appears nvidia's XvMC bob-deint compensation does try to adjust for overscan, but underestimates the amount of compensation required. I could not figure out the formula they are using, so I can't compensate exactly. But by only applying overscan compensation when the overscan is greater than or equal to 5% it appears that the bob-deint compensation error never gets too big.
Underscan works fine, the nVidia's bob-deint compensation is only wrong
when frame_height != imgh. And it appears that the newer drivers allow
a negative offset so we could actually get this right if we could detect
these drivers and not scale the source image in overscan (which is needed
by other hardware).


git-svn-id: http://svn.mythtv.org/svn/trunk@6324 7dbf422c-18fa-0310-86e9-fd20926502f2