Added simple detection rutines for opengl, fftw, sdl, aac, transcode, and exif options. By default I'm not using aac nor trancode even if detected, because I think we need to check for more than we just the presence of the libraries. I also don't know how to write detection rutines for vcdsupport and festival; maybe someone else will have tackle those...
BTW I discovered that the reason fftw isn't working for me is because we
only support the 2.x libraries, and not the current 3.x libraries. I've
added a detection rutine in ./configure for the 3.x libraries, but they
are, of course, disabled until someone has time to look at porting
MythMusic to the new libraries.


git-svn-id: http://svn.mythtv.org/svn/trunk@6375 7dbf422c-18fa-0310-86e9-fd20926502f2