Fixes printf that used %ld for ints, this causes problems on 64 bit systems.
Also replaced some printf's and cerr redirects with verbose macros,
and replaced the commercial flagging output ones with two sets of
macros print_v0 and print_v1. print_v0 is for non-error messages
that are always printed, and print_v1 is for non-error messages
that are printed only in verbose commercial flagging. I didn't
convert to VERBOSE macros on these because there are a lot of
C style formatting strings..


git-svn-id: http://svn.mythtv.org/svn/trunk@6529 7dbf422c-18fa-0310-86e9-fd20926502f2