ChannelBase cleanup.
This re-orders channelbase.h so it is easier to read, and documents
most of the methods. It also adds a SetChannelByDirection() method
which reduces the redundancy of the code in ChannelUp(), ChannelDown(),
and NextFavorite(). I've also added default implementations for SetFd()
and GetFd(), instead of leaving them abstract. The implementation of
GetFd() in FirewireChannel returned 0 which is a valid file descriptor
but completely bogus.

I've also replaced map with QMap in ChannelBase(), which along with
SetChannelByDirection() is why tv_rec.{cpp,h} has changes.

I've also made large method parameters that weren't changed in the
channel classes to 'const' references instead of copies, which in
the case of DVBChannel::SetPMT(PMTObject) trickled up through a few
of the dvb classes.

In channel.{h,cpp} I also found a bug in testing videofd, the code
assumed zero is not a valid file descriptor, but of course it is.
I fixed the tests and also got rid of the 'isopen' field, which was
redundant since we had videofd to test.


git-svn-id: http://svn.mythtv.org/svn/trunk@6590 7dbf422c-18fa-0310-86e9-fd20926502f2