Attempt to solve MacOS X linking problems. Someone with a OS X should look at this, but I've looked at the code and made guesses. It looks like signalmonitorvalue.cpp should be compiled in a frontend only compile, while pchdtvsignalmonitor.cpp should be in any backend only compile.
However scanwizard & friends are currently in a frontend only compile,
if this is desired then it pulls in a bunch of backend stuff, which should
be added to the frontend compile. If it isn't desired, then they should
be moved to the backend only compile and mythtv-setup should not be linked.


git-svn-id: http://svn.mythtv.org/svn/trunk@6666 7dbf422c-18fa-0310-86e9-fd20926502f2