This makes the changes needed to the MythTV Protocol needed for the signal monitoring patch, without actually adding signal monitoring.
This means MYTH_PROTO_VERSION is changed to "18", requiring upgrades
of all backend and frontends. The good news is that the version should
not need to be changed for the signal monitoring patch.

For the protocol this
* Renames QUERY_RECORDER PAUSE to QUERY_RECORDER PAUSE_RECORDER
  corresporting to PauseRecorder() in RemoteEncoder, EncoderLink, and TVRec.
* Adds symetric QUERY_RECORDER PAUSE, and QUERY_RECORDER UNPAUSE are added
  correspording Pause()/Unpause() in RemoteEncoder, EncoderLink, and TVRec.
* Adds QUERY_RECORDER SET_SIGNAL_MONITORING_RATE
* Adds the requirement to TVRec to reply to a
    SET_SIGNAL_MONITORING_RATE rate>0 notify>0
  at least until it sends a signal lock message or gets another message
  with rate<=0 or notify<=0.

In TVRec, this removes the Pause()/Unpause() calls in ToggleInputs(),
ChangeChannel(), and SetChannel(). The protocol change here is that
Pause and Unpause must be issue by whoever is using the recorder before
and after these commands.

In TV (tv_play), this commit adds PauseLiveTV()/UnpauseLiveTV() which 
perform all the work of pausing and unpausing everything in ToggleInputs(),
ChangeChannel(), and SetChannelByString().

I've also added kState_WaitingForLiveTV to tv.h and a dummy 
TVRec::SetSignalMonitoringRate(), these just to try to avoid
needing to update the MythTV Protocol for the signal monitor
patch. The dummy SetSignalMonitoringRate() returns a 
"signal locked" reply to the frontend to any request, which 
is basically what an analog backend does with the signal 
monitoring patch. This means you could use a frontend with
the signal monitoring patch while the backend does not have
the patch and MythTV would behave just like not having the
signal monitoring patch.



git-svn-id: http://svn.mythtv.org/svn/trunk@6739 7dbf422c-18fa-0310-86e9-fd20926502f2