Added a "Premieres" view for the New Titles search list (Home and End to change views). This is a short list for only the specials and series first episodes when they are shown on their original air date.
Added Custom Record example "First Episodes" which matches episode 1 of
any series when it is shown on the original air date. This results
in one or two dozen matches over two weeks. This can be used as a rule
to schedule any series premiere then "Don't record" for the things that
are not interesting. Better yet, set the rule to inactive so that all
series premieres are included on the Upcoming Recordings list in white
with status "x". You can then "Record anyway" for something of interest.


git-svn-id: http://svn.mythtv.org/svn/trunk@6793 7dbf422c-18fa-0310-86e9-fd20926502f2