This adds a "QUERY_RECORDER SET_SIGNAL_MONITORING_RATE" command to the MythTV Protocol for the backend, and adds a SIGNAL MythTV Event handler to tv_play. I have not incremented the Myth protocol version number because these commands are only used for providing information to the user and not using frontends and backend not supporting them fails gracefully when used with ones that do support them.
This commit is the first of several. I've made extensive 
modifications to TVRec and want to ease them in so each
can be tested individually.

The goal is to support signal monitoring without slowing
down tuning, so the "QUERY_RECORDER SET_SIGNAL_MONITORING"
command will only be used to toggle signal monitoring
once you are already watching a recording. Normally 
signal monitoring will be handled entirely in TVRec with
a dummy MPEG stream being sent in place of the video when
the recorder is paused, so that the TV class can present
an OSD and usable interface to the user.



git-svn-id: http://svn.mythtv.org/svn/trunk@6847 7dbf422c-18fa-0310-86e9-fd20926502f2