This is another channel scanning commit.
The main change in this revision is that the SIScan constructor 
now takes a card type, a ChannelBase class, and source id instead
of a DVBChannel class and source id. This is obviously so that
SIScan can use a Channel class in addition to a DVBChannel class.

The second change is to move the signalmonitor instance from 
scanwizard to siscan, this makes the new SIScan much cleaner
since it uses the signalmonitor for much more intensively than
the old SIScan. In this revision though, it doesn't make much
of a difference, it just makes the new scanwizard and old
scanwizard look more alike.

There are also various minor changes mostly in siscan to make the
sync with the new scanner easier. Most of the now remaining 
differences between the new and old scanner are contained in
siscan.



git-svn-id: http://svn.mythtv.org/svn/trunk@6889 7dbf422c-18fa-0310-86e9-fd20926502f2