Another scanner commit.
This breaks out the transport scanning code from HandleActiveScan()
into a new method ScanTransport().

Also adds 4 slots from the new scanner:
 HandleVCT(), HandleMGT(), HandleSDT(), HandleNIT()
And adds 2 methods from the new scanner:
 HandleATSCDBInsertion(), HandleDVBDBInsertion()
And stubs for 2 methods from the new scanner:
 UpdateVCTinDB(), UpdateSDTinDB()

With all this it should still be code neutral, it just 
moves things around and adds things needed for scanning
with the ScanStreamData class.



git-svn-id: http://svn.mythtv.org/svn/trunk@6919 7dbf422c-18fa-0310-86e9-fd20926502f2