Another Channel Scanner commit
This fixes a problem, discovered by John Pullan, where the first
transport scanned would not have it's DVB/ATSC tables found. It
happened because the timer for the first channel was started
before we tuned to the first channel, so that that the first 
transport did not get to wait as long for tables as the other
transports.



git-svn-id: http://svn.mythtv.org/svn/trunk@6925 7dbf422c-18fa-0310-86e9-fd20926502f2