This addresses #145.
I discovered that the problem is that something in the libav or
avformatdecoder overwrites the avformatdecoder class with audio
data. I haven't figured out where this is happening, but by 
allocating the audio buffer on the heap, instead of as an
array in avformatdecoder, the SEGFAULT doesn't happen.

However the video has a few lines of bright green on the top of video.
And those blocks move about a bit on the P & B frames. It's almost as
if a slice was missing. The sound is working, so I'm not sure what is
going on.


This commit also cleans up variable initialization in the involved
classes and provides a couple Set's & Get's for the NVP and 
RecorderBase pointers in RecorderBase and NVP, respectively.



git-svn-id: http://svn.mythtv.org/svn/trunk@6959 7dbf422c-18fa-0310-86e9-fd20926502f2