Redo of changeset [7018].
Basically the difference is that DVBRecorder::SetDemuxFilters()
no longer puts the DVBRecorder in an error state if the PID's
are not set. The problem was that since DVBChannel no longer 
blocks until a good signal is seen, the recorder was more likely
to call SetDemuxFilters() before the recorder's PMT was set.



git-svn-id: http://svn.mythtv.org/svn/trunk@7027 7dbf422c-18fa-0310-86e9-fd20926502f2