Another merge from the Signal Monitor patch.
This removes the recorder starting code from HandleStateChange()
and moves it to StartRecorder() and StartRecorderPost(). The
reason for two methods is that in the signal monitor patch 
StartRecorder() is run before the frontend begins playing the
stream, and StartRecorderPost() run after the frontend begins
playing the stream. But in this commit StartRecorder() just calls
StartRecorderPost() giving us the same functionality as the
code in SVN before this commit.

There is a change in how zero length recordings are handled,
before an empty recording would be added to the recorded list,
but now these should not be added to the list so that MythTV
can attempt to record them later.



git-svn-id: http://svn.mythtv.org/svn/trunk@7030 7dbf422c-18fa-0310-86e9-fd20926502f2