This commit closes #158 by applying a slightly tweaked version of the mpeg2trans_v2.patch contained within.
It reimplements the MPEG2trans class in mythtranscode. The output stream
is DVD compatible if the resolution etc of the original meets the DVD
spec.

In my testing some of the cuts were slightly off, but that might be due
to not having cut on proper keyframe boundries...  Most importantly,
there were no A/V sync issues after having performed a transcode with
the -l parameter.

To use the MPEG2 transcoding you will need to add an entry to the
recordingprofiles table. Something like

update recordingprofiles set videocodec = "MPEG-2" where name = "MPEG2"; 

You can also use "--mpeg2" on the command line for mythtranscode.

Also with this commit, mythtranscode will attempt to retreive a
programinfo for a recording if the filename matches Myths format.


The dvdwizard patch will NOT be applied as it relies heavily on Qt based
controls. Aaron if you'd like to put together one using Myth controls
and the theme engine I'll work with you to get it in.


git-svn-id: http://svn.mythtv.org/svn/trunk@7057 7dbf422c-18fa-0310-86e9-fd20926502f2