Closes #181.
I need more info, such as a log with "-v channel,record" to determine what
exactly the problem was in #181, but this changes TVRec::SetChannel()
and ChannelBase::SwitchToInput() to return false on failure, and abandon
a recording if tuning fails.

My guess is that #181 didn't actually explicitly tune to the previous 
channel, but simply failed to tune to a new channel, and proceeded to
record anyway. If this is the case this commit should address that 
problem. If it doesn't, reopen the ticket and add more info.



git-svn-id: http://svn.mythtv.org/svn/trunk@7060 7dbf422c-18fa-0310-86e9-fd20926502f2