Documents RecorderBase, and DTVRecorder.
Changes WaitForPause() to accept a timeout value and
return false it it times out. Doesn't change behavior though,
there is already a default timeout in the recorders that
can timeout, and I haven't changed that, and NVR never 
times out.

Adds a couple debugging VERBOSE(VB_RECORD.. macros to DTVRecorder.



git-svn-id: http://svn.mythtv.org/svn/trunk@7068 7dbf422c-18fa-0310-86e9-fd20926502f2