This fixes the seek problem as it concerns the DummyDTVRecorder.
It turns out the problem is due to the av_seek fallback in
avformatdecoder. In particular this line:
    if (av_seek_frame(ic, -1, ts, AVSEEK_FLAG_BACKWARD) < 0)
I don't know if that seek backward flag is really appropriate
in DoFastForward.

The av_seek needs to be fixed, but the solution to the 
DummyDTVRecorder problem is much simpler...

The reason it was using the fallback method when you seek right
after LiveTV starts is that not every path calling SyncPositionMap
sets the recordingHasPositionMap variable, and seek may be called 
before a sync has been successfully completed in LiveTV mode. The 
solution is to set recordingHasPositionMap in SyncPositionMap, 
and always use the position map seek in LiveTV mode, where you 
should always be able to get one.



git-svn-id: http://svn.mythtv.org/svn/trunk@7076 7dbf422c-18fa-0310-86e9-fd20926502f2