Closes #109 by applying patch.
It needed a little editing due to all the scanner changes, 
and I added some "#ifdef USING_DVB" protection, but it is
functionally the same as the patch.

I decided to apply it as is. While getting the channel 
major and minor numbers would allow you to more easilly
use this for OTA channels, I don't really see the need
for that since a full scan takes so little time these
days and collects much more information. But a scan 
does not currently work for non-ATSC DTV channels on 
cable, and a channel.conf works...




git-svn-id: http://svn.mythtv.org/svn/trunk@7079 7dbf422c-18fa-0310-86e9-fd20926502f2