It appears that an event is only sent after a FE_SET_FRONTEND if the tuner actually has to do anything. Which means if the tuner already happened to be on that frequency we never get an event. But you can't just query the frontend with a FE_GET_FRONTEND if an FE_SET_FRONTEND hasn't been sent..
So we check for any tuning events in the queue. If there
are any we grab the last one and grab the tuning parameters 
from it. If there aren't any events we try a FE_GET_FRONTEND,
which should work when there aren't any events..

Then after we tune we check the old frequency against the
new one, if they match we don't wait for an event. 

We could probably avoid the tune if they match before the tune,
but I'm taking mini-steps here...



git-svn-id: http://svn.mythtv.org/svn/trunk@7134 7dbf422c-18fa-0310-86e9-fd20926502f2