Fixes channel changing problem for Torbjorn Jansson.
The problem was that his driver returns the actual center
frequency the hardware locked onto instead of the one we
tuned to. This is a neat feature, but we were only checking
for an exact match in frequency for when to not wait for a
DVB tuning event. With this commit any frequency within a
range -500kHz to +500kHz is considered equal as far as DVB
tuning events are concerned.



git-svn-id: http://svn.mythtv.org/svn/trunk@7161 7dbf422c-18fa-0310-86e9-fd20926502f2