Fixes #264.
The "VLSI VES1x93 DVB-S" and "Philips TDA10046H DVB-T" DVB
cards occasionaly munge the PMT. The PMT is still readable
but the CRC check fails. This commit changes MythTV so that
it does not do a CRC check on the PMT for these cards

I'm assuming these cards read the PMT and write it out again
much as MythTV does (probably to support CA since this only
happens on the PMT). But unfortunately they change something
(presumably the padding) without recalculating the CRC. 
If this assumption is correct these cards presumably verify
the original CRC before rewriting the PMT, so it is safe
to not check the CRC.

If I'm incorrect and these cards don't check the CRC then this
is still relatively safe since these packets had to get past
the error correction layer and the odds that the relatively
infrequent and small PMT's have enough bit flips to get past
EC with errors is very low.



git-svn-id: http://svn.mythtv.org/svn/trunk@7169 7dbf422c-18fa-0310-86e9-fd20926502f2