Close #236 . This would happen if the user set dupmethod=0 with a kFindWeeklyRecord or kFindDailyRecord rule. For all record types, "Don't match duplicates" meant that no showings should be marked as rsPreviouslyRecorded and all showings should record. For a Find* rule, this would mean that after one showing recorded, the other showings were not marked "P" so the scheduler would assign the next showing to be recorded also.
Due to popular demand, duplicates found as a result of a "findid"
match in the 'oldfind' table are now treated as an exception. A
FindID duplicate will be marked as previously recorded even when
the user selected "Don't match duplicates" so that only one showing
in the time interval (daily or weekly) will record regardless of
the duplicate method.

As before, showings with specific description can be marked as "P"
due to matching an oldrecorded entry. This is useful for things
like "Steve Martin (People Search)" set to FindWeekly. This will
record a different Steve Martin each week but would not record
"The Jerk" twice. After it recorded once, future showings would
be marked "P" and a different title would be selected for the week.

However, by setting "Don't match duplicates", even oldrecorded entries
will not match and any one showing will record in the interval
regardless of the descriptive info.


git-svn-id: http://svn.mythtv.org/svn/trunk@7195 7dbf422c-18fa-0310-86e9-fd20926502f2