Changed ScheduledRecording::doneRecording to send a Myth event when a recording finishes instead of calling ProgramInfo::AddHistory.  The scheduler will receive the event, log the status change and call AddHistory.
This accomplishes two things.  First, the scheduler doesn't have to
guess that a recording finished just because the end-time passed.
Second, and more important, it provides a mechanism for recorders to
report other status changes to the scheduler.

A likely future example of the latter is when a recorder can't detect
a signal.  The recorder can notify the scheduler and it can reschedule
the program for a later showing if possible.  Another example is when
a driver problem prevents the recorder from reading data like
sometimes happens with the ivtv recorder.


git-svn-id: http://svn.mythtv.org/svn/trunk@7197 7dbf422c-18fa-0310-86e9-fd20926502f2