This is a merge of the DVBChannel patch plus a new try at waiting for the backend to respond to frontend tune events.
Instead of waiting for DVB Events, which is broken on
several drivers, I'm issuing a FD_READ_STATUS ioctl.

From what I've read on the mailing list this should
block until the backend has returned from the tuning
request as of kernel 2.6.12. Because I don't expect
everyone to be running 2.6.12 or later, I've also
added a 2 ms sleep, which should allow the backend
to see the tuning event and leave the old frequency
on older kernels.

I tried using FE_GET_FRONTEND since unlike FD_READ_STATUS
it returns the frequency, but this doesn't work with the
Air2PC card, and probably others.

Since the FD_READ_STATUS is used in the signal monitor
the only fix for the problems pre-DVB-event-fix is the
2 ms sleep. So conclusion here is that those having DVB
problems should upgrade to Linux 2.6.12 or later.

PS This commit may cause new problems for DisEqC users,
   please let me know if it does.



git-svn-id: http://svn.mythtv.org/svn/trunk@7204 7dbf422c-18fa-0310-86e9-fd20926502f2