John Pullan's fix for Full Tuned Scan in the UK.
This adds the requirement that we be in the scanning state
before processing any SIParser events. This shouldn't be
needed because current shouldn't be set until after we
have tuned to a new transport. But somehow this isn't
what happens.

The SIParser stuff is kinda crazy with callbacks so I
think this is the best fix until we rip that stuff out.



git-svn-id: http://svn.mythtv.org/svn/trunk@7206 7dbf422c-18fa-0310-86e9-fd20926502f2