Changed the basename for new recordings to <chanid>_<starttime>.nuv and save the basename in the recorded table.  Always save the actual end time of a recording, including overrecord, in the recorded table.
This is mainly preparatory work to allow changing the end time of an
in-progrss recording.  Not having the end time in the basename
eliminates the need to rename files if it changes.  Also, having the
basename in the recorded table will allow for alternate naming schemes
in the future.  For example, the mpeg and dvb recorders could name
their files .mpg instead of .nuv.


git-svn-id: http://svn.mythtv.org/svn/trunk@7240 7dbf422c-18fa-0310-86e9-fd20926502f2