Refs #326.
This partially fixes the problem. This properly shuts down
signal monitoring on channel changes or when you exit the
recorder when you haven't gotten a lock.

This also increases the time between signal checks from 50 ms
to 250 ms. I hope to lower this again, but this should make
debugging this problem a little easier.

I think to fully fix this we first want to unify the
tuning process on channel changes and the initial/non-LiveTV
tune. This would also fix some longstanding problems with
the LiveTV code, such as bttv picture settings being set
in a different place after a channel changes than on initial
tune in LiveTV mode. And remove code duplication in the 
DVB/ATSC handling.



git-svn-id: http://svn.mythtv.org/svn/trunk@7259 7dbf422c-18fa-0310-86e9-fd20926502f2