Fixes #321; timeout not long enough for DVB over USB. Fixes #319; per tuner timeouts.
This changes the signal lock and channel lock timeouts
from constants in the code to setable values in the
backend configuration and sets reasonable defaults when
you configure a new card.

Those with existing cards may wish to fiddle with these
values, as the values are initialized to 1000 ms and 
3000 ms respectively. These values may not be appropriate
for your setup. i.e. if you use DVB-S, DVB over USB, or
have a low signal station and want to wait a bit longer
for a lock.

This also drops some old dvb_* tables, it is not really 
safe to regress from SVN to 0.16 now anyway, and 0.17+ 
does not use these tables.



git-svn-id: http://svn.mythtv.org/svn/trunk@7266 7dbf422c-18fa-0310-86e9-fd20926502f2