This changes the pausing in the recorders to use QWaitConditions instead of busy waiting.
There is still a timeout on the waitconditions,
but it is 100,000 ns rather than 50-1000 ns.
This is needed because someone should be able to
close the recorder without unpausing first.

I've tested the DVB, MPEG and framebuffer recorders.
But I don't have a Firewire/DBOX2 cable box so...

There is also a little bit for cruft removal in
dtvrecorder, dvbrecorder, and hdtvrecorder.

The DVBRecorder in particular had a pretty broken
pause, it had to finish an infinte loop before it
would pause when channel was off-air.



git-svn-id: http://svn.mythtv.org/svn/trunk@7288 7dbf422c-18fa-0310-86e9-fd20926502f2