Mostly functional free commit.
SetupRecorder() has the only functional change. The common
recorder intialization is done at the end instead of having
the same code repeated and instead of setting errored on an
error the SetupRecorder() now returns false on failure. This
means that if the recorder could start because of a permissions
problem you can fix that problem and it will work next time
without having to restart the backend.


The non-functional changes are basic refactoring things.


The general cardid based database options (videodev, audiodev, etc.)
are now in a structure like the dvb and firewire options.

The options structures are now classes instead of struct, which
basically means we can initialize the variables more simply.
This is why this extends past tv_rec.cpp, since these structures
are used to initialize the firewire and dbox2 classes (ugly btw).

I've also renamed the variables for these structures in TVRec
so that they are a little more terse.

I've also renames capturecardnum to cardid, which is what it is
called in the DB and in other classes.

Finally, I've added a LOC macro to tv_rec which is simply:
#define LOC QString("TVRec(%1): ").arg(cardid)
This lets us just prepend this to debug messages and know
which recorder is complaining by using VERBOSE(..., LOC + ...);
The rest of the changes are just adding this "LOC + " string 
to the VERBOSE macros. And the LOC_ERR version which just adds
"Error: " to the LOC string.



git-svn-id: http://svn.mythtv.org/svn/trunk@7301 7dbf422c-18fa-0310-86e9-fd20926502f2