Adds RecorderPaused() signal to recorderbase.h
This signal is emitted whenever the recorder actually enters
the paused state.

I've made an increadably silly formatting fix in libmythtv.pro
only so that I don't have to explain to a half dozen people that
they need to run "touch libmythtv.pro" because a new class using
Qt signals has been added.

If it isn't obvious from the last commit this is in sevice of a
busy-loop free TVRec...



git-svn-id: http://svn.mythtv.org/svn/trunk@7312 7dbf422c-18fa-0310-86e9-fd20926502f2