Adds simple EIT crawl to EITScanner. + deleteLater support.
The EIT crawl visits every channel, instead of just visiting
every transport. It is simple-minded in other ways too...
I figure someone else will want to optimize this once they
realize how simple it is. Until then we can have something
functional.

This commit does not enable it. It is enabled in the TVRec
patch I'll be releasing later today. (If you use it with
the current TVRec it will change your starting channel
as it scans.)



git-svn-id: http://svn.mythtv.org/svn/trunk@7323 7dbf422c-18fa-0310-86e9-fd20926502f2