GetFramerate() fix + refactor.
GetFramerate() is currently caching the framerate from the
recorder, but this can cause filelength estimation problems
with DTVRecorders, where the framerate can change, after
this change the value is returned directly from the recorder.
The file length estimation can still be incorrect, so I added
a todo comment for it.

As for refactorings, this moves "inoverrecord" and 
"overrecordseconds" from the state variables section to
the program section, as they are specific to a program.
This also changes StartedRecording() and FinishedRecording()
to take a ProgramInfo. These should make with Issac's 
LiveTV changes easier, and also allow for a less-blocking
StartRecording() function when combined with the tuning
patch.



git-svn-id: http://svn.mythtv.org/svn/trunk@7327 7dbf422c-18fa-0310-86e9-fd20926502f2