Removes busy loops from TVRec.
This implements WaitForEventThreadSleep(), which
releases the state lock and waits for the event
loop to send a triggerEventSleep.wakeAll().
Unless you pass false to it as the first param
it will kick the event loop to make it process
any events you have just created for it process.

This changes locking around a bit, which is why it
wasn't lumped in with the flags commit. I've been
using something very similar in the tuning patch
without problems; but if you have problems with
any commit today, try reverting this one first.



git-svn-id: http://svn.mythtv.org/svn/trunk@7331 7dbf422c-18fa-0310-86e9-fd20926502f2