This is heart of the much talked about tuning patch.
The missing piece is the RingBufferInfo removal. I'm
not using any of the pausing in there, but I'm ripping
that out in a seperate commit. I also promised someone
an antenna adjustment mode, but that will have to be
wait a little bit.

This does enable a very basic EIT Crawl. The intent here
is to inspire someone to improve it, but provide basic
functionality in the meantime.

What this change is really about is removing all the 
complicated locking and the multi-threaded tuning TVRec
had acquired over the last couple years. About half of
it was my fault, so I'm not blaming anyone. :) 

I've gotten very positive responses in the last few
days, esp from DVB users. I've also fixed all the
issues I could reproduce.

Please send any bug reports to me or the devel list
instead of opening tickets for the next week or so;
I will be reading the list frequently, and it is much
easier to miss a comment in a long ticket than a post
on the mailing list...

BTW I've set the sleep time in the event loop to 30
seconds. I will probably lower this to later, but it 
is high now so you will notice if any event isn't 
triggering the event loop to run.  If you want to 
change it, the wait is is on line 1188 of TVRec, but
please send me a bug report along with the output of 
   "mythbackend -v record,channel"   first...
I didn't get any reports of these types of problems from
the patch testers, but SVN testers are a large superset
of patch testers.

BTW2 I did test skipping the signal monitoring stuff
on channel changes. It made no measurable difference
in channel change time, so I ditched that code...



git-svn-id: http://svn.mythtv.org/svn/trunk@7334 7dbf422c-18fa-0310-86e9-fd20926502f2