Mostly this just gets rid of RingBufferInfo.
But it also adds a SetRingBuffer(NULL) on line 3636,
which gets rid of the RingBuffer if we fail to open
the backing file.

It also gets rid of any dummyRecorder using an old
RingBuffer in SetRingBuffer(). A new one is created
when it is absent and we need one (existing code).



git-svn-id: http://svn.mythtv.org/svn/trunk@7341 7dbf422c-18fa-0310-86e9-fd20926502f2