Adds a new channel ordering method "ATSC Channel", it orders by the atscsrcid column. It's a little smarter than "channel number (alpha)". That is assuming you just have ATSC channels, it's pretty dumb otherwise.
Until DoGetNextChannel() can handle multiple column ordering
You will probably get the best results by setting the channel
ordering to numeric. In LiveTV it will balk at a channel like
"5_1" and switch to "ATSC Channel" ordering if it is an ATSC
channel, or "Channel Num (alpha)" ordering if it it is not an
ATSC channel, but on a regular recorder it will use the
numeric ordering. In the program guide the channels will not
be in order, but it will be close enough (major channels will
be in order).



git-svn-id: http://svn.mythtv.org/svn/trunk@7355 7dbf422c-18fa-0310-86e9-fd20926502f2