Implements a basic "Antenna Adjustment Mode". (F7)
This turns off the recorder in LiveTV to let
you adjust the antenna and see a constant 
signal monitoring OSD.

This could be more useful with a better signal
monitoring OSD, but this is a start.

This may also fix the spurious "You should have
gotten a signal by now" dialogs.



git-svn-id: http://svn.mythtv.org/svn/trunk@7376 7dbf422c-18fa-0310-86e9-fd20926502f2