Fixed a long standing problem in the scheduler where a low priority program could prevent a higher priority one from being recorded. Because programs that could not initially be scheduled were not retried until the very end, the low priority program could fill the slot needed to rearrange higher priority programs.
The program and rectype priorities are now the primary factor used to
determine the order in which programs are initially tried.  The
channel and input priorities are now used as a secondary factor.
Programs that can not initially be scheduled are now retried right
after all prgrams at the same primary priority are initially tried.
This prevents any lower priority programs from grabbing the slots
needed to rearrange other programs.


git-svn-id: http://svn.mythtv.org/svn/trunk@7391 7dbf422c-18fa-0310-86e9-fd20926502f2