Condenses the two questions in when you start-up mythtv-setup onto one dialog. This also removes the warning in the howto and replaces it with a check for a running mythbackend on startup. If it is seen running locally, a warning about what isn't safe while the backend is running is added to this initial dialog.
The check for mythbackend is very simple, it just runs
"ps -ae | grep mythbackend". So if you rename the backend
executanble it won't detect the backend. It also won't dedect
a master backend running on a different host, etc. It just
catches the common case.

"ps -ae" works for me on Linux, IRIX and Solaris. But I don't
have a BSD to test. I know "ps -aex" doesn't work on Solaris.
If the check fails you only miss this informative message.



git-svn-id: http://svn.mythtv.org/svn/trunk@7392 7dbf422c-18fa-0310-86e9-fd20926502f2