Fixes #430.
This is a simpler fix for the time skew portion of the
consecutive recording problem. Part of the problem I 
was trying to solve was already solved by David Engel.
MythTV should no longer try to schedule something for 
recording that is already recording on a slave backend.
So the GetRecording() request is not needed in the 
scheduler. We can instead simply cancel any non-LiveTV
recording if the master backend requests a new recording.



git-svn-id: http://svn.mythtv.org/svn/trunk@7433 7dbf422c-18fa-0310-86e9-fd20926502f2