I bunched a three mythcontext changes to avoid too much recomplining.
I've added a ENO macro which when appended with "+" or "<<" to
the VERBOSE args will print out the strerror_r() message on the
next line. This should let us remove all various perror() and
strerror() calls over time, and standardize the errno messages.
The errno is a thread local variable. So those two functions
are not thread unsafe, but perror() in-particular can mess up 
debugging logs because it doesn't respect the verbose lock.

The second change is to make MythEvent accept an automatic
(i.e. on stack) QString. This is just for convenience of
being able avoid an extra variable declaration.

Third, I added more documentation to the MYTH_PROTO_VERSION define,
specifically adding a remininder that the mythbackend.php version
of that value also needs to be incremented. :)



git-svn-id: http://svn.mythtv.org/svn/trunk@7453 7dbf422c-18fa-0310-86e9-fd20926502f2