Fixes #408.
The Xhannel and NVR needed to be closed when calling 
NVR::Initialize() when using an MJPEG card.

This closes and reopens the channel when needed, and also
makes it safe for the NVR to be open. The later part isn't
needed right now, but there was also a problem in that the
MJPEG init closed the NVR file descriptor but didn't reset
it to -1. It was easy to allow the NVR to be open when
fixing this.

This also uses the new ENO macro instead of perror() to
print out errno based error messages.

mkundinger can you verify this fix?



git-svn-id: http://svn.mythtv.org/svn/trunk@7454 7dbf422c-18fa-0310-86e9-fd20926502f2