Makes it possible to use a X11 Xinerama screen with non-square pixels.
When using Xinerama it uses a new DB setting "XineramaMonitorAspectRatio"
to determine the physical aspect ratio of the screen. This is set right
under the XineramaScreen setting. If this is not set, we default to the
square pixel assumption. The aspect ratio's in the combobox are 4:3,
16:9 and 16:10. Tell me if I missed some common monitor aspect ratio.

I've also documented many of the variables in videooutbase.h...

Finally, when using X11 it is now possible to use a window dimensions
that are not proportional to the physical dimensions of the screen
and it should always work.



git-svn-id: http://svn.mythtv.org/svn/trunk@7463 7dbf422c-18fa-0310-86e9-fd20926502f2