Fixes #473. Combines three settings checkboxes into one combobox.
Instead of having checkboxes for libmpeg2, XvMC, and XvMC-VLD, there is
now one combobox that lets you select the type of playback to use for
MPEG2 streams. ffmpeg is listed as "Standard" while the others are 
selectable options. The benefits and drawbacks of each are described
in the help text.

If you are using anything other than ffmpeg you need to go into this
"Setup:TV Settings:Playback" and select your "Preferred MPEG2 Decoder".



git-svn-id: http://svn.mythtv.org/svn/trunk@7497 7dbf422c-18fa-0310-86e9-fd20926502f2