Fixes #328.
This fixes a problem in mpegts.c wrt PAT scanning, the old
code assumed that every PAT seen was brand new so it tossed
out everything we knew. And it also didn't add services to
the services list, so that list became stale quite quickly.
This merges the pat_scan_cb with pat_cb, so that we now
add services to that list. (I left the pat_scan_cb 
function alone (and unused) for ease of ffmpeg merging.)

This PAT issue isn't a problem with the MythTV DVB & HDTV TS
streams because we always use the same program number of "1".
But this is a problem with general TS streams.

I've also added another PMT searching pass, for when we fail
to find a PMT in the stream the first time. This pass looks
for the same PMT, but without the CRC check. This lets us
play back mangled "ZDF" streams from the DBOX2 in Germany.



git-svn-id: http://svn.mythtv.org/svn/trunk@7502 7dbf422c-18fa-0310-86e9-fd20926502f2