Closes #373, by applying patch + some editing.
Parts of this patch had already made it into SVN, plus I've
made this multi-threaded software encoding default to being
disabled. Those that want to experiment with it can pass 
"--enable-ffmpeg-pthreads" to ./configure. You then need to
"make clean" before recompiling.

I haven't added this to the --help text for ./configure for now.



git-svn-id: http://svn.mythtv.org/svn/trunk@7507 7dbf422c-18fa-0310-86e9-fd20926502f2