Fixes #500.
Adjust Picture-in-Picture dimensions and location so that
it works with the various letterboxing modes.

Tested with 16:9 video embedded in a 4:3 video, and
4:3 video embedded in a 16:9 video.

This does not apply the letterboxing to the PiP video,
but plays it back as if it were in the letterbox_Off mode.
So the video has to report the correct aspect ratio
and MythTV needs to know your displays aspect ratio
for it to be played back with the correct aspect ratio.



git-svn-id: http://svn.mythtv.org/svn/trunk@7553 7dbf422c-18fa-0310-86e9-fd20926502f2