References #494.
This moves reschedule out of eithelper into eitscanner.
It also only reschedule when we have new info and either
we change channels or haven't seen new info in over a
minute.

If we wanted to be even kinder to the DB we could also
track when these EIT's are for and only call reschedule
if they are for the next 3 hours, or we haven't rescheduled
in over 3 hours...



git-svn-id: http://svn.mythtv.org/svn/trunk@7558 7dbf422c-18fa-0310-86e9-fd20926502f2