Closes #494.
I'm rejecting the 10 minute per channel patch, it is much better
to scan the channel a few times. If you leave a long time between
the times a channel is scanned it makes it more likely that you
miss a programming change in the next couple hours.

I've also changed the eitcrawl-num-events-added patch, the 
"EITHelper Added X events" is a debugging message, that tells
you something is going on so it shouldn't be disabled when
events are inserted. If you want to total the inserts it
should be done in the scanner, maybe when it changes channels.
To make that possible I changed EITHelper::ProcessEvents() to
return the number of events inserted into the DB.



git-svn-id: http://svn.mythtv.org/svn/trunk@7559 7dbf422c-18fa-0310-86e9-fd20926502f2