Closes #335. Really.
I had hoped to integrate Wendy's patch in a separate commit,
but I overwrote the original..

Anyway this adds search of multiple fields, by adding a comma or ">"
between phrases. The fields searched are filename, artist, album,
and title.

And, I was wrong about the regular expressions. A regular expression
detector would be needed for this, but otherwise the support is there.



git-svn-id: http://svn.mythtv.org/svn/trunk@7596 7dbf422c-18fa-0310-86e9-fd20926502f2