Fixes #505, References #501.
This fixes #505 by not trying to switch to tuners when there is no
other tuner to switch to.

But the reason both #501 and #505 are failing is because of the 
order of LiveTV ringbuffer sets. 

Issac how far off is your LiveTV change?

I'm not sure if I should bother to fix TVRec::SetupRingBuffer()
if this code will be gone in a few days...



git-svn-id: http://svn.mythtv.org/svn/trunk@7633 7dbf422c-18fa-0310-86e9-fd20926502f2