Fixes #540. channelqueued was not always being set and the support for to switching tuners while changing channels wasn't very DTV friendly either.
I cleaned up the channel keys input stuff, which means no more
channelqueued variable. Also the channel input didn't work properly
between tuners if you had smart keys enabled. This was because the
channel entry is cleared once you enter a channel prefix not 
corresponding to a channel on the current tuner. I added a check 
for whether the queued channel is on the current tuner, if it is
not we use the entire queued input. This is the same queue we 
were using for teletext switching and arbitrary jump.

FYI In order to jump immediately to a channel not on the current
tuner, you do need to hit select even if smart keys is enabled.

Also for better DTV friendlyness I changed the channel on the new
tuner a bit. Instead of changing to a new tuner, then changing the
channel, the code now simply verifies that the channel is valid for
that tuner and if it is sets it as the starting channel. This way
we don't go through the rigmoral of getting a lock on one channel
only to change to another one a second or two later.



git-svn-id: http://svn.mythtv.org/svn/trunk@7636 7dbf422c-18fa-0310-86e9-fd20926502f2