Allows edit controls to have a preferred position for the popup keyboard either above or below the edit or at the top or bottom of the screen.
Add a MythPopupBox::ExecPopupAtXY() so you can specify where
a popup should be placed on screen.

Changed the MythSearchDialog to place the popup keyboard at
the bottom of the screen so you can still see the search results
when the keyboard is shown.

This commit seems to touch more files than it should because
I removed the include of uitypes.h from mythdialogs.h, there
is no point in forward declaring all those uitypes if you then
go ahead and include the uitypes.h anyway. It had a knock on 
affect to a few other files.


git-svn-id: http://svn.mythtv.org/svn/trunk@7637 7dbf422c-18fa-0310-86e9-fd20926502f2