Fixes #268.
The EPG wasn't really forgetting the current channel.
The problem was that it was looking up channum which is
not always unique.

This change makes quidegrid look for the chanid if it is 
passed one and fallback on channum if it is not.

Then in LiveTV mode this change adds code to look up the 
chanid of the current channel. But in progfind.cpp and 
programs/mythfrontend/main.cpp we still use the channum.
This is fine because in these cases we aren't trying to 
start on a particular channel on a particular tuner,
but rather just start in the same place whenever the 
grid is brought up.

This also does a little bit of refactoring in that it passes
chanid around as an unsigned integer rather than a string.



git-svn-id: http://svn.mythtv.org/svn/trunk@7643 7dbf422c-18fa-0310-86e9-fd20926502f2