Closes #541.
This adds the EIT only listings source created by John Pullan
and adds a column to videosource called useeit, which when set
will allow the EITScanner to scan channels on any DTV source.
I've disabled this by default so that if you compile with EIT
parsing enabled, it default to NOT using EIT information, so
you must set this to true if you want EIT events.

You can still disable EIT scanning for any particular channel
by editing the channel.

To enable/disable EIT scanning for a source edit the source in
the "Video Source" configuration pane of the backend configuration.



git-svn-id: http://svn.mythtv.org/svn/trunk@7659 7dbf422c-18fa-0310-86e9-fd20926502f2