Added support for the additional "FROM" clause information stored in the kPowerSearch subtitle field. ProgLister() and modifyPowerSearchByID() accept a subtitle argument so they won't choke on a rule that uses this field.
In Custom Record, added a text box and examples that demonstrate how to
include additional tables in power search rules. "Person named in the
credits" includes 'people' and 'credits' to emulate a People search but
with the advantage that other criteria can be added. "Only on channels
marked as favorites" included the 'favorites' table and could turn out
to be very useful. "All matches for a genre" demonstrates a more complex
"JOIN".


git-svn-id: http://svn.mythtv.org/svn/trunk@7666 7dbf422c-18fa-0310-86e9-fd20926502f2