References #512.
Adds code to disable Chromakey OSD on cards/drivers that do not support
chromakeying. This avoids using this OSD display method when it won't 
work. i.e. with the nvidia 6600+ cards with the latest closed source 
drivers, or with earlier cards when the user has XvmcUsesTextures set
to true in the xorg.conf file.



git-svn-id: http://svn.mythtv.org/svn/trunk@7677 7dbf422c-18fa-0310-86e9-fd20926502f2