Fixes #539.
gkruse, this is probably not what you want to hear, but the 
ChromaKey OSD only supports 32 bit ARGB visuals.. and you
are using a 16 bit RGB visual.

This commit adds a check that the visual is a 32 bits visual,
if it is not ChromaKey OSD is not enabled. ChromaKey OSD could
be extended to support other visuals, but it would involve
creating new OSD blend functions.

One could also make Qt pick a 32 bit ARGB visual if one is 
available, but a google search reveals that this requires
subclassing QFrame to use a better visual, and ALWAYS using
the subclass to create windows. This is because Qt stores
the visual information in global variables, so that all Qt
windows in an application must use the same visual.. (Ugh!)
That kind of change doesn't make sense before the new MythUI
stuff is in place.

Until then you can probably configure X11 to pick a 32 bit ARGB 
visual by default for the root window. QFrame uses the root 
window's visual for the windows it creates, so this should
make ChromaKey OSD work for you.



git-svn-id: http://svn.mythtv.org/svn/trunk@7682 7dbf422c-18fa-0310-86e9-fd20926502f2