A little more cleanup in SIParser.
Two important changes. First, is a proof of concept of merging the
two sets of DVB/ATSC table parsers. This makes SIParser use the same
code to parse the VCT table in as the signal monitors and the 
HDTVRecorder. Second, is a fix to the language parsing in 
ProcessShortEventDescriptor(), which may address some EIT issues.

In other changes this prints out more info about unused descriptors
if you use "-v siparser" but doesn't process the descriptor at all
if you don't use "-v siparser".

This also doesn't add user configured blank language settings
to the priority list; which should help users that haven't 
configured a guide language, or have nasty some cruft in their DB.

For the VCT parsing I need to create a zero copy version
of PSIPTable and make sure the PSIP parsers don't depend on
the TS packet header being present before I apply similar
changes to other parsers. I chose the VCT for the proof of
concept because we parse it infrequencly, and unlike many
DVB tables I could test this parser extensively.



git-svn-id: http://svn.mythtv.org/svn/trunk@7727 7dbf422c-18fa-0310-86e9-fd20926502f2