Fixes #579, by applying patch from John Pullan.
It looks like it was just a two line fix in the end that did it,
the descriptor caches were being cleared at the start of the parse
of each descriptor.

But I'm glad I cleaned up all that language selection stuff in
the meantime. After [7758], it now initializes the preferred
listings/subtitle/audio languages from the UI language, and it
also handles the alternate language strings such as "deu"tch for
"ger"man and "isl"enska for "ice"landic.



git-svn-id: http://svn.mythtv.org/svn/trunk@7770 7dbf422c-18fa-0310-86e9-fd20926502f2