More of jd's patch for DVB over USB 1.0 support in signal monitor.
The problem is that the way the DVBSignalMonitor reads from the
DVB device is not supported well by these devices. It assumes
that when you ask for the filtered TS stream that you are doing
a recording and turns buffering up to the hilt and uses block
transfers. But if all you want to do is get a 188 byte PAT, you
have to fill an entire buffer with these tiny tables before you
get the first one. With a 1MB buffer it could take 532 seconds
to get the first PAT, then you have to wait another 532 seconds
for the PMT.. i.e. something on the order of 17 minutes!

What this does is it waits for a PAT for 150 ms, and if it
doesn't get one it tries to use jd's section reader instead.
The section reader doesn't support raw MPEG streams or ATSC 
streams, but AFAIC there are not yetany non-DVB DTV-over-USB
devices. And hopefully future devices will use USB 2.0 and 
be able to support the fully support the DVB API...

Oh, I also added some caching so we only do this 150 ms wait
on the first tune to a DVB device. I also made it a select
instead of a sleep, so it should usually sleep a lot less than
150 ms...



git-svn-id: http://svn.mythtv.org/svn/trunk@7779 7dbf422c-18fa-0310-86e9-fd20926502f2