Fix #589.
AFAIC we've never accounted for overscan in OSD placement, but
the x & y for the upper right corner of the OSD visible size were
transposed in the NVP after a recent change so that made the
adjustments we do make for letterboxing modes completely bogus
for any of the zooming modes.

In addition to the fix for the regression, I've also made 
GetVisibleOSDBounds() calculate the visible screen based on
the params set by MoveResize(), which should prevent it for 
getting out of sync with MoveResize as it did with respect
to the 16/9 Zoom and 16/9 Stretch letterboxing modes.



git-svn-id: http://svn.mythtv.org/svn/trunk@7786 7dbf422c-18fa-0310-86e9-fd20926502f2