Update on my recent patch to limit unnecessary calls to PlaybackBox::FillList().  This patch based on suggestion by David Engel to use a timer instead.
Instead of limiting calls to FillList() to once a second, we set a timer for
1 second whenever we receive a RECORDING_LIST_CHANGE event and the timer is
not active already.  This means we will fire off FillList() at most once
per second because of RECORDING_LIST_CHANGE events, but also will not wait
more than one second after recieving a RECORDING_LIST_CHANGE event to
update the list.  If we are playingSomething, then the timer event is
ignored when it fires.



git-svn-id: http://svn.mythtv.org/svn/trunk@7824 7dbf422c-18fa-0310-86e9-fd20926502f2