Refs #340.
Simplify the TV class to use the TVChain-derived ProgramInfo for its OSD information, instead of using a query to the backend.  Store the channel name + input name in the TVChain table to make this easier.

Remove obsolete queries (GET_PROGRAM_INFO and GET_INPUT_NAME), and bump the protocol version to 21.


git-svn-id: http://svn.mythtv.org/svn/trunk@7826 7dbf422c-18fa-0310-86e9-fd20926502f2