References #574. Don't free frame until we are done with it.
I thought this might be the cause for #574. But it is not.

My guess is that something is broken with seeking when using libmpeg2,
so that it seeks too far into an I frame, and misses parts of the image.

Does this happen with both PVR-250 and DVB/ATSC recordings?



git-svn-id: http://svn.mythtv.org/svn/trunk@7834 7dbf422c-18fa-0310-86e9-fd20926502f2