Closes #623, by applying workaround patch.
This just allows you to set set the start of the scan for 1394 avc devices on a different node on the command line. This works around a problem probing some devices resulting in a crash. Before this patch you had to change a define in 6200.c to avoid scanning other nodes.



git-svn-id: http://svn.mythtv.org/svn/trunk@7836 7dbf422c-18fa-0310-86e9-fd20926502f2