Refs #340.
Seeking across file boundaries for livetv.  Possibly ff/rew too, didn't test yet.

Behavior is: When seeking forward, and not playing the most recent file, seek to the beginning of the next program.  When seeking backward, it'll go to 5 seconds before the end of the file.

There seem to be some issues syncing the position map for the chains.  Need to look into those later.


git-svn-id: http://svn.mythtv.org/svn/trunk@7838 7dbf422c-18fa-0310-86e9-fd20926502f2