References #551. Documentation, plus a little tweaking in DiskLoop().
I added code to lower the write size in DiskLoop() if TFW::Write() is already blocking. This should be small enough so that the Write() will unblock soon enough so that packets get don't get dropped, but large enough so that the next full size disk write completes before Write() blocks again.

I also added bufferHasData.wakeAll(); in Write() when it gets blocked, so that the DiskLoop doesn't just sit around waiting for more data when a buffer full situation has been reached... This added up to 100 ms to the block time.



git-svn-id: http://svn.mythtv.org/svn/trunk@7846 7dbf422c-18fa-0310-86e9-fd20926502f2