Refs #340.
ToggleRecord now works, and is now basically just a couple db calls.  It just marks the program being watched in LiveTV as a non-livetv program (normal autoexpire value, 'Default' group). 
So, now it saves the entire program that's been watched from the beginning instead of from when 'ToggleRecord' was hit.

Change the recording group 'LiveTV' to be untranslated in the DB, and translated on display like 'Default'.

Add some locking for the playbackinfo pointer in TV, since it can now be changed out of thread.


git-svn-id: http://svn.mythtv.org/svn/trunk@7861 7dbf422c-18fa-0310-86e9-fd20926502f2