Closes #648.
This extends PreviewGenerator to create the previews itself
if the recording file is local.

And, this allows you to call Run() directly instead of calling 
Start(), and starting a thread. If you do call Run() and the
recording file is not local the preview generator uses the
MythContext sockets to the backend rather than creating its
own sockets. This is safe so long as the thread you are calling
from is the one communicating with the backend.



git-svn-id: http://svn.mythtv.org/svn/trunk@7886 7dbf422c-18fa-0310-86e9-fd20926502f2