Fixes #650.
This increments starttime seconds until we have a unique chanid+starttime.

This also calls StartedRecording() early enough so that the RingBuffer
is created with the right filename.

This also adds debugging VERBOSE macros to LiveTVChain, and delays
the setting of kFlagRingBufferReset on channel changes until after
the RingBuffer has been added to the LiveTVChain. This isn't a 
complete fix for LiveTV digital recordings, but was just enough to
allow me to test the StartedRecording() code.



git-svn-id: http://svn.mythtv.org/svn/trunk@7892 7dbf422c-18fa-0310-86e9-fd20926502f2