References #551.
Implements a suggestion by gigem to break up the AddNewRecords
query by 'record' rows. This doesn't make the total query take any
less time, but allows us to throttle back a little on how quickly
we demand the data from the DB. As suspected the problem is that
mysql is running at the same priority as the recording thread,
so if it is running all out for 20-30 seconds at a go, the 
recorder doesn't get a chance to write enough data to disk.

This adds a 50 ms sleep between each record row, but this can
be changed by changing the 'sleepTime' variable in AddNewRecords().
I don't know yet if this is big enough or bigger than it needs to
be, so try increasing it if you still have problems, and lowering
if you do not.



git-svn-id: http://svn.mythtv.org/svn/trunk@7911 7dbf422c-18fa-0310-86e9-fd20926502f2