Closes #602
Hal Burch's patch to preview schedule changes plus enhancements.

NOTE: You need to run ./configure and may need to "make distclean"
if you have trouble compiling.

NOTE: Update myththemes to current SVN if you use any of those
themes. If you are using a third party theme, make the equivalent
changes as the latest diffs for myththemes.

This adds a new item to the recording options page labeled "Preview
schedule changes". When selected, a 'speculative scheduler' is run
which includes any changes to the rule that is being edited. The
result is compared with the real schedule and the differences are
displayed on a page similar to the normal schedule page except that
there are two status columns showing the old and new statuses for
these items that would change.

In order to make room on the options page, a new group "Schedule
Information" holds "Program details" and "List upcoming episodes"
plus "List previous episodes", a mini Previously Recorded page which
displays the oldrecorded items that match the recordid or title of
the current item.


git-svn-id: http://svn.mythtv.org/svn/trunk@7931 7dbf422c-18fa-0310-86e9-fd20926502f2