Fixes #680, and fixes #664.
This creates seperate ringbuffers for the signal monitoring stage and
the video stage. This also fixes #680, this required the ringbuffer
reset flag to be set (which was needed for #664 as well) and for the
ringbuffer to be created.

I had this code last week, but it didn't work until after the short
video fix Isaac committed in [7958].



git-svn-id: http://svn.mythtv.org/svn/trunk@7961 7dbf422c-18fa-0310-86e9-fd20926502f2