Refs #105.
Don't delete recordedrating, recordedprogram, and recordedcredits table
entries for a program when we delete the recorded table entry unless we're
deleting the last recording with that same progstart.  Previously if you
had 2 recordings covering the same program (such as when you stop and
restart a recording for some reason), the three tables above would be cleaned
up when the first part was deleted.  Should Close #105.  If it doesn't, then
the ticket can be reopened if someone has more detailed information.



git-svn-id: http://svn.mythtv.org/svn/trunk@7986 7dbf422c-18fa-0310-86e9-fd20926502f2