Optimize scheduling by pre-marking generic episodes. Listings for TV series' sometimes have no descriptive information or a generic description of the series as a placeholder when the the station does not declare which episode of the series will be shown. This adds a flag column to the program table so that mythfilldatabase can mark these while grabbing new listings. By checking program.generic, the scheduler can avoid some expensive string matching that was used to search for these generic episodes.
The also fixes a problem introduced in [7937]. While it seemed a good
idea, "Check gContext->IsConnectedToMaster() before calling..." does
not return TRUE until after ConnectToMasterServer() and therefore
signalChange(-1) would never be called. Unfortunately, it is the
ConnectToMasterServer() that causes the warning/error messages :-(.
I've reverted and added a message to the user about the warnings.


git-svn-id: http://svn.mythtv.org/svn/trunk@8025 7dbf422c-18fa-0310-86e9-fd20926502f2