Fixes #692.
FindKeyFrames() stopped looking for frames once it had seen
a keyframe stream id. But FindKeyFrames() doesn't just find
key frames, it also tracks the total number of frames seen.
So this meant that streams that contain GOP headers would be
undercounted by one frame per keyframe interval.

This doesn't effect all streams, so I put a little note of
warning into the comments for anyone wishing to optimize
this function in the future.



git-svn-id: http://svn.mythtv.org/svn/trunk@8080 7dbf422c-18fa-0310-86e9-fd20926502f2