Closes #740. This should fix the segfault.
The segfault happened because SwitchToProgram() didn't
check for the being errored before resuming playback.

This adds that check in SwitchToProgram().

It also adds it to the StartPlaying() loop. This part is just
so that an error message gets displayed before we exit.

For an extra level of paranoia this also adds a check for a null
context pointer in the packet reading portion of GetFrame(). 
This prevents a segfault if GetFrame() is called when there is 
no context pointer; so long as there it is reset clearing any
buffered packets.

There are still remaining problems in LiveTV that I think are 
sufficiently canvased by the existing tickets; this just prevents
a segfault if we switch to an empty file. Addressing the creation
empty files is a seperate issue.



git-svn-id: http://svn.mythtv.org/svn/trunk@8098 7dbf422c-18fa-0310-86e9-fd20926502f2