Fixes one of the prebuffering pause problems.
The OSD list button was using pixmaps to draw the button background images. This meant it need a qApp->lock()/qApp->unlock() which caused a deadlock in the decoding thread when you used the EPG to switch between channels with different resolutions or aspect ratios. With these changes the background images are generated without using pixmaps and so we do not need the lock. Hence we do not cause this deadlock.



git-svn-id: http://svn.mythtv.org/svn/trunk@8115 7dbf422c-18fa-0310-86e9-fd20926502f2