Rename OSD::StartPause/UpdatePause/EndPause to ShowStatus/UpdateStatus/EndStatus.  Add handling for new "statusposition", "progbefore", and "progafter" elements in the "status" OSD container.  The statusposition is a position image that can be displayed at the current position on the slide bar rather than requiring a statusslider fill image to be defined.  statusposition and the original statusslider may be used together or independently.  progbefore and progafter are images that will be displayed in LiveTV if there is a program in the chain before or after the current program being watched.  These prog* items are hidden when they do not apply such as when the status OSD is used to display the brightness control.
Add ability to mark a OSDTypeImage as hidden so the code can hide images
rather than having to load them with a transparent image to make them
seem hidden.  This could be extended to other OSDTypes as well, but
the only place that makes use of this is the OSD methods mentioned above.

As part of this, the LiveTV position OSD which used to show the amount of
time behind Live has been changed to reflect the current position in the
current program.  If theme designers utilize the progbefore and progafter
images described above in their OSD themes, then the OSD will also reflect
whether there is a program before or after the current program being
watched.  When watching a program that has already ended, the OSD displays
the current position and total length of the program.  When watching a
program still in progress, the current position and total time recorded
so far is shown (this will increment while playback is paused).

Also clean up a little code in a few places where OSD::ShowStatus() is
called.

Refs #718.



git-svn-id: http://svn.mythtv.org/svn/trunk@8119 7dbf422c-18fa-0310-86e9-fd20926502f2