GetFree[Disk]Space changes.
Fix PlaybackBox's free disk space calculations.  Previously, it was only
using the first entry returned from RemoteGetFreeSpace() so it was only
showing the Master's disk space.  Now, it adds up all the free space
from all entries in the vector.

Move StatusBox's eliminate_duplicates logic up into remoteutil's
RemoteGetFreeSpace().  This way the PlaybackBox can benefit also.
This should eliminate counting free space twice in PlaybackBox if any
backends are sharing storage.

In MainServer::HandleQueryFreeSpace(), keep track of which hosts
we've put on the list so we don't put the same host on multiple times
if there is more than one encoder on that host.  This should prevent
the same hostname from appearing twice in the hostname field after
the duplicate elimination code is run in RemoteGetFreeSpace().



git-svn-id: http://svn.mythtv.org/svn/trunk@8135 7dbf422c-18fa-0310-86e9-fd20926502f2