StatusBox updates and a remoteutil RemoteGetUptime bugfix.
Back out parts of Hal Burch's status page patch and make some enhancements
to other parts.  Bruce beat me by about 10 minutes, I think he committed the
patch while I was busy typing my monologue below.  Then I went to commit and
it failed so it took another 5-10 to remerge. :)

RemoteGetUptime() was always returning false even when it succeeded.

Since we are pretty confident we know who is sharing storage, we can use
that information to figure out how much recording time is left on each
storage device based on the average and maximum bitrates for recordings
stored on that device.  StatusBox now looks at each storage location it
knows about and queries the average and maximum bitrates for the
recordings recorded by the backends that share that storage.  This avg and
max are now printed instead of the old "typical" hard-coded value of 5.5Mbit
and the hardcoded HDTV value of ~19Mbit.  The status box also shows what
the average and max bitrates are for that storage.

Also make the StatusBox easier to read for the Machine Status page by
using the area at the top of the page to show information about the current
section highlighted.  If you are on a line for a server, that server's info
will be at the top of the screen, if you are on one of the storage lines,
the information about that storage will be printed at the top of the screen.
This was done mainly because the longer strings do not fix in the scrollable
area so some pieces of information were unreadable.



git-svn-id: http://svn.mythtv.org/svn/trunk@8141 7dbf422c-18fa-0310-86e9-fd20926502f2