Fixes some audio initialization problems + audio/subtitle selection References #259.
For the audio initialization problem, the problem was that the
audio was only getting reinitialized if the number of channels
or sample_rate before decoding was changed, but this doesn't 
work properly in the presence of AC3 since the number of 
channels output depends on the number of channels encoded
and is not reported until you begin decoding A/V packets. The
number of channels were also not detected properly with codecs
like AC3 either because of the late channel/rate reporting.

There still seems to be a clicking problem with channels > 2,
at least when using ALSA out, so I kept the 2 channel restriction
for AC3 streams, but it is now a documented define, MAX_OUTPUT_CHANNELS,
which can be set to a larger number or zero (to use the same
number of channels as the encoding.) The number of mp3 channels
was not restricted before, and I haven't added any restriction;
if you have a problem let me know, these could be easily left
out of auto-selection at least.

For the auto audio/subtitle stream selection, this adds a language
preference for audio, and saves the last user selected audio/subtitle
so MythTV should resume a manually selected stream after a commercial
break, and try to find a stream in the same language on channel
changes. The channel change feature will only work on properly on 
streams that have a language descriptor, but the resumption after
a commercial break should work even in the absence of descriptors.

If you have problems with the DVB subtitle parts of this please
attach a 'mythfrontend -v playback' log to the #259 ticket.



git-svn-id: http://svn.mythtv.org/svn/trunk@8152 7dbf422c-18fa-0310-86e9-fd20926502f2