Fixes a segfault in the backend, when the DVBRecorder is shutdown before it is started up. This happens if the frontend is killed at an inopportune moment. This was fixed by setting the DVBRecorder's ringbuffer is to NULL before it is deleted in tv_rec.
This also adds a parameter to DummyDTVRecorder::StartRecordingThread()
which makes it wait until a set number of frames is written to disk
before returning, this along with setting the ringbuffer ready flag
after a few frames are written to disk prevents the AvFormatDecoder
from telling the NVP that it hit the end-of-file before we have an
opportunity to write anything to disk.

The NVP still has problems in LiveTV when you switch from an on-air
channel to an off-air channel, but this commit should fix the backend
portion of this bug.



git-svn-id: http://svn.mythtv.org/svn/trunk@8155 7dbf422c-18fa-0310-86e9-fd20926502f2