Fixes audio lockup when switching to a video only stream.
Before this commit GetFrame() is being called requesting 
audio after the audio stream is gone. This is because we
don't always know about the stream change until after 
GetFrame() is called. This change tells the NVP about the 
lack of audio as soon as the AFD knows about it and disables
the request for audio within GetFrame() when the audio 
stream goes away while the frame request is in flight.



git-svn-id: http://svn.mythtv.org/svn/trunk@8157 7dbf422c-18fa-0310-86e9-fd20926502f2