Closes #643
In cases where two recording rules match the same showing, one
of them needs to take precedence. An rsRepeat or rsInactive will
not be considered for recording so we penalize these to force
them to yield to a rule that may record. Otherwise, more
specific record type beats less specific.

The patch suggests priority as a factor but neither gigem or I
are comfortable with this. The controlling rule could flip-flop
when changing a priority on a different showing for a different
reason. Also, it is unknown if there would mysterious behavior
due to the total priority calculations (input preference, channel
priority) that may lead to unexpected results. Therefore, this
will not go in unless someone can make a very compelling argument
for it.


git-svn-id: http://svn.mythtv.org/svn/trunk@8191 7dbf422c-18fa-0310-86e9-fd20926502f2