JobQueue & Commercial Flagging related changes.
Don't run jobs for LiveTV recordings, it was a bug that they were running
(and the fact that they didn't always run made it worse).  Prior to the
LiveTV rewrite, jobs did not run for LiveTV recordings either.  If you want
the benefits of commercial skip/detection during a live recording, you need
to do an actual recording and watch the recording in progress.

In mythcommflag, don't bother trying to flag recordings in progress if they
were less than 1 minute long if logo detection is Off or if the recording
was less than 6-7 minutes if logo detection (or ALL) is On.  The amount of
video required to detect a logo is currently set around 7 minutes.



git-svn-id: http://svn.mythtv.org/svn/trunk@8201 7dbf422c-18fa-0310-86e9-fd20926502f2