References #712.
This commit could cause problems. It changes how PIDs are tacked,
instead of using 5 QMaps and one vector to keep track of information
about the streams, this creates one entry in one QMap to track the
information. This means 4 less QMap lookups per TS Packet processed
by the recorder.

This also changes the TS->PS transform to use the same keyframe tracking
as the TS recordings, and to use the same source for constants instead
of using it's own defines.



git-svn-id: http://svn.mythtv.org/svn/trunk@8206 7dbf422c-18fa-0310-86e9-fd20926502f2