Fixes #707.
This removes the need to see a PAT packet in SupportsTSMonitoring. This breaks support for the broken DEC2000-t driver, but it should speed up initial tuning for everyone else. The fix to the DEC2000-t driver  is scheduled to make it into the kernel soon, until then the patches can be found in jd's message in mythtv-dev: http://www.gossamer-threads.com/lists/mythtv/dev/166172



git-svn-id: http://svn.mythtv.org/svn/trunk@8209 7dbf422c-18fa-0310-86e9-fd20926502f2