Fixes #574.
libmpeg2 seems to be delaying an I-frame block decodes if it can't do them in a certain time and just displaying what it has so far. The fix is to simply not use it for generating previews. Someone who knows more about libmpeg2 could perhaps disable this feature instead, but it is probably useful during playback, so this would be disabled for previews only anyway so forcing ffmpeg use for previews is probably just as good a solution.



git-svn-id: http://svn.mythtv.org/svn/trunk@8213 7dbf422c-18fa-0310-86e9-fd20926502f2