Closes #711.
Doug pointed me in the right direction on this. Linux systems that 
do not allow us to keep a privledged thread, allow for regular users
to set the scheduler priority using either rlimits or the realtime
module. I've updated the HOWTO sgml to reflect this; rlimits is the
preferred way since that module made it into the Linux in 2.6.12,
but some distributions supported the other contender, the realtime
module. I've added a short blurb on both, plus a small paragraph
on SUID for systems that don't support either of the new options.



git-svn-id: http://svn.mythtv.org/svn/trunk@8215 7dbf422c-18fa-0310-86e9-fd20926502f2