Honor the suggested next download time for DataDirect grabbers.  At his request, this is the last half of the support that Robert Kulagowski started in [8105].  This only applies when using the internal housekeeper to run mythfilldatabase, it has no effect when running mythfilldatabase via cron.
When the 'MythFillGrabberSuggestsTime' setting is turned ON (it is ON by
default), the DataDirect code in mythfilldatabase will query the data
provider for the next suggested download time.  This time is stored in
the settings table as 'MythFillSuggestedRunTime' and the 'MythFillMaxHour'
and 'MythFillMinHour' settings are modified to allow the housekeeper to
run mythfilldatabase at the suggested time.  I've noticed the next
suggested download time from zap2it may be less than 24 hours away, so
the code compensates for this and will allow running of mythfilldatabase
more than once in a 24-hour window if the data provider recommends it.
One example of this is running mythfilldatabase at 3:30am on day 1 and
the data provider then recommends running at 12:30am on day 2.  The
period between these two times is less than 24 hours.  Because
mythfilldatabase is modifying the settings table, it now sends out a
CLEAR_SETTINGS_CACHE message to notify the backend that some settings
may have changed.  If the data provider tells us to download again on
the same day we just ran, the next time will be set to 24 hours from
now instead of what the provider recommended.

Also add LOC and LOC_ERR #defines in datadirect.cpp.



git-svn-id: http://svn.mythtv.org/svn/trunk@8218 7dbf422c-18fa-0310-86e9-fd20926502f2