References #368.
More of Frank Muenchow's patch.

This adds decoding for field 2 w/ivtv, so we can now decode XDS.
XDS tells you about the channel, program name, V-Chip, time, etc.

This also adds decoding for VPS and WSS; these are printed out
when you run with '-v vbi'. The VPS info would of course be
more useful on the backend, but that is left as an exersize
for people in PAL land. The WSS info could be very useful on
the frontend as it tells you the aspect ratio of analog 
transmissons...



git-svn-id: http://svn.mythtv.org/svn/trunk@8241 7dbf422c-18fa-0310-86e9-fd20926502f2