References #689.
I don't think this is causing the zero byte recordings, but it could be a contributing factor. Basically, if the PES header code straddles TS packets in a certain way the scanner could have missed them before this change. Also I made the frame tracking variables unsigned, except the first_keyframe variable which is now set to -1 before it is set. This lets us eliminate a couple variables and simplify the code further.



git-svn-id: http://svn.mythtv.org/svn/trunk@8258 7dbf422c-18fa-0310-86e9-fd20926502f2