References #802.
2nd patch from nooneimprt4nt AT gmail , com

This fixes the problem, so long as the channum's are unique across
all inputs on a card.

nooneimprt4nt, can you look at SwitchToInput() and see if you could
make it perform the DiSEqC switch to the specified input even if
a channum exists on two inputs? You probably need to make some
modifications to GetChannelOptions()...



git-svn-id: http://svn.mythtv.org/svn/trunk@8277 7dbf422c-18fa-0310-86e9-fd20926502f2