Fixes #813.
 * Delays the ringbuffer ready flag in one instance.
 * Delays RingBuffer switch if more than 256000 bytes remain in the RingBuffer; even if there are less than three valid video frames.
 * Increases RingBuffer retries on RingBuffer switches to approximately 5 seconds.

The RingBuffer switch threshold of 256000 should really be keyed to the bitrate, but I'm not sure exactly what it should be. This is just double the value used then 3 valid video frames remain. This change is needed so we don't switch prematurely when using XvMC, or when playing back on a marginal system with XVideo (i.e. a 1080i stream on a debug build, running on a P4-2800).



git-svn-id: http://svn.mythtv.org/svn/trunk@8286 7dbf422c-18fa-0310-86e9-fd20926502f2