Closes #590 by applying patch.
This adds the ability to see the keys mapped for any specific contexts.
This new feature is accessible via the menu (hit the info or menu key 'i'/'m')
This is a purely informative display; it helps you avoid hiding a key usable in that context but possibly defined in a parent context.

If you see a context problem check out FindContexts() in mythcontrols.cpp.



git-svn-id: http://svn.mythtv.org/svn/trunk@8287 7dbf422c-18fa-0310-86e9-fd20926502f2