Fixes #689, fixes #816. Don't abort the scan for pictures too early...
Torbjorn Jansson actually reported to me that an earlier fix for #689
appeared to have worked for him. But this fixed a final problematic 
stream of David's so I'm closing this ticket with this commit. #816
appears to be a result of this same bit of code, basically if the 
picture code was transmitted too long after the seq or gop it wouldn't
be found. This is sort of the exact opposite of the problem in #799.



git-svn-id: http://svn.mythtv.org/svn/trunk@8290 7dbf422c-18fa-0310-86e9-fd20926502f2