Fixes #801, by applying patch from jwestfall at surrealistic dt net
This prevents avlib from crashing when a stream change occurs during the stream detection rutine. This can occur with default settings with the firewire or dbox2 recorders and could happen with the hdtvrecorder if you turn off 'wait-for-keyframe' and a stream change occurs within a couple milliseconds of the recording start.



git-svn-id: http://svn.mythtv.org/svn/trunk@8398 7dbf422c-18fa-0310-86e9-fd20926502f2