References #868.
This is a temporary fix while I work on a more efficient and elegant one.
This fixes the skipframe loop and forces a seek on FF.

The forced seek is the temporary part. Fixing the frame discard will take
me a little time and this should allow people to use single frame seeking
in the meantime.



git-svn-id: http://svn.mythtv.org/svn/trunk@8399 7dbf422c-18fa-0310-86e9-fd20926502f2