References #868.
Takes advantage of our new found knowledge of when a frame is in use by avlib for regular mpeg2 decoding. It puts these frames on the end of the available list after a ReleaseFrames() call and tries to avoid them if other frames are available when GetNextFreeFrame() is called. This should prevent frames from getting reused while avlib is still using them in a low buffer situations.

The GetStatus() debugging rutine now uses caps for frames not thought to be in use by the decoder, and lower case frames marked as in use.



git-svn-id: http://svn.mythtv.org/svn/trunk@8425 7dbf422c-18fa-0310-86e9-fd20926502f2