Some date and time format changes for the settings page.
Add the new date format from #958,  Closes #958.

Change the sample date used to be the current date unless the month is equal
to the day in which case tomorrow's date will be used to avoid ambiguity
between the day and month when both are specified in numeric format.  The
help text is changed accordingly to notify the user if the sample date is
today's date or tomorrow's date.

Change the sample time used to be the current time.

Fix and clarify the help text for the time format setting.



git-svn-id: http://svn.mythtv.org/svn/trunk@8516 7dbf422c-18fa-0310-86e9-fd20926502f2