Fixes #743.
This is a narrow fix for the reported problem. The problem is that some DVB drivers still don't change channels properly and send you data from the last transport after it has reported being safely tuned to a new transport. This causes MythTV to print out a warning.

This warning leads to lots of spurrious MythTV bug reports. To suppress these driver bug reports in the MythTV bug tracking system, this commit delays the warning for a little over one second, and if it sees that the data becomes valid for the current transport in that time, it suppresses the warning.



git-svn-id: http://svn.mythtv.org/svn/trunk@8551 7dbf422c-18fa-0310-86e9-fd20926502f2