Closes #866.
This applies the patch with two modifications.
 1/ Update recordid in recorded table to match the recordid in the
    scheduled recording in NotifySchedulerOfRecording().
 2/ If a LiveTV recording flagged to be made a regular recording ends
    while the user is still viewing the recording as a LiveTV recording
    then we now add a scheduled item for the recording so that any
    transcode, commflag, or user jobs can be run.

This does not implement Adrian's suggestion of allowing channel changes to other tuners because it is non-trivial to implement.

This also does not implement Bruce's suggestion that the scheduled recording item be added and deleted immediately when the user toggles the 'R'ecorded flag. To do this properly we would need to implement Steve Adeff's suggestion of a conflict OSD when you toggle the flag. This would be very nice, but requires a bit of coding to create the OSD for it and resolve all the corner cases such as if a scheduled recording wants to start while you are examining the conflicts. I think this should be implemented as an additional feature, and probably pushed to 0.20.



git-svn-id: http://svn.mythtv.org/svn/trunk@8553 7dbf422c-18fa-0310-86e9-fd20926502f2