Closes #1002, by optimizing GetQueuedChanNum().
If there is nothing in queuedChanNum we just return it without
running the regular expression matching at all.

We also use a static regular expression which means we don't rely on the Qt regular expression cache to avoid recompiling the matching tables, this Qt feature appears to be broken or missing in the MacOSX Qt implementation.

Finally, we use minimal rather than greedy matching so the expression can be evaluated more quickly.



git-svn-id: http://svn.mythtv.org/svn/trunk@8568 7dbf422c-18fa-0310-86e9-fd20926502f2