Fixes #1005.
Makes sure the kFlagCancelNextRecording flag is cleared when needed.

Chris, I believe this is how it has always worked, overrecord is applied unconditionally to the recording. Then the overrecord time is overridden by StartRecording() when the scheduler wants to record something. This allows the overrecord to work in the presence of rescheduled or canceled follow-on recordings, without needing to update the end time. Updating the end time wasn't a feature when overrecord was first added...



git-svn-id: http://svn.mythtv.org/svn/trunk@8577 7dbf422c-18fa-0310-86e9-fd20926502f2