Fixes #743, by applying slightly modified patch.
This is to cope with some wackyness some DVB cards exhibit. Basically, if stale data is seen we wait 400 ms before issuing a warning. If in that time we see good data we just the bad data.

The problem with the drivers is that they are missing some fences which became a requirement for DVB drivers about 6 months ago. If anyone recalls the tuning problems with DVB Events, these fences were meant to do away with that problem, but the implementation lags the spec a bit...



git-svn-id: http://svn.mythtv.org/svn/trunk@8619 7dbf422c-18fa-0310-86e9-fd20926502f2