This commit adds the ability to control mythfrontend via a socket interface. Commands are sent and replies are received in text mode, there is no fancy xml interface or anything, it is meant to be able to be used via something as simple as telnet.
This feature is disabled by default, you need to turn it on via the mythfrontend
setup screen under the Default settings page and then restart mythfrontend.  The
default port is 6545, so after restarting mythfrontend, you can test by
telnetting to localhost on port 6545 ("telnet localhost 6545").  For a list of
supported commands, type "help".  For help on a specific command, type
"help COMMAND" (where COMMAND is the name of the command like "help jump").

Notes:

The integer 1x, etc. play speeds come from the FFRewSpeed settings in the
database, otherwise they use the defaults of (3, 5, 10, 20, 30, 60, 120,
180) defined in tv_play.cpp.

The pre-defined fraction speeds (1/2, 1/4, etc.) are listed in the
"help play" help text.

There can be only one socket connection to mythfrontend at a time, if a
new connection is opened, the old one will be automatically closed.

Closes #1092.  If there are any bugs found, they can be submitted under
another ticket.



git-svn-id: http://svn.mythtv.org/svn/trunk@8688 7dbf422c-18fa-0310-86e9-fd20926502f2