Fixes a ringbuffer switching problem when the stream format changes.
When OpenFile() is called due to a stream change and the ringbuffer is paused it can get stuck waiting for more data which will never come.

Also, if the stream is coming from a DTV type recorder we need to mark all discontinuous streams as being stream format changes, since the format may have changed drastically. Doing a format change while opening the file causes an additional slew of problems which this fixes by just reiniting the decoder with the new format during the ringbuffer switch.



git-svn-id: http://svn.mythtv.org/svn/trunk@8703 7dbf422c-18fa-0310-86e9-fd20926502f2