Fixes #774.
The "Extra Audio Bufferring" option was not being passed to
a decoder unless the video output needed to change. Fixing
this combined with the prebuffering fix from rob.r appears
to fix the problem entirely. I also added an override for
the extra audio buffering so that it is always applied if
you use XvMC output; it also prints a message if you do not
have the setting enabled by default and are using XvMC.



git-svn-id: http://svn.mythtv.org/svn/trunk@8769 7dbf422c-18fa-0310-86e9-fd20926502f2