References #894.
In some cases the DVBRecorder takes much longer to start up than it should (especially when 'wait_for_seq_start' is enabled, which is the default). This increases the RingBuffer retry counts until this can be properly fixed. I'm referencing #894, so I'll know to revert this once this ticket is fixed.

BTW this appears to happen when the PATs and/or PMTs are far apart, when a DVB-S rotor is involved in the tuning, when there are no GOP headers and the Sequence headers are far apart, or when the reception is very poor. Those factors are additive so the current timeouts are usually fine when one or two of these problems exist, but not if three or more exist. This is why people have been reporting tuning problems with particular channels, such as "Sky Travel".



git-svn-id: http://svn.mythtv.org/svn/trunk@8805 7dbf422c-18fa-0310-86e9-fd20926502f2