Fixes unreported segfault.
I came across this testing the atscscan import. It imported a lot
of bad channels, so I was switching from one off air-channel or
encrypted channel to the next and exiting playback frequently, and
twice it crashed. The second time I got a backtrace. It turns out
that some of the customEvent() handlers didn't check if the recorder
exists, so if you exit and get one of these events during shutdown,
things go boom. This commit adds the needed null pointer checks..



git-svn-id: http://svn.mythtv.org/svn/trunk@8858 7dbf422c-18fa-0310-86e9-fd20926502f2