References #942.
This commit moves the pthread_join on the NVP thread outside the osdlock, so that it is safe for the NVP to grab the qApp lock.

While trying to reproduce the bug in #942, I got the "Error encountered while displaying video" dialog. This was expected with what I was doing, but I didn't expect it to stop responding to input. Turns out that the NVP grabs a qApp lock when displaying this error dialog, but it does it while we are waiting for the thread to finish while holding the osdlock. If anything else is holding the qApp lock and tries to get the osdlock (which is the normal lock order), then we get into a deadlock situation between those two threads.



git-svn-id: http://svn.mythtv.org/svn/trunk@8862 7dbf422c-18fa-0310-86e9-fd20926502f2