Fixes #1231.
The MPEGStreamData stream verifier was checking the adaptation field control field in the PES packets and tossing PSIP packets that contained them, since that flag is invalid for ATSC streams. But that ADF is allowed in DVB and this class is used for both ATSC and DVB streams, so this commit removes the check.



git-svn-id: http://svn.mythtv.org/svn/trunk@8902 7dbf422c-18fa-0310-86e9-fd20926502f2