Fixes #1249
Checking the recstatus turns out to be unreliable and may not update
the oldrecorded status to rsRecorded for real recordings. Instead
borrow the approach in CheckForRecGroupChange() to verify the recgroup
stored in the 'recorded' table.

If the real recgroup is not LiveTV when it is asked to finished, then
UPDATE_RECORDING_STATUS else skip the update for LiveTV recordings.


git-svn-id: http://svn.mythtv.org/svn/trunk@8913 7dbf422c-18fa-0310-86e9-fd20926502f2