Added ViewScheduled::viewCards(). The default keybinding for VIEWCARD is "Y" which is the as SWITCHCARDS in the playback context.
On the viewschedule page, all Will Record items are highlighted in
the "record" color, usually green or orange for G.A.N.T. Pressing
VIEWCARD once will highlight only the items assigned to the lowest
numbered card. "Y" again highlights items for the next next higher
card that has assigned items and so on until the highest number is
reached then it cycles back to highlighting all Will Record items.

This is obscure to the point of being an Easter egg but can be useful
to help visualize which shows follow each other on a specific card,
for a dual input card, it is easier to see which shows from which
input occupy the card, and helps to see when it is safe to shutdown
a slave.


git-svn-id: http://svn.mythtv.org/svn/trunk@8956 7dbf422c-18fa-0310-86e9-fd20926502f2