References #1037.
Small include refactoring. This doesn't change the code very much, it just moves the CardUtil and SourceUtil functions into their own files.

This is makes recompiles after changes to videosource.h significantly faster, this file had grown to a behemoth that was included in many places. It is still a behemoth, but now it is only included in three files since most of the places it was included only needed the CardUtil or SourceUtil classes.

This also includes a small fix for the DishNet legacy switch support patch I commited a couple days ago. There was an off by one error, so that you could only select the 21 switch and only by selecting the 64 switch in the UI. I've changed the places that used raw numbers to use the enum names of the switches to prevent this type of error in the future.



git-svn-id: http://svn.mythtv.org/svn/trunk@8967 7dbf422c-18fa-0310-86e9-fd20926502f2