Fixes a minor annoyance in the Tuner Status dialog.
This adds a tiny detail to the detailed tuner status. It tells
you the device "Tuner 34" corresponds to. In that case on my
backend this adds "[ MPEG: /dev/v4l/video0 ]". This is the same
info it tells you when you are connecting up inputs so the
code for this was trivial, just call the GetDeviceLabel() with
the right info (and add 2 additional items to query).



git-svn-id: http://svn.mythtv.org/svn/trunk@8990 7dbf422c-18fa-0310-86e9-fd20926502f2