JobQueue changes:
Allow the user to specify a command to run for transcoding and commercial
flagging jobs, for instance to run a wrapper around mythtranscode or a
different transcoding program.

The command settings support all the normal %CHANID% type substitutions that
normal User Jobs support including new %JOBID% and %TRANSPROFILE% keywords.

Move the "Save original file after transcoding" setting from the Host-specific
backend setup page to the JobQueue global settings page.  This is a global
setting and I think it makes more sense with some of the other transcoding
settings pertaining to the JobQueue.



git-svn-id: http://svn.mythtv.org/svn/trunk@9006 7dbf422c-18fa-0310-86e9-fd20926502f2