References #944. Fixes disappearing OSD problem in SVN.
I'm keeping the ticket open to remind me to apply this to 0.19-fixes, if no problems are reported.

The OSDListBtnType has been edited much more heavily than needed for this fix. It needed the rewriting, I found several double free's and other potentially serious problems.

The thing I'm not sure of is the DVD menu change in the NVP. I assumed the highlights were scaled for a 720x576 window, my guess is these are actually scaled for different sizes depending on the DVD standard. Highlights weren't working for me before the change, so I couldn't even test with an NTSC DVD.



git-svn-id: http://svn.mythtv.org/svn/trunk@9021 7dbf422c-18fa-0310-86e9-fd20926502f2