Move some transcode logic/functionality out of the JobQueue into mythtranscode.
The code moved includes the code responsible for renaming transcoded files,
deleting or keeping the original copy, clearing recordedmarkup and the
cutlist and bookmark fields in the recorded table, and the status updates
for the jobqueue table.  This code belongs more in mythtranscode than in
the JobQueue, and with the new ability to run a different executable than
mythtranscode, we may not want to perform these actions in the JobQueue
itself.




git-svn-id: http://svn.mythtv.org/svn/trunk@9028 7dbf422c-18fa-0310-86e9-fd20926502f2