References #1095. Adds EIA-708 caption decoder.
This is far from a complete implementation of EIA-708.

But it does decode the plain left-to-right, scroll up captions still in use by most networks. The advantage over the the EIA-608 caption decoder already in SVN for ATSC is that this decodes additional caption streams available on some networks (PBS) and supports all unicode characters, not just an extended version of ASCII.

Some caveats:
 * This does not decode the ATSC caption descriptors, so you won't know what streams are available, and in what languages they are in.
 * This does not work when the captions are out of order, this mean FOX, UPN and others probably won't work. This is because we are not yet reordering the captions out of picture order. So only networks that use only B frames for captions like CBS, NBC and others work.

The default key to toggle the EIA-708 captions is '#', once I've resolved the caveats I'll integrate this into the 'T' binding. Defaulting to EIA-708 captions when they are available. The second caption stream is available via the menu.



git-svn-id: http://svn.mythtv.org/svn/trunk@9140 7dbf422c-18fa-0310-86e9-fd20926502f2