Things now use libmythui for the main window.  Nothing is actually drawn with mythui yet.
Remove MythMainWindow from libmyth.  Update the MythMainWindow class in libmythui for the changes since it was started.  Rename/move a few files/classes in libmythui so they don't conflict with libmyth for now.

Note the install location for the headers changed (from <prefix>/mythui to <prefix>/mythtv/libmythui).  You may want to delete the old installed headers.

A distclean may be needed to compile everything.

Refs #12.


git-svn-id: http://svn.mythtv.org/svn/trunk@9156 7dbf422c-18fa-0310-86e9-fd20926502f2