References #1133. Adds UI for configuring DViCO card's analog part.
There is still a problem with newer DVB drivers.

Since 2.6.14, the DVB drivers do not tell you which device you are talking to, so we can't tell if the current card is a DViCO Fusion HDTV5 PCI card or DViCO Fusion HDTV5 USB device or an Air2PC HD-5000 card. So if any of these cards is detected I assume that it is a DViCO Fusion HDTV5 PCI card. If someone knows a reliable way to disambiguate these cards with the latest drivers please let me know.



git-svn-id: http://svn.mythtv.org/svn/trunk@9172 7dbf422c-18fa-0310-86e9-fd20926502f2