References #1403. This hopefully fixes the EIA-708 buffer fill problem.
EIA-708 has "delay cancel" and "reset" commands which must be processed out-of-band. But we we're only catching the first of several delay cancels each time through and so letting the buffer fill-up. With this change we find the last of the out-of-band commands. Also if the buffer ever does fill up, we pretend there is a delay cancel at the end of the buffer.



git-svn-id: http://svn.mythtv.org/svn/trunk@9204 7dbf422c-18fa-0310-86e9-fd20926502f2