Closes #1422. Cleanup of CardUtil.
This also fixes the problem where DVB input connection options were always being shown, hiding the external channel changer setup.

This also adds an edit and delete binding to the DVBTransportEditor, so you don't need to use the menu to delete a transport.

This also beautifies the transport editor for ATSC and analog transports.



git-svn-id: http://svn.mythtv.org/svn/trunk@9209 7dbf422c-18fa-0310-86e9-fd20926502f2