Closes #1433, by applying patch.
This tweaks some params in the pes memory allocator to get a big bump in performace.

The method causing most of the allocations will actually be a zero memory copy routine when the SIParser refactor is done, but these optimizations will still be useful for when we need to clone tables.



git-svn-id: http://svn.mythtv.org/svn/trunk@9238 7dbf422c-18fa-0310-86e9-fd20926502f2