References #1341. Bulk of Dave Matthews' MHEG implementation.
The MythTV hooks will be added seperately and in steps so that any problems can be isolated to a changeset.

MHEG is a type of Interactive TV used in DVB countries. It includes such things as an enhanced channel guide, access to additional video streams, and games. So far the implementation only contains support for the UK profile.



git-svn-id: http://svn.mythtv.org/svn/trunk@9316 7dbf422c-18fa-0310-86e9-fd20926502f2