Closes #1485. SDT was being disabled in siscan.
We used to disable the SDT until we had seen the NIT. This was because we weren't caching the SDT's at the time and we wanted to see the UK channel number map, if it existed, before inserting any SDTs. But since we do cache the tables now this code was just causing problems without any benefits.



git-svn-id: http://svn.mythtv.org/svn/trunk@9344 7dbf422c-18fa-0310-86e9-fd20926502f2