References #368. Fixes a few problems with DVB/IVTV teletext decoder.
I have not re-enabled this decoder as the default decoder, when it is possible to run it. The code for doing this is in NVP::SetCaptionsEnabled(), but I commented it out on purpose in the last commit. First, because it doesn't decode captions (at least when you first enable it), and second because it is still very very unstable. But, this teletext decoder can be accessed by binding "NEXTCC" and pressing that key until you get past the subtitle streams.



git-svn-id: http://svn.mythtv.org/svn/trunk@9346 7dbf422c-18fa-0310-86e9-fd20926502f2