Closes #1502. Adds a workaround for broken DVB drivers recommended by DVB guy Manu Abraham.
Basically, this adds a user specified hard delay after telling the driver what transport to tune to, but before we ask if it has finished tuning yet. This defaults to off, and is configurable in mythtv-setup's "Recording Options" pane for a DVB card.

If we detect a frontend name of "DiBcom 3000P/M-C DVB-T" when initially setting up a DVB card, we auto-magically set this delay to two seconds.

"DiBcom 3000P/M-C DVB-T" corresponds to a revision of the Hauppauge Nova-T with currently broken Linux DVB drivers. Kenneth Aafloy previously reported that this would be fixed in the 2.6.12 release of Linux, but since it hasn't been fixed yet and we are at 2.6.15.6 now I presuming that fixing the tuning bug for this card has a very low priority for the DVB driver folks.



git-svn-id: http://svn.mythtv.org/svn/trunk@9349 7dbf422c-18fa-0310-86e9-fd20926502f2