Referneces #1593. Backend portion of Channel Editing patch.
This implements the GetChannelInfo() and SetChannelInfo() calls.
The first allows the frontend to query a backend recorder for some additional information about the current channel (or another channel).
The second allows the frontend to change some of the channel information, and updates that data in all the currently running recorders.



git-svn-id: http://svn.mythtv.org/svn/trunk@9524 7dbf422c-18fa-0310-86e9-fd20926502f2