References #1485. Adds a method HasCachedSDT() to address one of the scanning issues.
We were waiting for the transport to contain all the SDT's associated with the NIT, but these can of course be located on different transports. HasCachedSDT() just checks to see that we have at least one complete set of SDT sections.

Also, if we do not find all the SDTs we are waiting on. I've enabled a bit of code in HandlePostInsertion() which will insert what we do have and not insert stations based on the PAT if an NIT is found.



git-svn-id: http://svn.mythtv.org/svn/trunk@9532 7dbf422c-18fa-0310-86e9-fd20926502f2