References #1485. Adds duplitate checking code to insert_dtv_multiplex.
This is less trusting of the pre-existing transports than the patch in #1485, but should still prevent most duplicate insertions of transports. Probably for this to work better in DVB land an existing transport that didn't match the insert data exactly should be updated, but I didn't address that issue in this changeset.



git-svn-id: http://svn.mythtv.org/svn/trunk@9537 7dbf422c-18fa-0310-86e9-fd20926502f2