References #1507. Adds first transport to scan list.
This doesn't quite fix the tuned scans, but addresses the ScanWizard portion of the problem.

With this commit the first transport should be added and scanned; if followed by several scan all transports actions the scanner should find all the channels. I need to do a bit more work in HandleNIT() to get this to be fully automatic again.



git-svn-id: http://svn.mythtv.org/svn/trunk@9541 7dbf422c-18fa-0310-86e9-fd20926502f2