Fixes #927, refs #1049, refs #1552. Retunes when we fail to see video after a while (for DVB-S).
This is mainly intended to help DVB-S, where sometimes tuning fails due to DiSEqC and other hardware and driver issues.

This should be backed out and moved to DVBSignalMonitor once #1049 is addressed. At the moment we never know if we are on the 
right transport, which is especially problematic when there is a rotor involved. Once #1049 is implemented we will know, and
then we can safely move this code to the signal monitor which also means the UI won't freeze up while tuning DVB-S channels.



git-svn-id: http://svn.mythtv.org/svn/trunk@9565 7dbf422c-18fa-0310-86e9-fd20926502f2