Adds some calls to DataDirectProcessor to support channel editing after a QAM scan.
The most important new calls are GrabFullLineup() and SaveLineup(). GrabFullLineup() allows you to grab all the channels in the lineup, including the currently unchecked ones. SaveLineup() adds all the channels you pass to it to the lineup, and removes the ones not on your list.

The idea is to allow you to get all the channels when you are mapping data direct channels to your scanned channels, and also to remove the channels from your datadirect listings which you are not using after you finish mapping the channels.

There was talk on the mailing list about preventing datadirect added channels from being added to the DB. SaveLineup() could also be used for this purpose, but this is not on my TODO list in case someone still wants to work on this.

I'm not done with this, in particular I would like to standardize the temporary file creation, and maybe simplify getNextSuggestedTime() by using the new Post() function which supports the Lineup methods. But since there was expressed interest in preventing datadirect added channels from being added to your lineup I thought it would be best to make the API available before anyone went ahead and reinvented it.

I did all the testing with an external program, so there may be some uncaught minor merge errors with the new functionality, but none should be too difficult to fix if encountered. The old functionality all checked out in my tests.



git-svn-id: http://svn.mythtv.org/svn/trunk@9621 7dbf422c-18fa-0310-86e9-fd20926502f2