Some more DataDirect refactoring, improvements.
This uses some of the functions added in [9621] to get the right data for channel editing in tv_play.cpp

This also changes some of the container classes in DataDirectProcessor to make things more efficient + other refactors.

Finally, this speeds up the "Fetch channels from listings source" button in mythtv-setup by only fetching listings for one day and only on the specific sourceid we're interested in.



git-svn-id: http://svn.mythtv.org/svn/trunk@9625 7dbf422c-18fa-0310-86e9-fd20926502f2