Disables channels.conf import for ATSC capture cards in mythtv-setup's channel scanner.
I'd like to get bug reports on the channel scanner. But from looking at the wiki, it looks like people are using the import when they should be using the scanner to get all the info for MythTV to tune properly. This causes lots of problems as atscscan picks up encrypted channels and otherwise untunable channels, plus atscscan is really really slow.



git-svn-id: http://svn.mythtv.org/svn/trunk@9633 7dbf422c-18fa-0310-86e9-fd20926502f2