Merge in a mythfilldatabase fix that went into head a month ago in [9295].
"Don't bail out of mythfilldatabase when we fail to retrieve some data.
This causes issues for people who have multiple sources and an error occurs
with one source. The scheduler would not get triggered and other data fixups
would not occur because mythfilldatabase would exit if we had any errors at
all."



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-19-fixes@9665 7dbf422c-18fa-0310-86e9-fd20926502f2