Fixes #1692. Fixes the UTC->localtime conversion for ATSC + adds manual override.
The reason why cizek's patch didn't work for me was because the data I am getting OTA is incorrect. I've implemented the correct offset from cizek's patch and made it the default, and added a global EIT time offset override. The override makes things work here, but we really should have a per-source + per-channel setting if this is a common problem.



git-svn-id: http://svn.mythtv.org/svn/trunk@9764 7dbf422c-18fa-0310-86e9-fd20926502f2