References #1719. Update initial GPS to UTC offset in MythTV to current offset.
On Jan 1st, 2006 the International Earth Rotation Service added a leap second to UTC, but GPS has a fixed offset from TIA time and does not add leap seconds, hence the need for the adjustment. This is the 14th leap second added since the GPS epoch on Jan 6th, 1980.

See http://maia.usno.navy.mil/ser7/tai-utc.dat for a TIA-UTC table that goes back to 1961.



git-svn-id: http://svn.mythtv.org/svn/trunk@9782 7dbf422c-18fa-0310-86e9-fd20926502f2