Closes #1736. Refs #1682. Fixes autoselection of subtitle/caption streams when the streams are not immediately available.
Caption streams are often not detected until a few seconds into playback. This means the autoselect was usually selecting Teletext 888 or CC-1 captions rather than the more desirable options, such as a caption stream in one of your preferred languages.

This references #1682 only because I accidentally placed a version of the patch there.



git-svn-id: http://svn.mythtv.org/svn/trunk@9805 7dbf422c-18fa-0310-86e9-fd20926502f2