Fixes #1766. This fixes DVB Teletext scanning and supresses a warning.
The teletext PMT scanning would skip teletext if there was a private stream other than teletext listed before 
the teletext stream in the PMT. This patch on #1766 from jyrki.n fixes the problem.

The warning was just about skipping some filler data, which we don't need warnings of even in "-v vbi".



git-svn-id: http://svn.mythtv.org/svn/trunk@9905 7dbf422c-18fa-0310-86e9-fd20926502f2