Fixes #906. Fixes #1461. Refs #137. Merges revisions 9854:9914 in mythtv-vid to head.
This brings over fixes for XvMC running out of frames and XvMC rendering pause frames (#906).

The OSD fading (#1461) is fixed as far as I can tell. Tj NG is still reporting problems, but I can't reproduce his problem with the same setup so I'm closing the ticket.

The rewind+play wierdness in #137 is not entirely fixed, but it's different now; the immediate problem with the pause frame is fixed, but playback does not resume normally because we still throw away the I and P frames while we still need them.



git-svn-id: http://svn.mythtv.org/svn/trunk@9915 7dbf422c-18fa-0310-86e9-fd20926502f2