Fixes #1788. Refs #1787. Adds much of the code needed for H.264 (aka MPEG-4 AVC) TS decoding in MythTV.
In particular this adds it to the MythCodecID and other enums, and adds a H.264 keyframe/frame finder to avformatdecoder. You would still need to get a H.264 video into MythTV with a keyframe map for proper playback with skipping (see #1787).

The patch was submitted by Jean Michel Bouffard on behalf of the Queen of Canada and the Canadian Communications Research Centre.



git-svn-id: http://svn.mythtv.org/svn/trunk@9919 7dbf422c-18fa-0310-86e9-fd20926502f2