Fixes #1807. Crash on uninitialized OSD cache pointer in 0.19-fixes.
I was able to reproduce this (through a slightly more circuitus route). It looks like this was do to a lost change in the backport, but I also noticed that this depends on "delete NULL;" being safe. I added a null check before the delete which I'll port back to SVN head.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-19-fixes@9927 7dbf422c-18fa-0310-86e9-fd20926502f2