Fixes #1828. Adds recorder keyed picture attribute adjustments.
This fixes a long made complaint about using one source for several analog recorders, with different picture attribute defaults. This automatically uses the hardware defaults for V4L2 compatible devices. And for devices whose defaults are not reasonable, this allows you to set the picture settings on for a recorder, instead of just for a channel.

The default picture attribute keybindings are "F" for playback picture attributes, "G" for recorder picture attributes and "Ctrl-G" for channel picture attributes.

This changes the MythTV protocol, so you will need to update all connected frontends and backends.

This also gets rid of the "UseOutputPictureControls" option. I don't see any reason why you would ever want to disable picture attributes if your hardware supports it, and it does no harm to have them enabled if the hardware does not support them.



git-svn-id: http://svn.mythtv.org/svn/trunk@9968 7dbf422c-18fa-0310-86e9-fd20926502f2