Refs #1104. Removes a SoundTouch 1.3.1 change that induces clipping.
This restores the 1.3.0 behavior of passing 1.0 timestretch samples through putSamples() unmolested. The 1.3.1 version treats 1.0 as a special case of slowdown, where the slowdown is 0%. But this causes clipping for Cougar when MMX is enabled.

The clipping is probably dependent on source volume since no one else has reported it, but it's a regression and we don't experience the problem that the SoundTouch 1.3.1 is trying to avoid because we don't use the features of SoundTouch which trigger the problem.



git-svn-id: http://svn.mythtv.org/svn/trunk@10003 7dbf422c-18fa-0310-86e9-fd20926502f2