Fixes #1847. Make automatic input switching (based on channel number) work with DVB cards w/patch from Mark Buechler.
When we made DVBChannel input aware a few months we forgot aspect of this required so that you could switch to a channel on another input without explicitly switch inputs using the "C" key first. It actually worked as a side effect of the full retuning we were doing on each DTV channel change, but this broke when I added the quick intra-multiplex channel changing code.



git-svn-id: http://svn.mythtv.org/svn/trunk@10023 7dbf422c-18fa-0310-86e9-fd20926502f2