Refs #1502. Refs #1552. Backport DVB tuning delay for broken DVB drivers.
This slows down tuning when you use the Nova-T or TerraTech USB 2.0 DVB-T devices.

A more flexible version of this is in SVN head, but it requires a DB schema change. This instead hardcodes the two devices with known broken drivers.

The DVB folks do not intend to fix these bugs anytime soon, so we have to work around the bugs.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-19-fixes@10127 7dbf422c-18fa-0310-86e9-fd20926502f2