Refs #1552. Backports various tuning delays to 0.19-fixes.
This backports some tuning delays recommended by Marcus Metzler.

In particular we 10 ms sleep after every frontend set, and if we are dealing with any DVB-S device or the Nova-T DVB-T device we sleep for 300 ms or 100 ms, resp between status querying ioctl calls.

This delays are needed due to numerous bugs in the i2c bus handling in many DVB-S drivers and the Nova-T DVB-T driver.

Again, these are bugs that the DVB driver folks do not intend to fix anytime soon so we need to work around them.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-19-fixes@10128 7dbf422c-18fa-0310-86e9-fd20926502f2