Refs #906. Refs #137. Backports XvMC fixes from SVN head to 0.19-fixes.
This is slightly more hackish than the fix in head, because I didn't backport some of the NVP fixes that simplify the XvMC fixes there.

But I've made hundreds of channel changes and many skips and playback speed changes and it only gave me stuttering once vs. 10% of the time without the fixes.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-19-fixes@10143 7dbf422c-18fa-0310-86e9-fd20926502f2