Add a status char column on the proglist and previouslist pages.
NOTE:

You should update your myththemes module and must update these if
you are using a theme from myththemes.

If you are maintaining or using a third party theme, you need to
add a fourth column in the "proglist" section. Look for:

<listarea name="proglist"...

Subtract 20 pixels from the third column and add a column 4 like this:

-        <column number="3" width="420" context="-1"></column>
+        <column number="3" width="400" context="-1"></column>
+        <column number="4" width="20" context="-1"></column>



git-svn-id: http://svn.mythtv.org/svn/trunk@10158 7dbf422c-18fa-0310-86e9-fd20926502f2