This commit allow you choose what parameters to pass to ffmpeg when a re-encode is required. There are currently four encoding profiles (HQ, SP, LP, EP) to choose from which give approx. 1, 2, 4 or 6 hours of video on a regular single sided DVD. You can customise the profiles or add new ones by simply editing the xml file containing the profile parameters.
I'm not sure how accurate the bitrates are for each profile to produce the correct
file size required. If you think the bitrates are wildly off let me know.

This just enables the script to use the encoding profiles. The UI side will follow
later. For the moment there is a setting at the top of the script 
(defaultEncodingProfile) that contains the name of the profile to use when a file
needs to be re-encoded - the default setting will use the SP profile approx. 2 hours
per DVD.

git-svn-id: http://svn.mythtv.org/svn/trunk@10191 7dbf422c-18fa-0310-86e9-fd20926502f2