Closes #1598. Adds AC3/DTS passthough device setting.
This is needed because some hardware doesn't detect the output type or uses different audio connectors for different types of audio.
By default the old behaviour of sending all output to the same device is used.

This also makes the configure autodetection of ARTS support find the arts headers in Gentoo's new file location regime.

This changes the myth binary revision so a distclean and plugin recompile is required.



git-svn-id: http://svn.mythtv.org/svn/trunk@10201 7dbf422c-18fa-0310-86e9-fd20926502f2