Closes #1915. Heads -display param in non-Qt based XOpenDisplay() calls.
This grabs the "-display" param before QApplication steals it from us and saves it for later use in the context. We then use this in MythXOpenDisplay() which is used in place of XOpenDisplay() in the various places where we open the display ourselves.

This changes the binary revision so you will need to recompile the plugins.



git-svn-id: http://svn.mythtv.org/svn/trunk@10257 7dbf422c-18fa-0310-86e9-fd20926502f2