Closes #1908. Fixes SEGFAULT by adding a null pointer check to TV::UnMute().
This is on a Qt Timer and can be triggered at the wrong time during a channel change. This will fix the segfault, but there may still be a problem with using the timer like this. It is possible that sound will not be unmuted after a channel change, though reading the code it looks like the unmute is done directly as well when needed.



git-svn-id: http://svn.mythtv.org/svn/trunk@10261 7dbf422c-18fa-0310-86e9-fd20926502f2