Added "first" and "last" flags to the program table. These are set by mythfilldatabase. If programids are present, the first and last showings for each unique programid are flagged. Otherwise, the first and last showings for each unique title,subtitle,description are flagged.
For now, there are Custom Record examples for "First showing of each
episode" and "Last showing of each episode". This may be useful for
things like live sports events that have later edited rebroadcasts.
If you are only interested in the live coverage:

Rule Name: Tour de France

program.title = 'Cycling' 
program.subtitle LIKE 'Tour de France%' 
AND program.first > 0 

This should match only the first live broadcast for each Stage of
the race.

These may be useful for future features. Possibly adding a first
showing of each episode dupin option but I'm still pondering issues.
If a show is marked both first and last, this is the only showing so
possibly add a priority factor for things that are known to be shown
only once (within the period of the known listings =).

"Last" by it self can be useful for procrastinating to space out an
influx of episodes. A few times I've started recording a cable series
where there were several episodes repeated each week. By recording
the "last" showing, it spreads things further into the future as new
listings are added but will record an episode before it disappears
from the listings.

"Last" showings can move as new listings are added. There is also
an ambiguity with "first". Say an episode is shown Tue and Fri, On
Mon, the Tue showing is first. On Thur, the same episode on Fri
may now be marked as "first" but should also be marked as previously
recorded from the Tue recording.

One potential problem is resolved by the implementation. Say an MLB
Baseball game is shown on both ESPN and FSN. Because the starttime and
programid are the same, both showings get marked as "first" rather
than randomly one or the other, This way, both live showings are shown
as availabe and other factors, such as channel priority, can determine
which showing should be recorded.



git-svn-id: http://svn.mythtv.org/svn/trunk@10341 7dbf422c-18fa-0310-86e9-fd20926502f2