Fixes #1899. Fixes segfault, caused by drawing at the wrong time in the teletext menu code.
Peter Holmes found the problem and modified my initial fix so that it actually works (I'm not in teletext land). Basically this makes the teletext display class cache up any data to be drawn and draws it when in the draw method instead of just drawing to the osd surface willy nilly whenever data arrives, and adds some locking to prevent the the data update and the drawing threads from accessing the data at the same time.

This also adds the file to translate.pro, one error string was always shown in English...



git-svn-id: http://svn.mythtv.org/svn/trunk@10355 7dbf422c-18fa-0310-86e9-fd20926502f2