Refs #2011. Fixes "deadlock" when many deletes are pending.
The delete method was grabbing a DB connection before waiting for it's turn to actually delete the file, and then releasing it after the delete completed. If a large file delete is scheduled, followed by many deletes of any sized files, such as after channel flipping in LiveTV, then the number of open connections grows quickly. This can make MythTV unresponsive until the big delete completes because most MythTV actions require a DB connection and none (or too few) are available. With this change the delete function lets the DB connection go before it blocks on the delete mutex.



git-svn-id: http://svn.mythtv.org/svn/trunk@10370 7dbf422c-18fa-0310-86e9-fd20926502f2