Closes #1966. Merges mythtv-atscsrcid to head.
This splits the atscsrcid column into two columns atsc_major_chan and atsc_minor_chan to support ATSC channels in QAM which uses much larger numbers for minor channels.

This introduces a smart channel sorting method that can handle mixed system channel lineups much better. i.e. DVB-S + ATSC + NTSC channel are sorted inteligently when they are in the same lineup.

This adds code to TVRec::InitChannel() to fall back to a channel in your lineup when your startchannel does not exist in the lineup and so is not tunable. This assures that LiveTV is not broken when a channel disappears from your lineup, or the user forgets to set it after the initial mythfilldatabase run.

These changes are explained in more detail in the mythtv-atscsrcid commit logs attached to Ticket #1966. I've been running a version of this for about two weeks without serious problems (only sorting problems which have been fixed), but I'll be on the lookout for any error reports in the coming week.



git-svn-id: http://svn.mythtv.org/svn/trunk@10392 7dbf422c-18fa-0310-86e9-fd20926502f2