Fixes #1702. Refs #1648. Implements a very basic signal monitor class for the freebox recorder.
This tries once to connect to the streaming server and reports a lock if it succeeds. It then waits for a PAT and a PMT with a program number of 1.

Ideally we would try to reconnect if the connection failed the first time, and we would make rtspcomms.cpp a little more efficient.

This also adds "-v record" debugging to the freebox recorder. When I was testing I found some problems with handling error conditions. I think these are all fixed, but the additional debugging should make any additional problems easier to track down.

BTW The freebox recorder will work with any Internet TV station broadcasting a TS MPEG2 stream using RTSP. You just need to create a m3u file as described the readme on ticket #1702. H.264 (aka MPEG-4 AVC) support could be added without much work, if there is any interest in it.



git-svn-id: http://svn.mythtv.org/svn/trunk@10434 7dbf422c-18fa-0310-86e9-fd20926502f2