Fixes #1872. Fixes a tuning problem that caused failed tuning under some circumstances.
An overzealous MPEGStreamData::Reset() was removing the ATSC PSIP PID from the list of PIDs we were listening to if we got a VCT table before we got a PAT table. When we had also not gotten an MGT table and this was a channel we had never gotten an MGT on before, then we would never get a lock without switching to another channel on the same transport and then switching back to the problem channel.



git-svn-id: http://svn.mythtv.org/svn/trunk@10437 7dbf422c-18fa-0310-86e9-fd20926502f2