Fix mythtranscode exporting raw video frames from 1920x1080 video.  The video frame is actually 1920x1088 so part of the V data was getting cut off.  The full 1920x1088 size frame is now sent out the fifo.
Currently the only app I know of that uses the mythtranscode in this raw mode
with fifos is nuvexport.  Nuvexport is being patched separately by xris to
accomodate for this fix.  Once that is done, nuvexport should work with
mythtranscode's fifos on 1920x1080 video.  If any other applications use the
fifos, they will need to be modified to realize that they need to accept a
1920x1088 sized video frame when transcoding 1920x1080 video.

I was told there may be a ticket for this, but I've searched twice and haven't
seen it, so I'm assuming it was closed already if it existed.



git-svn-id: http://svn.mythtv.org/svn/trunk@10578 7dbf422c-18fa-0310-86e9-fd20926502f2