Fixes #1848. Refs #2053. Allows DTV cards to be hooked up to DataDirect without manual editing again.
There is still some work to do to prevent mythfilldatabase from mucking with the datadirect lineups when it doesn't need to so I'm keeping #2053 open.

The only time you really need to touch the datadirect lineups are:
 * right after a digital channel scan to get the xmltvid matches
 * when datadirect tries to add channels to a digital source (so you don't download listings data for an untunable channel).
 * when the user explicitly asks you do so to with --remove-new-channels and there are actually new channels to remove.
 * when you start up the frontend channel editor, and sometime after you leave it if xmltvid changes are made.

The code is still touching the lineups twice every time mythfilldatabase is run for digital sources, even when every channel has an xmltvid, and after any time you run the frontend channel editor (whether changes were made or not).



git-svn-id: http://svn.mythtv.org/svn/trunk@10612 7dbf422c-18fa-0310-86e9-fd20926502f2