Fixes #1783. Fixes the last stream data leak.
This is in the same piece of code from [10188] that caused segfaults and needed to be reverted in [10238]. I've added a check to make sure the MPEGStreamData is not still in use before deleting it. I've been running this for a while without any segfaults here, and Stuart A has confirmed that it fixes the leak with Valgrind.



git-svn-id: http://svn.mythtv.org/svn/trunk@10613 7dbf422c-18fa-0310-86e9-fd20926502f2