Fixes #1076. This just makes MythTV stop trying to write to a file once it has it a file size resource limit.
The signal is ignored rather than doing a coredump, and a useful error message is printed out once for each file which hits it's resource limit. This can be triggered when the mythbackend user has a ulimit set, if the c library is configured for 32 bit file pointers, or if the filesystem does not support "large" files.



git-svn-id: http://svn.mythtv.org/svn/trunk@10633 7dbf422c-18fa-0310-86e9-fd20926502f2