Refs #2087. Removes NTSC Standard Cable Channel 1.
This channel was removed from the broadcast standards in 1945 (on which this portion of the cable channel assignments is based).

According to what I have found via google, it has not been used since the 1970's.

This applies "mythtv-digital_cable-remove_channel_1-before_fix_standard.patch" and "mythtv-ntsc_cable-remove_channel_1.patch" 



git-svn-id: http://svn.mythtv.org/svn/trunk@10648 7dbf422c-18fa-0310-86e9-fd20926502f2