Fixes #2101. Allow EIT data to be collected when the section reader is in use by the DVB signal monitor.
I'm not going to apply the other patch to use the section reader exclusively. The section reader code can not fully support both ATSC and DVB without being made significantly more complicated, and does not support "tuned" DVB channel scanning. It also does not work with some of the DVB-S and DVB-C broadcasters which improperly label their NIT and SDT tables as "other NIT" and "other SDT" tables.

The section reader as written is a functionally limited fallback for problematic USB devices, extending it to be fully functional would need to be a 0.21 project.



git-svn-id: http://svn.mythtv.org/svn/trunk@10651 7dbf422c-18fa-0310-86e9-fd20926502f2