Optimize upcoming lists for existing search rules. Include upcoming lists for both per title and per rule in the Schedule Information section of the recording options page.
Added plRecordid in proglist to lookup showings associated with a rule
in recordmatch. Existing complex search rules used to have to run the
complex query for each proglist request. However, the match results
are already known in 'recordmatch'. By using these match lists, even
the most complex rule can return the upcoming list immediately.

Because there is now a plRecordid and lists of just the matching
showings can be useful for non-search rules such as Channel or
Weekslot, I've put "Upcoming episodes for this title" and "Upcoming
episodes for this rule" in the Schedule Information section. I also
renamed the misleading "List previous episodes" to "Previously
scheduled episodes". This is not a list of all episodes that have
appeared in the listings but is a list of entries in the oldrecorded
table that match this title or rule.


git-svn-id: http://svn.mythtv.org/svn/trunk@10689 7dbf422c-18fa-0310-86e9-fd20926502f2