Fixes #2084. Fixes #2103. Fixes some mythfilldatabase problems with respect to inserting channels.
For #2084, this passes the cardtype in to mythfilldatabase when it is called from withing mythtv-setup, this is needed because at this point the connection between the card input and the source has not yet been saved to the database if the user has changed the source since entering the configuration pane. This also prevents mythfilldatabase from running in mythtv-setup for scan required sources if there are no channels in the lineup. This is not strictly needed after the fix, but it prevents us from wasting time doing this and making the UI unresponsive in the meantime.

For #2103, this creates a new class of DTV devices for Firewire and DBOX2, "unscanable" devices, this is for devices which should be scanned but we do not yet have the code to scan them.



git-svn-id: http://svn.mythtv.org/svn/trunk@10698 7dbf422c-18fa-0310-86e9-fd20926502f2