Fixes #2126. Expands startchan and tunechan in cardinput to match channum varchar(10) in channel table.
The startchan change avoids getting stuck on digital cards which don't support signal monitoring, and the tunechan has been expanded for consistency.

I did not make the suggested change to allow direct entry of channels larger than 99999, for the same reason as when I expanded the channum varchar. This would makes channel entry annoying for the rest of us, and 100 thousand channels should be enough for anyone :)



git-svn-id: http://svn.mythtv.org/svn/trunk@10720 7dbf422c-18fa-0310-86e9-fd20926502f2