Expanded the Not Listed status to include Timeslot and Weekslot rules.
This only applies to daily or weekly rules where there are no matching
showings in the TV listings.  It is normal that a kTimeslotRecord may
not match every day such as weekends or The Daily Show on Friday.  A
kWeekslotRecord or kTimeslotRecord might match reruns only. However,
if there are zero matching showings for any of the dates and times,
one rsNotListed will be added to the reclist for the next usual time
slot. This will warn the user that the show may have been cancelled or
that the time, title, or channel info has changed.


git-svn-id: http://svn.mythtv.org/svn/trunk@10757 7dbf422c-18fa-0310-86e9-fd20926502f2