Fixes #1193. Work around bug in Qt QRect(QPoint, QPoint) constructor properly.
Janne discovered that this Qt constuctor inexplicably adds 1 to the x and y of the lower right point. This removes the overscan hack we used before and uses a non-broken QRect constructor instead.

Removing the old hack fixes #1193 (OSD offset for large overscan values) and this makes the non-overscan case pixel exact so we don't lose the last column of the OSD when not using overscan.



git-svn-id: http://svn.mythtv.org/svn/trunk@10763 7dbf422c-18fa-0310-86e9-fd20926502f2