Fixes #2177. Hides channels hidden in EPG when channel flipping.
I'm not sure if this really is better. I read the arguements from last year and restored the behaviour that was implemented then. However, if you don't like it there is a boolean "skip_non_visible" in channelutil.cpp, if you flip that these channels will only be hidden in the EPG and not in channel flipping.



git-svn-id: http://svn.mythtv.org/svn/trunk@10773 7dbf422c-18fa-0310-86e9-fd20926502f2