Refs #1866. Adds FindBestMplexFreq.
This add some fuzzy matching to the frequency for dvb-t
systems. Previously we would ask the card to tune on a
frequency from the frequency list. This would often not
be the actually frequency that the mplex is on, but one
of the offset frequencies, and the card just managed to
pull itself together and get lock.

The NIT will have populated that DB with the mplexes
and their frequencies. So we can check the currently
tuned frequency against what the NIT provided. If this
frequency equals either the centre frequency or one of
the offset frequencies then we use the value we already
have. If we don't already have a frequency or the frequency
isn't one of the offsets (as would happen on a repeater mux)
then just use the frequency we tuned to.



git-svn-id: http://svn.mythtv.org/svn/trunk@10847 7dbf422c-18fa-0310-86e9-fd20926502f2