Fixes #1970. Adds pid listening priorities. This allows recording to work for cards with very limited pid filtering capabilities.
Tuning to a new channel on the same transport may take a little longer if you use one of these cards, but it should still be faster than tuning in 0.19. Most of the time, tuning should be just as fast since we still monitor other "low priority" pids, we just visit them in a round-robin fashion so the tables are not updated as quickly.



git-svn-id: http://svn.mythtv.org/svn/trunk@10860 7dbf422c-18fa-0310-86e9-fd20926502f2