Refs #2202
dvdnav_get_active_spu_stream may return a stream number with the eigth bit set as a visibility flag. The actual stream number is the first 5 bits (<32) so we just mask out the rest.

This fixes the most common cause of #2202 but there is a second problem with some dvds which I'm still trying to figure out.


git-svn-id: http://svn.mythtv.org/svn/trunk@10870 7dbf422c-18fa-0310-86e9-fd20926502f2