Fixes some of the closed captioning breakage caused by [10702].
 * Fixes a logic error that caused the EIA-708 decoder to only be called on EIA-608 data that failed the checksum check.
 * Removes global variable.
 * Makes checksum table init threadsafe.

Greg, this checksum code should be removed entirely after 0.20, the EIA-608 checksum should be done in the 608 decoder. If for some reason the checksum there isn't working on your recordings the existing checksum routine should be fixed. My guess is CC data is getting confused with XDM data, you're extra layer of EIA-608 checksumming is probably just tossing some of the XDM data before it can get missinterpreted as CC, but of course this messes up the XDM decoding.



git-svn-id: http://svn.mythtv.org/svn/trunk@10930 7dbf422c-18fa-0310-86e9-fd20926502f2