Don't reset the recstartts when we start a pseudoLiveTVRecording.
Normally LiveTV recordings have their recstartts set to the exact time,
to the second, that the recording starts.  For pseudoLive recordings
this is bad because it can cause the scheduler to lose track of the
recording since the recstartts in the recorder (and starttime in the
database) do not match the recstartts that is in the scheduler for
that recording.

This fixes #2038.  The Watch Recordings screen will now properly highlight
these psuedoLive recordings as in-progress.

References #1371 because #1371 ticket mentions the same issue as #2038
among other things.



git-svn-id: http://svn.mythtv.org/svn/trunk@10951 7dbf422c-18fa-0310-86e9-fd20926502f2