Bugfix for the slow truncating delete code.  Estimate the filesize using a different formula that doesn't use st_blocks.
Patch by Michael Dean.  I slightly modified it to put the decrement after the
truncate instead of having to have a priming increment before the while loop.

Closes #2242.

I'm not sure why this code even needs to estimate the file size.
It's not going to make that much of a difference whether we're truncating
to even block boundaries or not so I don't see much difference in using the
real size versus an estimated size (which in this case is the real size
rounded to the next lowest block).  We could just set fsize = buf.st_size
and be good.



git-svn-id: http://svn.mythtv.org/svn/trunk@10956 7dbf422c-18fa-0310-86e9-fd20926502f2