Add a rsFailed recording status and use this when the recorder fails to tune in the channel.
When a recording fails to start, the scheduler is notified and the recording
metadata and any file created are deleted.  The oldrecorded table retains
the history that the recording was tried but had a recstatus of rsFailed (-9).

This should really fix #2244.  Once it has some time in head, I'll merge it
over to -fixes since this is more invasive than my original non-fix.



git-svn-id: http://svn.mythtv.org/svn/trunk@11211 7dbf422c-18fa-0310-86e9-fd20926502f2