Add dispatch() and dispatchNow() wrappers to MythContext so that we can print out events for debugging purposes.
"-v network" now also prints out MythEvent messages since we are normally only
able to see these when they are sent over the network.  This allows us to
see them on the master backend as well now.



WARNING!!!  This modifies MythContext and the MYTH_BINARY_VERSION so it
REQUIRES a "make clean" of mythtv and mythplugins.






git-svn-id: http://svn.mythtv.org/svn/trunk@11212 7dbf422c-18fa-0310-86e9-fd20926502f2