Carry over [11262] from head.  References #2368.
Commit log for [11262]:

In ProgramInfo::StartedRecording, we need to check more than the chanid and starttime when copying the program info from the program table to the recordedprogram table. This is because a manually scheduled recording can mask a normal program listing, but both are still in the program table.

This mod makes StartedRecording check the program title in addition to the chanid and starttime when copying data to make sure we are copying the right data. Previously we'd copy the wrong data which caused duplicate rows to appear on the Watch Recordings screen for the manually scheduled recording.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11263 7dbf422c-18fa-0310-86e9-fd20926502f2