Carry over [11211] from trunk.  Here is the original commit log:
====
Add a rsFailed recording status and use this when the recorder fails to
tune in the channel.

When a recording fails to start, the scheduler is notified and the recording
metadata and any file created are deleted. The oldrecorded table retains
the history that the recording was tried but had a recstatus of rsFailed (-9).
====



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11331 7dbf422c-18fa-0310-86e9-fd20926502f2