Close #2472 - Add a thumbnail offset and bookmark image option to MythArchive.
Apply a slightly modified version of the patch on the ticket. It allow 
for a thumbnail offset to be used when grabbing the thumbnail image used
on the main menu and also for the first chapter menu image. You can also 
tell the script to use the preview image created for recordings if available.

There's a couple of settings at the top of the script - 
'thumboffset' (defaults to 10 seconds) and 
'usebookmark' (defaults to True).




git-svn-id: http://svn.mythtv.org/svn/trunk@11361 7dbf422c-18fa-0310-86e9-fd20926502f2