Fixes #2238. Workaround for Linux kernel bug in v4l-compat.
The problem is that the Linux kernel doesn't properly report the presence or absence of audio capture hardware when using v4l-compat. But if you try to get the volume level and mute status it will return it if audio hardware is present. So the workaround is fairly simple, we just check the volume first and use the sometimes flaky flag to decide whether a failure should be reported as an error or a warning. If the kernel says there is audio hardware we report an error when the audio get fails, otherwise you need to run with "-v audio" to see the error message (since this will fail normally with most frame-grabbers).



git-svn-id: http://svn.mythtv.org/svn/trunk@11376 7dbf422c-18fa-0310-86e9-fd20926502f2