Some 'watched flag' fixes and tweaks.
Set the flag when playback is exited by reaching the end of the recording or when exited through the network control interface.

Don't try to set the flag when exiting a video as it won't work (yet).

Reduce the % used in the calculation to 14 and the minimum time to 4 minutes. Feedback would be welcome on how well this works.


git-svn-id: http://svn.mythtv.org/svn/trunk@11381 7dbf422c-18fa-0310-86e9-fd20926502f2