Refs #2087. Finetunes Cable HRC table.
Our HRC table entries are a few kHz off. This corrects for that with a patch from mtdean at thirdcontact com.

Generally the lower channels were within the rounding error of the bttv and ivtv drivers, but the error accumulates so that at 125 it is three bttv/ivtv finetune ticks from the correct value. If you are using HRC and are not finetuning the upper channels this should improve your analog reception a bit. If you were finetuning, you should clear for your HRC channels.



git-svn-id: http://svn.mythtv.org/svn/trunk@11390 7dbf422c-18fa-0310-86e9-fd20926502f2