Fixes #2109. Adds support for AC3 passthrough audio with OSS drivers.
This sets the format flag appropriately, since many filters & decoders don't detect digital data robustly.

In addition to the code that was added in [10718] and reverted in [10762] this adds a sanity check to make sure the drivers support AC3 passthrough before setting the passthrough bit. Some drivers support AC3 passthough, but don't support the official OSS API for it and die a horrible death when you set the AC3 bit.



git-svn-id: http://svn.mythtv.org/svn/trunk@11392 7dbf422c-18fa-0310-86e9-fd20926502f2