Adds a frequently asked for feature. Allow you to change the final order that files will be shown on the DVD title menu.
On the last page of the create DVD wizard you can now change the order of 
the items in the selected items list. Press 'Select' on the item you want 
to move. Use the up/down arrows to move the item to its new position. Press 
'Select' again to exit move mode.

git-svn-id: http://svn.mythtv.org/svn/trunk@11402 7dbf422c-18fa-0310-86e9-fd20926502f2