Refs #1095. Copies ATSC Caption descriptor from VCT to PMT in DVBRecorder.
Terrestrial ATSC is not required to have a caption descriptor in the PMT, this copies it over from the VCT when it isn't in the PMT. Since we don't save the VCT in our DVB recordings we need these descriptors in the PMT so we can inform the user of the language and other properties of their caption streams.

This mostly effects PBS, other networks already copy the caption descriptor to the PMT before transmission.



git-svn-id: http://svn.mythtv.org/svn/trunk@11403 7dbf422c-18fa-0310-86e9-fd20926502f2