Fixes #1771. This allows the "All Programs" list to have a different order from other program lists by applying patch from John P. Poet.
The main benefit of this is to allow you to see the latest recorded program in "All Programs" while still having series recordings ordered however you want.

The only non-formatting change from the patch is to use a little simple math to set listOrder & allOrder from DB field instead of a switch.



git-svn-id: http://svn.mythtv.org/svn/trunk@11423 7dbf422c-18fa-0310-86e9-fd20926502f2