Fixes #2453. Backports HDHomeRun QAM change to 0.20-fixes.
Just before the HDHomeRun was released to the public the firmware tuning API changed. The MythTV code was ready for this change, this just changes the tuning set so that the modulation and frequency is reported to the device instead of the 8-VSB frequency index.

All production HDHomeRun devices are compatible with the change, but engineering samples will need the latest firmware.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11424 7dbf422c-18fa-0310-86e9-fd20926502f2