Add the ability to have a 'Preserved' flag icon on the Watch Recordings screen. This icon will show whenever a recording is preserved as long as the theme supports the preserved icon.
No themes are being modified with this patch because there already a large
number of flag icons.  An example of using the preserved flag is in the
blue theme's ui.xml file.  Theme designers can decide for themselves whether
to use the preserved flag icon in their theme design.  Users wanting to use
the flag in an existing theme can user their own oopy of the theme ui.xml file.

Closes #2124 by applying most of patch submitted by Rich and updated by Daniel.



git-svn-id: http://svn.mythtv.org/svn/trunk@11455 7dbf422c-18fa-0310-86e9-fd20926502f2