Refs #2556. Tightens the criteria for a private stream to be declared as a video stream.
Now a private stream is considered to be a video stream
only if the first descriptor is the component tag (0x52)

The component tag isn't actually an indicator of a video
stream, as it's present on all of the streams. Normally
it is the last descriptor on a stream, with the other
descriptors identifying the actual content.

The streams we are now declaring as video stream have
the component descriptor as the first, and generally
only descriptor.

Previously any stream marked as a private stream (type 0x6)
would be marked as a video stream if it could not be 
identified as anything else.

This should allow people having trouble with Channel 4
on various muxes to be able make recordings without
bogus streams.



git-svn-id: http://svn.mythtv.org/svn/trunk@11510 7dbf422c-18fa-0310-86e9-fd20926502f2