Fix some typos, formatting, trasnlation and unnecessary changes from 11523.
Reverted the wording on the exit options, for example "Bookmark
This Position" says nothing about exiting playback and there is
no reason to retranslate all the existing messages.

For the delete options, 're-record' is now second so the default
is to confirm the normal delete request. This is not the same order
as the GUI popup. The default in this context is to confirm and
if the order of the choices were to be switched, the focus would
need to be set to the correct item rather than the first option.


git-svn-id: http://svn.mythtv.org/svn/trunk@11527 7dbf422c-18fa-0310-86e9-fd20926502f2