Fixes #2547. Adds outline to EIA-608 captions when they are not on a black background and adds a bold font option for better visibility with low resolution streams.
I grabbed the bold font from VLC because I couldn't source the font in the patch.

I didn't use the rest of the patch because it would have enabled outlines even when the captions were on a background, disabled font selection for the EIA-608 captions and Teletext (where the new bold font will not work), and would have unconditionally added a shadow which didn't look good with my HDTV test streams.



git-svn-id: http://svn.mythtv.org/svn/trunk@11557 7dbf422c-18fa-0310-86e9-fd20926502f2