Fixes #2541 in the 0.20-fixes branch. This backports [11521] to allow tuning raw MPEG channels in HDHRRecorder.
From the original commit message:

The HDHomeRun recorder classes were originally based off the HDTVRecorder (at least partially). The HDHRRecorder only tunes ATSC streams. Since most cable operators have many channels which do not contain ATSC info this means that QAM tuning is only likely to work with HDTV channels, assuming the cable operator is obeying the law (a big assumption in some markets). This copies over some of the pertinent bits of code from the DVBRecorder to the HDHRRecorder, and makes hdhrchannel look at the MPEG Program Number if the ATSC channel info is missing.

The change is based on the patch submitted by jwestfall. 



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11561 7dbf422c-18fa-0310-86e9-fd20926502f2