Fixes #2143 in 0.20-fixes branch. This fixes the "Auto" time-offset in ATSC EIT.
Unkike the original changeset in SVN Head [11430], this may manual intervention to change the offset from the manually set time-offset in mythtv-setup to "Auto" again (assuming you set this override in the past).

If you are enabling ATSC EIT for the first time, "Auto" is the default and you will not have to do anything.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@11563 7dbf422c-18fa-0310-86e9-fd20926502f2