Refs #2567. Fixes the ProgramInfo::showDetails() ratings string.
showDetails was only showing the first ratings string that showed up in the ratings query, but often this is just one of the advisory ratings, rather than something more important like the VChip or MPAA rating. This changeset adds a static function which creates concatinates the ratings strings with the advisory ratings always appearing after the VChip, MPAA, ABA, and similar ratings systems.

git-svn-id: http://svn.mythtv.org/svn/trunk@11620 7dbf422c-18fa-0310-86e9-fd20926502f2