Fixes #2316. Adds checks of dvb-utils imported channels before adding them to the DB.
This also moves the dvb-utils scanner output parsers from dvbtypes.{h,cpp} to dtvconfparserhelpers.{h,cpp} so that dvb-utils channel maps can be imported for non-DVB cards like the HDHomeRun without compiling MythTV with DVB support.

This uses the list of channels generated by the dvb-utils channels.conf parser to seed the internal channel scanner. It then scan the inserted transports like it does normally, except that before inserting channels it checks the parser output and uses it's name for the channel rather than whatever the scanner picked up. Any channels we find on the same transport as a channels.conf channel are reported but not added to the database.

This also avoids use of the dvb-utils channels.conf parser's DB manipulation routines which were causing the problems reported in #2316, notably channels inserted with the same channel id as an existing channel.

git-svn-id: http://svn.mythtv.org/svn/trunk@11682 7dbf422c-18fa-0310-86e9-fd20926502f2