Transport editor cleanup.
This is basically a rewrite of large parts of the transport editor. We now use the pre-existing classes that do the same things instead of re-implementing everything. This also fixes a bunch of bugs, you can now create transports if none existed previously, you can now edit the modulation of DVB-T transports, we now print the correct Hz/kHz label for the frequency in all cases, we prevent editing the transport if it is connected to two completely different card types (a PVR-250 and a DVB card for instance).

git-svn-id: http://svn.mythtv.org/svn/trunk@11747 7dbf422c-18fa-0310-86e9-fd20926502f2