Add a new QUERY_RECORDING backend command that can be used to grab the ProgramInfo for a single recording using either the baseneme of the recording or the chanid/starttime.
To use the QUERY_RECORDING command, you would issue a command similar to one
of the following:

QUERY_RECORDING BASENAME 1003_20061117210000.mpg

QUERY_RECORDING TIMESLOT 1003 2006-11-17T21:00:00

In the event of an error or the recording can not be found, "ERROR" is
returned.  If the query was successful, the word "OK" is returned followed
by a single ProgramInfo in Myth's normal over-the-wire format.

The MythTV Protocol version is not being changed because this command is
optional since it is not used by Myth internally at all.  It is being added
to SVN for use by MythWeb and the perl bindings.




git-svn-id: http://svn.mythtv.org/svn/trunk@11794 7dbf422c-18fa-0310-86e9-fd20926502f2