Add a new MythZoneMinder plugin.
ZoneMinder is a system for monitoring cctv cameras. It uses a web interface
which doesn't work very well on a MythTV only system so the plugin provides
a remote friendly interface to some of the functions provided by ZM.

Please read the wiki carefully to make sure you are aware of the limitations 
of the current version. If there is enough interest I will look at adding new
features.

You'll need to re-run the configure script to enable the plugin.

There's a lot more info on the wiki :-
www.mythtv.org/wiki/index.php/MythZoneMinder


git-svn-id: http://svn.mythtv.org/svn/trunk@12144 7dbf422c-18fa-0310-86e9-fd20926502f2