A few AutoExpirer fixes required for yesterday's Storage Groups patch.
Since the Master is in charge of Storage now and slave backends don't run
the AutoExpirer, we can't reference the expirer on a slave.  This was causing
segfaults on slaves with truncating delete turned on because
MainServer::m_expirer was initialized to NULL.

The expirer on the master needs to know where any truncating deletes are
occuring, so truncating deletes are now flagged in the inuseprograms table
while they are running.  The expirer looks for any in-progress truncating
deletes and will skip checking filesystems that have running deletes until
the trunating delete is finished.  The expirer ignores stale inuseprogram
entries.

Since slaves don't have an expirer to ask about the local encoders, just set
the truncating rate based on the number of capturecards configured.  If this
turns out to be a problem, we could setup a mechanism for the slave to ask
the master how many capturecards can record to this filesystem, but that is
probably not necessary.



git-svn-id: http://svn.mythtv.org/svn/trunk@12167 7dbf422c-18fa-0310-86e9-fd20926502f2