Fixes #2756. Refs #2695. Allows scanner to work properly when DiSEqC input is connected to a source which is also connected to another DiSEqC input.
Before the refactors in #2695 we did not know which input we were scanning so each input had to be connected to a different source, even if the two sources contained the same channels. Now we pass the inputname to the scanner so that it knows which input it should scan if a source is connected to multiple inputs.

git-svn-id: http://svn.mythtv.org/svn/trunk@12183 7dbf422c-18fa-0310-86e9-fd20926502f2