Add a thumbnail image finder to MythArchive.
Allows you to choose exactly what thumbnail images to use for the main menu
and chapter menu images.

If the file is a MythTV recording and there is a cutlist available and you 
tell MythArchive to use it the thumb finder will skip over the commercials 
when selecting the images.

Accessed from the last page of the 'Create DVD' wizard by pressing INFO when
the file you want to find thumbs for is selected in the list.

It works fine with mpeg2 files but other file formats may need a little more
work to get it working properly.

Requires MythTV r12194 from trunk to be installed before compiling MythArchive.

More info on the wiki page:-
http://www.mythtv.org/wiki/index.php/MythArchive#Using_the_Thumb_Image_Selector

git-svn-id: http://svn.mythtv.org/svn/trunk@12195 7dbf422c-18fa-0310-86e9-fd20926502f2