Close #2838 by applying the patch.
Allows the time format sent to the MythShutdownNvramCmd to be specified
rather than using a hardcoded format. Should allow for using commands other
than nvram-wakup to set the wakeup time in the BIOS.

I added a DB update to change the existing MythShutdownNvramCmd settings to 
use the new format so everything should work just the way it was before this 
patch. Also updated the README.


git-svn-id: http://svn.mythtv.org/svn/trunk@12374 7dbf422c-18fa-0310-86e9-fd20926502f2