Refs #2695. Removes some unused parameters in DiSEqC setup.
These were added to try to simplify DiSEqC setup, but it turned out that fine tuning a DiSEqC setup is a common operation for FTA hobbists so I removed the simplifying code. This removes the paramaters that were used to enable and disable the simple input connections panel but are now unused.

git-svn-id: http://svn.mythtv.org/svn/trunk@12419 7dbf422c-18fa-0310-86e9-fd20926502f2