Rewrite the MythZoneMinder plugin to use a client server architecture. The server runs on the machine where ZoneMinder is running and accepts connections from any frontend running the MythZoneMinder plugin.
The client(s) and server communicate over a socket so they can be on the 
same machine or on different machines as long as there is some sort of 
network connection between them. If you run the server as either root or
as the same user that ZM uses (usually the same as the apache user) then 
there is now no need to patch the ZM source you can use a stock 
installation of either version 1.22.2 or 1.22.3.

Since a few people have indicated to me that there ZM machine is a separate 
server not running Myth at all the server has been designed to be as lean as
possible and doesn't need Myth or even QT to be installed on the server 
machine.

You will need to run ./configure in the plugins directory to get things working 
(./configure --previous should suffice).

You will also need to go into the MythZoneMinders settings page to make sure 
the IP address and port settings are set to point to your server machine.

There is a README in the mythzmserver directory with some additional info on 
how to install/setup the server and I will update the wiki page later.

git-svn-id: http://svn.mythtv.org/svn/trunk@12437 7dbf422c-18fa-0310-86e9-fd20926502f2