Refs #1777
Adds support for the xmltv --capabilities and --preferredmethod arguments to mythfilldatabase.

This removes most of the hard-coded support for specific xmltv grabbers. I expect there to be teething trouble with this change. Please keep discussion to the -dev mailing list and not in Trac.

Note: This commit does not include the support for the cache capability, it may or may not be added later.


git-svn-id: http://svn.mythtv.org/svn/trunk@12499 7dbf422c-18fa-0310-86e9-fd20926502f2