Call the ChannelWizard to update info from the channel priorities page.
The TV Setting->Recording Priorities->Channel Priorities page
lists all channels in the channel table. Pressing "SELECT",
"MENU", "INFO", "CUSTOMEDIT", bumping or glaring at the keyboard
will bring up the ChannelWizard for the selected channel. These
are the same pages as you see when using mythtv-setup's Channel
Editor then selecting a specific channel. The advantage is that
you can make minor updates to channel information from within
mythfrontend without having to run mythtv-setup.

This does not allow deleting or adding channels which should be
done with mythtv-setup. Changes are immediately reflected on the
channel priorities page and exiting this page runs the scheduler
so changes that affect the schedule will take affect when you are
done.


git-svn-id: http://svn.mythtv.org/svn/trunk@12508 7dbf422c-18fa-0310-86e9-fd20926502f2