Fixes bug in PMT stream normalization for HDHomeRun.
We were looking at the input PMT for the stream data since it is absent in the PMT we write to disk, but this means we need to find the correct index in the input PMT in case we are not recording all the streams. This moves the normalization code into the mpeg stream data and mpeg tables code and fixes the bug.

git-svn-id: http://svn.mythtv.org/svn/trunk@12564 7dbf422c-18fa-0310-86e9-fd20926502f2