Fixes #3004. Fixes race with "Watch as it records".
If scheduler wants to use the recorder you are watching LiveTV on and you decide to watch as it records, there can be a race with normal end of program ringbuffer switching. If the program you are watching ends while the scheduler is running the code in StartRecording and is on line 463 (the event flush), then the ring buffer switch for the next program will be started, but if the recorder is a digital recorder and has not found a keyframe before StartRecording gets to line 533 then the ringbuffer change for the scheduled recording will partially overwrite the in flight ringbuffer switch.

This change avoids this race by moving the delete of the pendingRecording until after the new scheduled recording is setup so that the normal end of program ringbuffer switching is not triggered while StartRecording is running.



git-svn-id: http://svn.mythtv.org/svn/trunk@12675 7dbf422c-18fa-0310-86e9-fd20926502f2