Fixes #2954. Various updates to ivtv recorder.
This makes sure that closed captions work with driver with revisions 0.2.0 through ~0.10.0.
This removes recording modes that don't work from the configuration UI, these are MPEG2-TS and MPEG-x layer 2 audio.
This makes full NTSC/PAL resolution the default for IVTV cards (only), the ivtv temporal filter only works at this resolution and has been reported to greatly improve recording quality.
This removes the --enable-ivtv-sys-header configuration option, this wasn't working correctly, is a pain to maintain and is only potentially useful for a very small number of ivtv driver developers.

git-svn-id: http://svn.mythtv.org/svn/trunk@12740 7dbf422c-18fa-0310-86e9-fd20926502f2