Refs# 2655. Adds a new duplicate checking method "subtitle then description".
This uses the subtitle for the dup match if there
is a subtitle. If the subtitle is empty then the
description is used to do dup matching.

I haven't added the new method to mythweb, i'll
leave that for xris to do.

This also adds a new string that will need
translation.

Thanks to Martin Long for the original patch


git-svn-id: http://svn.mythtv.org/svn/trunk@12808 7dbf422c-18fa-0310-86e9-fd20926502f2