Refs #2557
Instead of deleting the AudioOuput class every time we change track, re-use the same instance until we leave mythmusic.

This may or may not have an effect on the SIGSEGV a number of people have experienced when using mythmusic. I doubt it's the direct cause of the problem, but may reduce how often it occurs and will probably allow some better backtraces.

Other benefit are that mute won't get turned off on track changes and the transition between tracks seems a little faster.

Original patch by PaulH with some fixes to make it work.



git-svn-id: http://svn.mythtv.org/svn/trunk@12872 7dbf422c-18fa-0310-86e9-fd20926502f2