Fix bogus original air dates stuffed into the recorded table.
Not all programs have original air dates. The absence of a
date does not mean that this is the first date a program has
ever been shown and a quarter of all television programs
ever produced were not all premiered on New Years Eve '70.

The hasAirDate flag will be set if the date is blank, null,
year zero, before the history of television or whatever. This
flag should always be checked before displaying or relying on
the date stored in the originalairdate. If this flag is not
set, the date is unknown and whatever garbage date exists in
the field is meaningless.


git-svn-id: http://svn.mythtv.org/svn/trunk@12906 7dbf422c-18fa-0310-86e9-fd20926502f2