Add the ability to save a PNG file of the currently displayed video frame by using the SCREENSHOT keybinding or the "play save screenshot FILENAME WxH" command in the Network Control socket interface.
When using the SCREENSHOT keybinding, files will be saved in the user's
~/.mythtv/ directory as CHANID_STARTTIME_FRAMENUMBER.png, so if you take
multiple screenshots, you will end up with multiple files.  The Network Control
interface allows specifying a specific filename and optionally an image size.

NOTE: This does not save the overlayed OSD in the png file, it only saves
the actual video frame, so if you are trying to make screen captures when
designing themes, you'll still have to use the old "NO_XV=1 mythfrontend"
trick.



git-svn-id: http://svn.mythtv.org/svn/trunk@12908 7dbf422c-18fa-0310-86e9-fd20926502f2