On the 'Select Recordings' screen check that all the required UI types can be found and bail out if not.
Also check that after processing the recording list that at least one
recordings is available. Add some debugging to show that a recording
has been rejected because it's not available locally. MythArchive needs
direct access to the files so wont work with files that have to be 
streamed to it.

Refs #3158.

git-svn-id: http://svn.mythtv.org/svn/trunk@12952 7dbf422c-18fa-0310-86e9-fd20926502f2