Fix a problem with the current shutdown code in the scheduler where if for some reason ShutdownServer() fails it gets into a limbo state and never recovers.
It will now wait for 60 seconds for the shutdown to complete and if it doesn't
puts a warning in the log, resets the idle time and the shut down flag so
the shutdown sequence will start again from the beginning.

It also adds some additional debugging to help to see what's going on during
the shutdown sequence.

Fixes #2670.

git-svn-id: http://svn.mythtv.org/svn/trunk@13100 7dbf422c-18fa-0310-86e9-fd20926502f2