Refs #3041
Add reading of length from id3 TLEN tags.

Add calculating length using the frame count in the Xing header for VBR mp3s.

For mp3s using either TLEN or Xing headers this reduces the time to scan the file. A full scan of collections with many such mp3s should be noticably faster - the bigger the mp3, the better the overall improvement.
 


git-svn-id: http://svn.mythtv.org/svn/trunk@13104 7dbf422c-18fa-0310-86e9-fd20926502f2