Add a configurable playback resume feature to MythMusic.
This commit adds the ability to remember the current position in the music
tree on the playback screen and move back to that position when it is 
restarted. It will work anywhere in the music tree and can also resume 
playback from the exact position in the track if you wish.

It adds a "Resume Mode" settings that allows the user to select between 
"Off", "Track", and "Exact" modes.

If the user is playing music and then exits while the music is still playing 
the exact position is stored and then upon re-entering MythMusic it will,
depending on the above setting, either:-
 
* Off    - resume playback from the beginning of the active play queue
           (existing behaviour)
* Track  - resume playback from the beginning of the track that was playing
           before exiting.
* Exact  - resume playback from the point in the last track before exiting.

I didn't use the patch on the ticket because in most cases it failed to 
restore the correct track because it didn't take into account the fact that 
playlists can themselves contain playlists and also it wouldn't work if the 
user had 'Show entire music tree' turned on.

Closes #3025.



git-svn-id: http://svn.mythtv.org/svn/trunk@13153 7dbf422c-18fa-0310-86e9-fd20926502f2