Add an album art image viewer to MythMusic's metadata editor.
Following on from Stuart's work to save the details of any album art found 
during the scanning process this adds an additional column to the 
music_albumart table to indicate what each image represents (Front Cover,
Back Cover, CD, Inlay etc).

The DB update will attempt to guess the image type from the filename for any
images that are already in the DB and the file scanner will also do the same
for any new images found during a scan.

The metadata editor has a new page that shows you what Album Art images are
available for the current track and also allows you to change the type of
each image if needed (Press MENU to change the type).

git-svn-id: http://svn.mythtv.org/svn/trunk@13176 7dbf422c-18fa-0310-86e9-fd20926502f2