Added Search Words->Stored Searches. This allows you to select a search stored from Custom Record with the convenience of a popup list and the view keys (Home and End) to cycle thru all stored searches.
Commit [10131] added a Store button to the custom editor for
creating reusable example clauses and saving a work in progress.
This also allows for creating a reusable search to find specific
things without creating a record rule. The "Store" button
dialog now has buttons to store or replace "...as a search" or
"...as an example". The only difference is that things stored
as a search will be included in the list on the Stored Searches
page.

A customexample "New Flix" is inserted so that there will be at
least one view for this page. This can be removed or modified and
other stored searches can be created from the Custom Record page.


git-svn-id: http://svn.mythtv.org/svn/trunk@13452 7dbf422c-18fa-0310-86e9-fd20926502f2