Closes #3474, #3475
Tanks to Michael T. Dean for these patches. I've modified them slightly so any
bugs introduced are my own.

Allow the mythfilldatabase log location the housekeeper uses to point to
a file or a directory. Prior to this commit a directory in this setting
would cause the housekeeper run mythfilldatabase to fail.



git-svn-id: http://svn.mythtv.org/svn/trunk@13508 7dbf422c-18fa-0310-86e9-fd20926502f2