Add a way to defer auto-transcode jobs so that they can be automatically scheduled for days after the original recording completes rather than immediately after.
You can now specify auto-transcode jobs to run from 0-365 days after the
recording completes, so you can for instance watch the HD version of a show
if you watch it within 7 days, but after 7 days have it automatically
transcoded down to a lower-resolution and smaller file on disk.  If the
setting is left at the default of 0, then auto-transcode jobs are run as soon
as possible like they currently are.

Modified patch by Patrick Donnelly.  I added the GUI portion and reworked the
JobQueue settings page in mythfrontend to make things fit better as well as
modifying the patch to use a single DB setting instead of two.

Closes #3503.



git-svn-id: http://svn.mythtv.org/svn/trunk@13517 7dbf422c-18fa-0310-86e9-fd20926502f2