Revert the libavcodec/mpegaudiodec.c patch I made in [13511] in favor of the 1-line fix that the ffmpeg guys made in their SVN tree after I contacted them. This way we don't have to merge in the change later when we resync again.
I'm leaving in an if statement to squelch most of the "incorrect frame size"
messages since these will occur very very frequently when playing an old .nuv
file that was transcoded with mythtranscode before I committed [13510].  We
can remove this if statement at a later time if we want to or maintain it in
our libavcodec tree for a while if we want so people's logs don't fill up with
these warnings.

References #2972.



git-svn-id: http://svn.mythtv.org/svn/trunk@13523 7dbf422c-18fa-0310-86e9-fd20926502f2