Allow the user to set a mythfilldatabase run period where the max hour is less than the min hour, for instance if they want the run period to be from 8PM till 4AM.
This is loosely based on a patch by WP van Paassen in #3310.  I shortened
it up a bit, and removed a part unecessary with current SVN head.
While I was in the housekeeper, I also converted a few things like
now.toString(QString("d")).toInt() to a much cleaner now.date().day().

Closes #3310.



git-svn-id: http://svn.mythtv.org/svn/trunk@13524 7dbf422c-18fa-0310-86e9-fd20926502f2