References #3520
Adds a new option "Let transcode determine frame rate" (defaults to on)
which allows transcode to determine the frame rate. Prior to this patch
"coded 24fps" content was expected (for NTSC).

Note: After changing the option you will need to restart mtd.

With thanks to Rod Nayfield for being that final person needed to shame
me in to settling on a quick fix.


git-svn-id: http://svn.mythtv.org/svn/trunk@13536 7dbf422c-18fa-0310-86e9-fd20926502f2