Adds the mythmovies plugin by Josh Lefler to mythplugins.
It provides show times and cinema listings based on Zip/Post code and a given radius. It uses external scripts to grab times and so can be used in any country so long as a script is written for a local data source.

It ships with a grabber for the USA which uses the ignyte website.

I noticed some minor issues/bugs but nothing serious and the problems will be easier to fix now that it is in SVN.

Closes #2482


git-svn-id: http://svn.mythtv.org/svn/trunk@13582 7dbf422c-18fa-0310-86e9-fd20926502f2