Move the seektable writing code out from the main recorder threads and into the TVRec main thread.  Now the recorder threads fill in the positionMap and positionMapDelta and the TVRec class calls RecorderBase::SavePositionMap() to consume the data.
If it is determined that we should write the delta to the database, a copy of
the delta map is made and the positionMapLock is unlocked.  This allows us to
keep the position maps for as short of time as possible and allows the DB
update to happen without impacting the recorders collecting data.

As part of this, I moved the *Recorder::SavePositionMap() code into
RecorderBase because the only difference between the three was the
positionmap type.  The type is now stored in a variable set by the three
*Recorder classes affected.

Closes #1660.



git-svn-id: http://svn.mythtv.org/svn/trunk@13619 7dbf422c-18fa-0310-86e9-fd20926502f2