Fix for bug where preview video would popup a dialog whenever you move the cursor over a broken video.
The NVP was launching a full screen modal error dialog on any fatal playback error, but this is an annoyance in the 'Watch Recordings' screen. With this changeset we instead just print an error message to the console when we encounter a fatal error while using the 'Null' video output method.

git-svn-id: http://svn.mythtv.org/svn/trunk@13836 7dbf422c-18fa-0310-86e9-fd20926502f2