Make the default transcoding video height 576 if the users video standard is PAL.
This is a minor change and almost pointless but it relates to a recent discussion about sensible defaults for such settings in different localities. In PAL/SECAM countries 576 lines is the default and user who is ignorant of these things will find it easier if the defaults in the transcoding profiles reflect that.



git-svn-id: http://svn.mythtv.org/svn/trunk@13853 7dbf422c-18fa-0310-86e9-fd20926502f2