Refs #3710. Fixes error handling in ffmpeg file.
I've just sent a version of this patch to ffmpeg. Until this get's fixed upstream I'm applying this patch here. If we fail in alloc_picture() after successfully calling s->avctx->get_buffer we call s->avctx->release_buffer so that we don't leak buffers like crazy.

The underlying problem is that ffmpeg isn't picking up on the resolution change, this will be fixed separately.



git-svn-id: http://svn.mythtv.org/svn/trunk@14016 7dbf422c-18fa-0310-86e9-fd20926502f2