Refs #3409. This fixes a rare SEGFAULT.
This is triggered by the Ohio News Network stream because they do not properly clear the trailing portions of their TS packets. But the code this changeset disables would have caused a SEGFAULT even if it has ever been triggered correctly. No one reported this SEGFAULT in years so I presume the corner case the code is designed to handle is very rare.

I'm not closing the ticket itself because this just avoids the SEGFAULT, a theoretical MPEG stream that properly triggers this branch might not play with this changeset applied (but it also won't cause a SEGFAULT like before).

git-svn-id: http://svn.mythtv.org/svn/trunk@14053 7dbf422c-18fa-0310-86e9-fd20926502f2