Added a factor for the Watch List based on the average number of hours that recordings are time shifted per recording rule.
This adds avg_delay to the record table as a weighted average
of hours from when episodes are recorded until they are deleted
(watched, auto-expire turned off). The average ...average(?)
is 100 (about four days) with a max of 200 counted (more than
a week). This stat may be used for other things in the foreseeable
future. A score approaching zero would nearly double the watch
list score. An average of over a week would cut the score nearly
in half.

In other words, the Watch List 'learns' which shows you like to
watch right away and which ones you usually put off for a few
days or weeks and reflects these habits in its scores and ranking.

NOTE: This includes a db update "1192" from mythtv-vid so that
updates don't overlap numbers or get out of sequence. Perhaps
any db update in branches should be reflected in trunk to avoid
differences down the line for those who did or didn't use that
branch.


git-svn-id: http://svn.mythtv.org/svn/trunk@14074 7dbf422c-18fa-0310-86e9-fd20926502f2