Renames the G.A.N.T. folder to G.A.N.T (with permission of Oscar) because windows doesn't like filenames ending in a fullstop.
Includes a database update to change the name in the database. Does not change the name of the mytharchive theme, also called G.A.N.T.

This change will only affect the folder name and internal name of the theme. I'll shortly be committing changes so that the "public" name of a theme is read from the themeinfo.xml



git-svn-id: http://svn.mythtv.org/svn/trunk@14193 7dbf422c-18fa-0310-86e9-fd20926502f2