Refs #3826
Initial work on a new themeinfo class to parse and hold information on a theme in themeinfo.xml. Still a lot of work to done.

This replaces the check for -wide in a theme name and instead looks at the <aspect> tag. As a temporary fix for third party themes it will fall back to the old method.

In appearance settings we now use the themeinfo class to get the name, theme type, aspect and preview image path.

Widescreen themes no longer need to be named "-wide".



git-svn-id: http://svn.mythtv.org/svn/trunk@14197 7dbf422c-18fa-0310-86e9-fd20926502f2