Refs #3842. Applies "3842-multilineup-fixes-v1.patch" (sorry about the botched patch upload earlier),
This should be the last change to 0.20-fixes before the 0.20.2 release unless some major bug turns up.

I'll attach the testing log to the ticket. I haven't tested every combination of flags, so if you experience any problems or think some additional combination of flags not noted in testing.txt should be tested please note it on the #3842 ticket.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@14276 7dbf422c-18fa-0310-86e9-fd20926502f2