Adds the ability to skip commercials in DVDs created with MythArchive of MythTV recordings.
It does this by adding chapter points at cut end points that have been set
for a recording.

It only does this for DVD themes that don't have a chapter menu and you have 
told MythArchive to not remove the commercials from the file.

To enable this feature you have to change addCutlistChapters = False to True
at the top of the mythburn.py script.

Closes #3655.

git-svn-id: http://svn.mythtv.org/svn/trunk@14310 7dbf422c-18fa-0310-86e9-fd20926502f2