Refs #3842. Fixes validation so that firewire channels can be added without the workaround; this also saves an input connection when you fetch channels (validation needs this to work as intended without invasive changes).
I'm not going to apply this to fixes until it's baked in SVN for a while. There are too many card combinations and connection states for me to be sure I've tested everything and I don't want to have any risk that I'll contaminate fixes with new bugs when so many people are updating.

git-svn-id: http://svn.mythtv.org/svn/trunk@14355 7dbf422c-18fa-0310-86e9-fd20926502f2