Refs #3842. Backports source validation changes from head.
Fixes validation so that firewire channels can be added without the workaround; this also saves an input connection when you fetch 
channels (validation needs this to work as intended without invasive changes).

This also fixes a bug in 0.20-fixes SourceUtil::IsUnscanable(uint sourceid) which had already been fixed in head; this fix was needed for the firewire channel fetch to work (since it has an "unscanable" input).



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@14397 7dbf422c-18fa-0310-86e9-fd20926502f2