Adds a three day grabber for mythweather using the BBC rss feed.
Some minor changes to the current conditions BBC grabber.

Both grabbers support worldwide locations just search for the nearest large town or city. Within the UK you can use a postcode in the search dialogue to get an observation and forecast based on the closest weather station.

I'm actually discarding a lot of the forecast information provided because the 3 day screen is pretty limited. I hope to change that when I have time.



git-svn-id: http://svn.mythtv.org/svn/trunk@14445 7dbf422c-18fa-0310-86e9-fd20926502f2