Fixes #2909. Fixes segfault in channel scanner caused by usage unsafe QDialog handling.
This was applied and reverted about 8 months ago because it broke most of the playbackbox dialogs. I've addressed that issues in this changeset and searched the codebase for all similar constructions, finding one in welcomedialog.cpp. This was also tested in the multirec branch for a couple weeks.

This changes the binary revision and so requires plugins to be rebuilt.



git-svn-id: http://svn.mythtv.org/svn/trunk@14478 7dbf422c-18fa-0310-86e9-fd20926502f2