Refs #3486. Fixes for pause bug in ivtv 1.0.0 drivers with patch from Ian Armstrong.
The bug affects the ivtv driver in all 2.6.22.x Linux kernels. This works around the bug by ramping up the speed after a pause instead of trying to jump directly to normal playback speed.

The underlying bug should be fixed in the 2.6.23 Linux kernel release when it comes out, so this workaround is only employed with the 1.0.0 ivtv release included with the 2.6.22.x kernels.



git-svn-id: http://svn.mythtv.org/svn/trunk@14589 7dbf422c-18fa-0310-86e9-fd20926502f2