Fixes #4009. Adds 14:9 aspect ratio to overrides.
This is a compromize aspect ratio used by some broadcasters when broadcasting 16:9 material over PAL/NTSC, it makes the black bars at the top and bottom slightly smaller without looking terrible to most viewers. By telling MythTV the actual aspect ratio is 14:9 and not 4:3 you can zoom into this type of video and see it with the correct aspect ratio and without any letterboxing on a 16:9 display.

git-svn-id: http://svn.mythtv.org/svn/trunk@14593 7dbf422c-18fa-0310-86e9-fd20926502f2