References #4011
This fix stops empty mount directories from MediaMonitor being expanded
to the current working directory. This was causing scanning of CWD n times
(where n is the number of removable devices that were found but not mounted).

This bug could only be hit with "file browse" views.

Thanks to Kiran Adimulam for reporting this issue, continuing to send
me more information and testing this patch.


git-svn-id: http://svn.mythtv.org/svn/trunk@14659 7dbf422c-18fa-0310-86e9-fd20926502f2