Change the way the encoding profiles are found in MythArchive.
First look for a custom profile file in the config directory 
(~/.mythtv/MythArchive on a default setup) then if not found use the default
encoding profiles in INSTALLDIR/share/mythtv/mytharchive/encoding_profiles.

This allows you to customise the encoding profiles as you like with any
parameters that will be passed to ffmpeg without fear of your changes being
overridden the next time you update.

It would be nice to add an editor to allow you to change the parameters from
the UI at some point.

I'll add some additional notes on the MythArchive wiki page later.

Fixes #3914.

git-svn-id: http://svn.mythtv.org/svn/trunk@14713 7dbf422c-18fa-0310-86e9-fd20926502f2