Fix the EPG always displaying the channel icons even when the setting to display them is unchecked.
I've also fixed a problem were the channel icon in the info panel was always
trying to load the image from the local harddrive and wasn't using the cached
image which was fetched from the master BE. This was causing problems for all
FE's not running on the master BE if there wasn't a copy of the channel icons
available locally in the same location as those on the master BE.

Based on the patch by Alex Halovanic. Closes #4031.

git-svn-id: http://svn.mythtv.org/svn/trunk@14725 7dbf422c-18fa-0310-86e9-fd20926502f2