Refs #3966. Makes optional smart channel changing a little smarter.
This makes the code a little more aggressive in finding a match so that when there is only one possible channel that can be selected from the input the user has typed it will be committed without waiting for the OSD to fade or for the user to hit select. Before we would only autocommit completely unambiguous channel changes.

git-svn-id: http://svn.mythtv.org/svn/trunk@14728 7dbf422c-18fa-0310-86e9-fd20926502f2