Refs #4101. Fixes ivtv audio sample rate setting.
I noticed the problem a few weeks back, but ticket 4101 reminded me to finish testing the patch. There were two problems, the AudioCompressionSettings class was saving all the triggered groups not just the selected one, so the values from all put the uncompressed audio type were being overwritten, and the other problem was that once the value was set incorrectly the SampleRate class would allow it.

This fixes both problems and also adds a dbcheck update to fix existing misconfigurations.

git-svn-id: http://svn.mythtv.org/svn/trunk@14748 7dbf422c-18fa-0310-86e9-fd20926502f2