Increase the granularity of MythMusic's artist tree.
Adds a new setting in the global settings for "Number of Artist Tree Groups".
Allows you to have either the existing setting (A B C D E), or two letters 
per group (A B) or a single letter per group (A).

I've split the General Settings into two pages and put the new setting on
there with the Tree Sorting setting. Seems a more appropriate place for it
to me.

Patch from Paul Lambert. Closes #4092.


git-svn-id: http://svn.mythtv.org/svn/trunk@14751 7dbf422c-18fa-0310-86e9-fd20926502f2