Fix a 'feature' in MythMusic where if you have no tracks in the active play queue and are using the simple list mode you are shown a list of available playlists. If you then select one of the playlists the first track of the playlist starts to play but the list view doesn't change to show the tracks.
The behaviour has been changed so that when you select a playlist the tracks
from that list are copied to the active play queue and the list view then 
switches to show the active play queue. I think this is more like what 
you would expect to happen in this case.

Fixes #4106.

git-svn-id: http://svn.mythtv.org/svn/trunk@14767 7dbf422c-18fa-0310-86e9-fd20926502f2