Fixes #3794. Fixes a race condition which can cause 'E'dit to get stuck waiting for a pause to take effect.
The problem is that a Play could get executed while the after the 'E'dit processing function in the NVP had determined that the playback is paused. With this change GetPause() has been renamed IsPaused() and has an added param which tells you if someone has just executed a Play so that the pause is about to be undone, combined with locking and a check for whether we are starting editing in the Play() function this allows us to ensure that playback is really paused when we need it to be.

Tested by myself and Stuart M & reviewed by Stuart M.

git-svn-id: http://svn.mythtv.org/svn/trunk@14821 7dbf422c-18fa-0310-86e9-fd20926502f2