Clear the preview image before preview video playback starts. This means even if the aspect ratios of the two are different the preview image won't be visible beneath the video.
Also scale the preview image to the size of the video container. This
was already done for the preview video. A warning though to anyone who
wants to use a video container larger than the previous default of
160x120, the preview images themselves are still created at that width
and will be scaled up meaning a decrease in quality.

I'm going to propose that the preview images are at least doubled in
size, but that change isn't included in this commit.



git-svn-id: http://svn.mythtv.org/svn/trunk@14823 7dbf422c-18fa-0310-86e9-fd20926502f2