Refs #4132. Fixes two bugs in the playbackbox recording group password dialog handling.
In noticed this when testing dialogs for #4132. First initial focus was being set, this adds a little code that puts the focus on the old password entry if a password has been set, and on the new password entry otherwise. Second while setRecGroupPassword() did set the password in the db it did not update the recGroupPassword variable so if you pulled up the dialog again in the same session the new password you had just entered would not work. Both these problems are fixed in this changeset.

git-svn-id: http://svn.mythtv.org/svn/trunk@14831 7dbf422c-18fa-0310-86e9-fd20926502f2