More preview image scaling tweaks and fixes. This should now correctly scale and position 4:3 images in 16:9 frames as well as 16:9 in 4:3.
Previews will now use a blackhole widget instead of the container area
if it is specified. Not only is this more logical from a theme point
of view, it also allows additional images and text to be placed in the video
container for display with the preview image e.g. A frame around the
preview. Anything placed in the video container will not be displayed
unless preview images or video are enabled.  

See http://miffteevee.co.uk/imagebin/theme_fullscreen59.png for an example.

Some credit should go to Bill cizek@ as he wrote and submitted a patch to do the
same thing long before I independantly started working on fixes. His
patch was not used because I saw it too late.

Closes #3771



git-svn-id: http://svn.mythtv.org/svn/trunk@14865 7dbf422c-18fa-0310-86e9-fd20926502f2