Don't delete the shared decoder we acquire to test play some tracks while importing them.
This fixes a segfault spotted while doing some testing for #4132. If you exit
out of the music import dialog after playing a track and then re-enter it and
play another track you get a segfault. That's because we deleted the shared 
decoder given to us by Decoder::create() when we shouldn't.



git-svn-id: http://svn.mythtv.org/svn/trunk@14886 7dbf422c-18fa-0310-86e9-fd20926502f2