Fixes #4115. Backports tv_play.cpp portion of [14794] to fix the major cause of segfaults using network control in 0.20-fixes.
The problem here seems to be that Network Control was never updated when we changed how LiveTV worked in MythTV. For keyboard input we check the ignoreKeys boolean and ignore unsafe actions while it is set. But Network Control completely ignores this and so you can do anything you want during the ringbuffer switching. 

Another problem is that if both keyboard control and network control are active two incompatible actions can be triggered in a single iteration of the TV main loop, this is fixed by simply skipping network control events if we have a keyboard event, until the next iteration of the loop.

Finally, there were a couple thread safety issues which were trivially fixed.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@14921 7dbf422c-18fa-0310-86e9-fd20926502f2