Fixes #4084. Code tidy from candrews.
As of Qt 3.1 using the QApplication thunks for QEventLoop::enterLoop() and QEventLoop::exitLoop() is depreciated. Instead you get a pointer from QApplication to it's event loop, check that it isn't null, and then call enter and exit directly.

This does not offer any power saving enhancements or anything like that discussed in the ticket comments, but it should be one less thing to worry about for the Qt 4.x port.

git-svn-id: http://svn.mythtv.org/svn/trunk@14951 7dbf422c-18fa-0310-86e9-fd20926502f2