Allow jump points that don't return to the main menu before jumping.
This adds a new REG_JUMPEX macro to mythmainwindow.h that adds an extra
parameter that allows you to indicate whether a jump point should return to
the main menu before jumping. It will be used by the jump point that shows
the popup mini player in MythMusic.

This requires a 'make clean' and rebuild of the plugins.

Refs #4222.

git-svn-id: http://svn.mythtv.org/svn/trunk@14969 7dbf422c-18fa-0310-86e9-fd20926502f2