Fixes a segfault playing back certain video with H.263 decoder reported to me by greg.
The problem was with the allocator for Direct Rendered frames. Before the mythtv-vid branch we ignored the direct rendering directive for frames whose width wasn't an mpeg-2 width. The code to handle direct rendering for other frame sizes was added as part of getting DVDV decoding to work, but we didn't have a test case for one of the five allocation paths. Thankfully we added a debugging warning message when this path was run which allowed me to find the problem quickly.

With greg's test case I was able to test this code path and added the CODEC_FLAG_EMU_EDGE flag which prevents the segfault greg was seeing.



git-svn-id: http://svn.mythtv.org/svn/trunk@15033 7dbf422c-18fa-0310-86e9-fd20926502f2