Modify the backend to send out a new format of the RECORDING_LIST_CHANGE event when a recording is deleted.  The format is as follows:
    RECORDING_LIST_CHANGE DELETE <chanid> <starttime>

starttime is in Qt::ISODate format.

The frontend handles this in PlaybackBox to allow the Watch Recordings screen
to dynamically update its in-memory program cache rather than having to
reload the full list of all recorded programs every time a single program
is deleted.  This applies to user-initiated deletions as well as things
like the AutoExpirer since they all use the same deletion code in the backend.

This should give people a more responsive GUI when deleting programs because
the whole program list does not have to be reretrieved from the backend.

All other code that sends out RECORDING_LIST_CHANGE still uses the original
format and causes the frontend to reload the complete list.  This happens
for instance when a new recording starts.

This does _NOT_ fix a current known issue with the delete popup hanging if
you try to delete a program while the scheduler's BUQ is running.  A fix
for that is being discussed.



git-svn-id: http://svn.mythtv.org/svn/trunk@15059 7dbf422c-18fa-0310-86e9-fd20926502f2