Allow MythMusic to continue playing in the background after exiting.
When you exit the MythMusic playback screen you are given the option to exit
and stop playing or exit and continue playing. If you choose to continue 
playing you can then use other parts of Myth while your music is still
playing.

When a new track starts and you are not in the main music playback screen a
mini popup player is shown for a few seconds then disappears. While the mini
player is on screen you can press ESCAPE to remove it or press SELECT and it
will stay on screen until you press ESCAPE. There is also a new jump point
(not bound to anything by default) that will instantly show the mini player
anywhere in Myth. While the miniplayer is on screen you can use most of the
key bindings available in the main playback screen to change tracks, fast
forward, rewind, pause, stop, mute the audio, increase/decrease the volume
etc. Like other part of Myth the miniplayer is fully themeable.

Thanks to Otto, Mike, Benoit, Robin, Bill and anyone else I may have forgot
for testing and providing some feed back.

Note to testers: I changed the name of the jump point to "Show Music Miniplayer"
so you will have to remove the binding on the old jump point and set it again
on the new one.

Closes #4222.

git-svn-id: http://svn.mythtv.org/svn/trunk@15091 7dbf422c-18fa-0310-86e9-fd20926502f2