Closes #3684
Applying a work-alike patch to the one submitted by jmortensen A verizon net.

This patch adds a new "dialog" for manual title searches (saving one the
arduous journey to the metadata editor). It isn't remote friendly and
will not be until moved to MythUI.

Additionally it moves one step closer to interruptible searches, the bulk
of this patch is work done with that goal in mind.

This also adds a test and artwork directory.
  * test
    * This directory contains scripts used while testing some of the
      async changes. Eventually to contain other MythVideo tests.

  * artwork
    * Raw artwork, currently the .xcf file for the new search dialog background.

A note to theme maintainers:

The new "entersearchtitle" container is optional.

Also mv-search-popup.png represents the zenith of my artistic abilities,
any better default would be welcomed.


git-svn-id: http://svn.mythtv.org/svn/trunk@15175 7dbf422c-18fa-0310-86e9-fd20926502f2