This update allows an xbox360 to stream video from a mythbackend and is the start of getting music streaming working as well.
 As well the video portion of the upnp server has been changed
to allow for navigation just like a directory tree 
rather than the precanned "reports" previously used.

 To use this you need to have your videos available on your
masterbackend in the same location you have defined in the mythvideo setup 
menu. 

 Currently a list of the videos is grabbed at backend startup and 
stored in a new table.  My next update will add a housekeeper
job to update/rebuild the list every 30 minutes or so. I'm doing
it this way to avoid users having to run the videomanager
to make new files available. If at some point we start auto 
updating the mythvideo tables periodically I'll change it to use
that instead.

 



git-svn-id: http://svn.mythtv.org/svn/trunk@15257 7dbf422c-18fa-0310-86e9-fd20926502f2