Add option to either scale images to full screen or scale to fit in MythGallery.
This is based on the patch by Joep Admiraal. I've modified it so it doesn't
need a new setting instead you can toggle the behaviour in the single view
screens using the FULLSCREEN key binding ('W' by default). The setting is
remembered when you leave the single view screens and will be restored when
you next use it.

Closes #4364.

git-svn-id: http://svn.mythtv.org/svn/trunk@15271 7dbf422c-18fa-0310-86e9-fd20926502f2