Fix the "All Programs" count on the Recording Group chooser popup so that the cound does not include programs in the special "Deleted" Recording Group.
This commit should also fix a bunch of places where the special case "Deleted"
recgroup name was being translated when it shouldn't be.  Since this is a
special case recgroup, it needs to be treated the same as the 'LiveTV' and
'Default' recgroups are treated regarding translation.  Can someone who
uses a non-English frontend verify this please and report back if it is
not working correctly?



git-svn-id: http://svn.mythtv.org/svn/trunk@15393 7dbf422c-18fa-0310-86e9-fd20926502f2