Allow you to stop the mini-player popping up on track changes.
Sometimes the mini-player popping up on track changes can be annoying for
example when you are entering some text and suddenly the focus is take from
you so this allows you to temporarily, or permanently if you wish, stop it
showing on track changes.

While the mini-player is on screen you can press 'MENU' to toggle the option
on or off. The music player remembers the last setting and will restore it
on a restart.

Closes #4353.

git-svn-id: http://svn.mythtv.org/svn/trunk@15426 7dbf422c-18fa-0310-86e9-fd20926502f2