Fixes #4472. Refs #4480. Fixes a channel initialization problem.
The problem was causing segfaults in the firewire recorder as the channel was sometimes reinitialized on channel changes before the recorder had successfully paused. This was causing channel change scripts to be executed twice per channel change rather than once.

This adds a new function to ChannelBase, IsTunable() which only verifies whether the channel is tunable in terms of being a valid channel and not conflicting with input groups, but does not actually try to tune it. This also adds a parameter to ChannelBase::Init() called setchan, which when set actually intializes the channel, and when cleared does all the verification and has all the side-effects of Init() except for actually performing the initialization of the hardware.

git-svn-id: http://svn.mythtv.org/svn/trunk@15470 7dbf422c-18fa-0310-86e9-fd20926502f2