Improved scheduler decisions when two or more overlapping shows have the same priority where there are different input priorities.
An example would be an HD card 1, input 1, source 1 and SD
card 2, input 2, source 2. Both sources have channels that
carry the same stations. Because of input, HD, channel, or
power priority, showings on card 1 are +1 better than card 2.
Two rules have the same per rule priority. "A" is shown at 8
and 11 and is higher in the pecking order than "B" with one
showing at 8.

Until now, "A" would win the 8pm showing on the HD input and
"B" would be placed in the only open slot at 8 on the SD input
which is fine because there is no conflict. However, the user
did not declare a preference for one title over the other but
did express a preference for using the HD input. Therefore,
the correct solution would be to schedule "B" in HD at 8 and
"A" in HD at 11.

This fix is based on a patch from David Engel. A new check
is added to the beginning of conflict resolution which tests
if the show in question (B) overlaps a show with the same
priority (A) and that show has other showings with the same
priority (A on card 1 at 11 but not the lower priority
showings on card 2). If this does not resolve the overlap
then we continue with the existing checks to find alternatives
at lower priority or eventually move a higher priority show
if allowed and necessary in order to prevent a conflict.

In testing, this found good solutions in simple or complex
situations and the results were better reflections of the
user choices.


git-svn-id: http://svn.mythtv.org/svn/trunk@15528 7dbf422c-18fa-0310-86e9-fd20926502f2