Check the MySQL version before attempting to upgrade the main DB schema. If the user is not running MySQL v5.x, then the program will exit giving the user the option to upgrade MySQL or downgrade MythTV. Patch by Michael T. Dean, modified slightly by me.
As part of this, clean up some of the MSqlQuery declarations in
UpgradeTVDatabaseSchema, we don't really need a new instance and DB
connection for each query in this function.

Also modify mythtv-setup to check the result of UpgradeTVDatabaseSchema
the same way that mythbackend and mythfrontend do, rather than blindly
continuing assuming the DB was upgraded.



git-svn-id: http://svn.mythtv.org/svn/trunk@15579 7dbf422c-18fa-0310-86e9-fd20926502f2