Add option to add subtitles to DVD's created by MythArchive.
There's a new 'Add Subtitles' setting that when turned on will cause
the script to automaticaly add subtitles to the DVD for any files that
have them.

Only DVB subtitles are currently supported and requires that you tell
MythArchive to use ProjectX as the cutter/demuxer.

It's only been very lightly tested with a few DVB-t recordings I had to hand
so there are probably still issues to be fixed.

This commit also should fix #4475.

git-svn-id: http://svn.mythtv.org/svn/trunk@15609 7dbf422c-18fa-0310-86e9-fd20926502f2