Carry over [15670] from trunk.
Reset idleSince and m_isShuttingDown if the halt command is empty or returns
control back to mythbackend when the scheduler tries to shutdown the server.

This could happen because the halt failed or because the user is suspending or
hibernating the OS using the halt command instead of shutting it down.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-20-fixes@15671 7dbf422c-18fa-0310-86e9-fd20926502f2