Merge MythHorizListButton with MythListButton and introduce the concept of layouts. Added a third grid format layout as well.
This turns MythListButton into a verstile widget that can be used for
everything from simple lists, to the gallery views in
MythGallery/MythVideo. Even better, the layout used is themeable
meaning that all sorts of different things can be done with themes.

The "buttonlist" definition now includes a <layout> element which
accepts the values vertical, horizontal and grid.



git-svn-id: http://svn.mythtv.org/svn/trunk@15721 7dbf422c-18fa-0310-86e9-fd20926502f2