Adds some more ivtv driver work arounds....
Adds hack for video format setting and input switching with PVR-350 and PVR-250 rev 1 cards using ivtv 0.10.0 and later. a patch has been submitted upstream to fix this in the driver, but unfortunately 2.6.24 was released with a driver which breaks all existing applications using the ivtv cards. This hack is keyed to the ivtv driver and driver version, so it can be disabled for future ivtv driver releases.

This also updates Hans' VBI bugs warning for ivtv drivers before 0.10.0 and after 0.10.5 due to new VBI bugs in more recent ivtv drivers.

git-svn-id: http://svn.mythtv.org/svn/trunk@15734 7dbf422c-18fa-0310-86e9-fd20926502f2