Add a new 'localAction' to the configuration of a jumppoint.  This defaults to empty.  This is used to allow local actions to take precedence over jumppoints when both are bound to the same key.  Currently only the Program Guide and Program Finder jumppoints have localActions defined (GUIDE and FINDER respectively).  Below is how it works for the Program Guide.
If the Program Guide jumppoint is bound to 'S' and the GUIDE action in the
"TV Playback" context is also bound to 'S', and the user hits 'S' while
playing back a recording or LiveTV, the local 'GUIDE' action will now be
used instead of the Program Guide jumppoint.  When 'S' is hit in a context
that does not have a 'GUIDE' action defined, the Program Guide jumppoint
will be executed.  In order for this to work, both the jumppoint and the
local action MUST be bound to the same key.

This code is flexible and can easily be extended for new actions by
changing the proper jumppoint registration from using REG_JUMP to REG_JUMPLOC
and defining the localAction that equates to the jumppoint.

WARNING: This updates the MYTH_BINARY_VERSION because of the new argument for
    MythMainWindow::RegisterJump, so you must recompile all plugins after
	upgrading to this SVN version.



git-svn-id: http://svn.mythtv.org/svn/trunk@15736 7dbf422c-18fa-0310-86e9-fd20926502f2