Allow the PREVCHAN keybinding to act the same as JUMPPREV while watching a normal recording.
Closes #4434, but applies a different patch and almost reverse logic.

The patch in the ticket removed PREVCHAN and made JUMPPREV serve double-duty
making it go to the previous channel in LiveTV and the previous program while
watching a recording.  This would have lost functionality though, because it
is a valid action to jump from LiveTV to a previous recording.

This patch allows the user to map only PREVCHAN if they want to and use it
in LiveTV to jump to the previous channel and use the same key to jump back
to the previous recording when watching a recording.  It does not prevent
the user from mapping both PREVCHAN and JUMPPREV to different keys, it does
allow both to be used to jump to the previous recording when not in LiveTV.



git-svn-id: http://svn.mythtv.org/svn/trunk@15795 7dbf422c-18fa-0310-86e9-fd20926502f2