adds decoding CPU count to video display profiles
Default for all profiles is 1, so the displayu profile has to be changed
if libavcodec's multi threaded decoding should be used. Multithreding
in libavcodec has to be enabled in configure with --enable-ffmpeg-pthreads.
With CPUs set to 2 MPEG4 AVC (H.264) HDTV with multiple slices should be
playable on most x86 dual core cpus.


git-svn-id: http://svn.mythtv.org/svn/trunk@15816 7dbf422c-18fa-0310-86e9-fd20926502f2