Fixes #4566. Makes sure that the DeviceReadBuffer stops reading the dvb device when the DVBStreamHandler is stopped.
Before it was possible that the DRB never received the signal because the stream handler was waiting for the DRB to stop blocking before sending the signal and the DRB would only stop blocking if it recieved data on the file handle, or got the stop/paused signal. This change makes sure we send the stop signal to the DRB before we try to stop the stream handler so that shutdown does not depend on there being data in the DVB driver's buffers.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16058 7dbf422c-18fa-0310-86e9-fd20926502f2