Refs #4515. Fixes regression in handling off-air MHEG channels from the multirec merge.
This uses a patch originally written by Bradley Kite and modified by me to be less intrusive. This adds creates a fake keyframe entry for an MPEG stream with no audio or video streams so that the frontend can read the data and tell you that the channel is off-air instead of timing out waiting for data.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16135 7dbf422c-18fa-0310-86e9-fd20926502f2