Merge mythtv part of [16153] from trunk.
Fix a feature? of the background music player were if an external program is
run like mythbrowser using myth_system() the music will stop at the end of
the current track and will only resume again when the external program is
exited.

To work around this this commit adds an additional flag to myth_system that
allows you to run a command waiting for it to finish but doesn't block the
parent thread so it can continue dispatching messages etc.

It also updates MythBrowser and MythNews to call myth_system() with the new
flag and blocks all input while the external program is running.

Refs #4222.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16154 7dbf422c-18fa-0310-86e9-fd20926502f2