Fix to allow MythEvent with extradata to be used safely.
MythEvent was making a safe copy of the message but not the extradata before using a copy in another thread. This fixes that problem and also moves some of the implementation from the mythevent.h to mythevent.cpp. This also makes sure that if you use the int constructor message and extradata are explicitly initialized in the constructor, and that the ExtraData(int) method does not segfault on a bad index value.

NOTE: MythEvent itself is still not threadsafe nor reenterant, but it is now safe to create a MythEvent or the strings used to initialize it in a thread other than the Qt thread (and then use it in any one thread at a time).

Unfortunately this requires a binary revision bump, so plugins will have to be recompiled, I also recommend a 'make distclean' for mythtv itself.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16233 7dbf422c-18fa-0310-86e9-fd20926502f2