Bugfix from Yeechang Lee to the PageTop/Middle/Bottom code in PlaybackBox.
This commit still doesn't add default keybindings for these since we have
been trying not to add new defaults for optional items such as this because
of the large number of existing actions and jumppoints that already exist.
We prefer to leave optional actions like this up to the user to bind rather
than risk potential conflict with an existing binding which the user may have
configured.

Closes #4451.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16286 7dbf422c-18fa-0310-86e9-fd20926502f2