Merges [16357] and closes #2662
This is a work around for when libav doesnt return a valid DTS
after a call to av_seek_frame().  This allows us to seek around
in MKV/MOV/MP4 files. If at some point libav changes to return
some info we can remove it but for now this does the job.




git-svn-id: http://svn.mythtv.org/svn/trunk@16362 7dbf422c-18fa-0310-86e9-fd20926502f2