Fixes #4567. Fixes rendering of chaff in XVideo renderer when the video buffer size != video frame size.
This is the same as the earlier reverted commit, except that I made sure that the NVP and classes using it don't use the real video frame size unless they know how to handle it. All others use the video buffer size, which is what was being used before this changeset. They are usually the same thing, but when they are not the same you can get rendering artifacts, and if the video frame size is used for buffer allocations you can overwrite memory or if lucky get a segfault.

The earlier reverted commit caused a a portion of the preview videos of videos with a frame size not divisible by 16 in both dimensions to have a green hue over a portion of the image.


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@16434 7dbf422c-18fa-0310-86e9-fd20926502f2