Refs #4064. Adds ability to create a special backward compatible MythTV get_gl_proc_address() on distros whose drivers all support glXGetProcAddressARB.
The ./configure option is not printed in help because it allows you to create a MythTV binary which won't run on your computer. But it is available for distro maintainers who wish to create a single binary that works with many GLX libs, which they are responsible for testing for compatibility with glXGetProcAddressARB().

Leaving ticket open for backport.

git-svn-id: http://svn.mythtv.org/svn/trunk@16877 7dbf422c-18fa-0310-86e9-fd20926502f2