Fixes #3504. Make sure we always find the EPG start channel.
This adds a GuideGrid::FindChannel() method which finds the EPG start channel. Before multirec we only stored the chanid for the currently shown channel, even if we had two or more chanid represented by a single row in the EPG. Now that we keep the chanid info we can use it to more reliably locate the EPG start channel.

git-svn-id: http://svn.mythtv.org/svn/trunk@16882 7dbf422c-18fa-0310-86e9-fd20926502f2