Adds an image reflection effect to MythUI. This is more proof of concept than the final implementation, I'll be optimising and tweaking it.
See http://miffteevee.co.uk/imagebin/mythui_imagereflection3.png for examples of reflection effects.

The following element belonging to <imagetype> defines the look
{{{
<reflection axis="vertical" shear="50" scale="50" length="50" />
}}}

Axis  - the reflection axis {vertical,horizontal}
Shear - Percentage shear of the reflection shear, affects the angle of the reflection
Scale - Percentage to scale down the reflection, affects the angle between the image and the 'surface'
Length - Percentage of image to retain in the reflection, affects the length of the reflection

Image transformations will be added shortly as well, allowing images to be sheared, rotated etc

N.B. It's not really intended that it be used for static theme images, reflection effects for those images would be better 'hardcoded' using your favourite graphics editor. This is intended to be used with dynamic content, album art, movie posters, preview images and gallery images.

Refs #12



git-svn-id: http://svn.mythtv.org/svn/trunk@17123 7dbf422c-18fa-0310-86e9-fd20926502f2