Merge [17146] from trunk.
Don't treat not being able to remux the subtitles as a fatal error. spumux can
fail to add the subtitles if the images are too complex. At the moment we just
ignore the error and finish the DVD without the subtitles until a better fix
can be found.

Patch from Alex Butcher.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@17147 7dbf422c-18fa-0310-86e9-fd20926502f2