Refs #5267. Fixes two chicken and egg problems with DB initialization.
The first is not to do the datadirect programid update if the DB is empty. This didn't do any harm, but caused some error messages to be logged.
The second is not to do the database backup on an empty database, since it will fail to find settings and housekeeping tables.

There are some other problems running mythtv-setup to initialize the database, but these seem to be Qt4 port related.
svn commit -m Refs

git-svn-id: http://svn.mythtv.org/svn/trunk@17160 7dbf422c-18fa-0310-86e9-fd20926502f2