Allows input groups to work properly when each recorder has more than one input and so can be in multiple mutually exclusive input groups.
The scheduler was handling this correctly, but then the channel class was blocking a tune when it saw that you were using an input that was in an input group with an input on a busy recorder, even if the busy input on the other recorder was not in the same input group.

The example of this in practice is a Cable firewire STB which is plugged into the S-Video input of a capture card which also plugged directly to the cable and the same STB is also hooked up as a firewire recorder. When the capture card was busy recording directly off the cable, it was blocking the firewire recorder, even though it was only the S-Video input which was in the same input group as the firewire recorder's input.

git-svn-id: http://svn.mythtv.org/svn/trunk@17172 7dbf422c-18fa-0310-86e9-fd20926502f2