Refs #5346. Adds an optional ignoreDB param to MythContext init.
This optional param allows you to disable DB access by the GetSettings/SetSettings code. This allows you can write simple programs that use MythContext, but set the settings via command line parameters.

This setting is also respected by the two classes that create a VideoDisplayProfile, AvFormatDecoder and VideoOutput. (And VideoOutputNull, which uses the VideoDisplayProfile created by VideoOutput).

I'm using this for the commercial flagger to flag videos as a standalone program, but it would also be used for things like a firewire channel changer that used our firewire smarts but didn't touch the DB, or with the command line interface to the channel scanner in the channel scanner branch (which can emulate dvbscan output).

Unfortunately this changes the myth binary revision, so plugins will need to be relinked.

git-svn-id: http://svn.mythtv.org/svn/trunk@17312 7dbf422c-18fa-0310-86e9-fd20926502f2