Fix a possible problem with mythshutdown --lock/--unlock.
The lock/unlock commands can be nested so if lock is called twice unlock
would have to also be called twice to unlock shutdown. Each lock/unlock
requires 2 db queries one to get the setting and one to save it back again.
The problem is if two instances of mythshutdown is running the reads and
writes could overlap each other resulting in the setting getting out of
sequence.

This fixes the problem by using only one query to update the setting and also
locks the table making it impossible for two mythshutdown processes to update
the lock setting at the same time.

Refs #5269.


git-svn-id: http://svn.mythtv.org/svn/trunk@17342 7dbf422c-18fa-0310-86e9-fd20926502f2