Significant speedup for the opengl painter for some videocards.
Some cards (like the 7050pv in my main frontend when run at 1080p) didn't do 
so well with the way text was renderered, and did _especially_ poorly with 
outlined text. Menu cross-fades could be painfully slow.  

This makes it speedy.

Qt4 allows us to render text directly to a transparent QImage, so do that to
pre-cache the entire string, instead of splitting it up into many GL draws for
the shadow, outline, etc.  This means color-cylcing text (which is completely
unused, as far as I'm aware) will be slightly slower than before, but the
tradeoff is worth it, IMO.  Also gets rid of a bit of special-cased text
rendering functions, as text is essentially treated the same as an image now.

I may have missed a corner-case or two, so please let me know if anything
looks wacky after this checkin.



git-svn-id: http://svn.mythtv.org/svn/trunk@17611 7dbf422c-18fa-0310-86e9-fd20926502f2