Merge [17723] from trunk.
Fix one of the causes of mytharchivehelper returning a file duration of
0 length.

This one is caused by files with more than one video stream and
mytharchivehelper using the duration of the last one found which could be
an inactive stream. This fix just uses the duration of the first video stream
found that has a non zero duration.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@17724 7dbf422c-18fa-0310-86e9-fd20926502f2