In MythContext::PromptForSchemaUpgrade(), if we created the temporary main window and later EndTempWindow(), then we need to reset the database connection info using SetDatabaseParams() because the creation of the temp window set the hostname to the empty string.
This was causing segfaults in mysql_next_result during ~MSqlQuery() in some
DB updates if you were upgrading using mythtv-setup or mythfrontend.

This doesn't solve the real bug, but prevents it from occurring.  The bug
occurs after we create a MSqlQuery while the hostname is empty and then later
destroy that that MSqlQuery.



git-svn-id: http://svn.mythtv.org/svn/trunk@17853 7dbf422c-18fa-0310-86e9-fd20926502f2