Allow <area> tags to hold percentage values.
This feature has only been added to ease the creation of scalable popup dialogs and lists. It is not envisioned as a replacement for all absolute values in a theme. That approach would be not recommended, not least because of the loss of pixel precision in the conversion between percentages and integer coordinates.



git-svn-id: http://svn.mythtv.org/svn/trunk@18014 7dbf422c-18fa-0310-86e9-fd20926502f2