Add a new MythUISearchDialog that provides a dialog to quickly choose an item from a list.
You pass a QStringList containing the list you want to search. As the user
enters a text string in the edit the list of entries changes to show only
the entries that match the input string. You have the option to search
anywhere in the string or only the start.

When the user either clicks an entry in the list or presses the OK button
the dialog will either send a DialogCompletionEvent or an haveResult(QString)
signal will be generated. Both pass the selected string back to the caller.

I'll be using this in MythBrowser's edit bookmark screen.


git-svn-id: http://svn.mythtv.org/svn/trunk@18192 7dbf422c-18fa-0310-86e9-fd20926502f2