Re-write myththemedmenu to use mythui directly instead of using it's own theme format.
This change will break third party themes. You will need to use one of the following until they are converted to the new format: Blue, G.A.N.T, MythCenter, MythCenter-wide, Retro, Minimalist-wide, Iulius and Titivillus

The spacing of buttons is a little different for now, work is being done on allowing buttons to be spread to fill available space, please don't file tickets for this.

Plugins will need rebuilding.



git-svn-id: http://svn.mythtv.org/svn/trunk@18606 7dbf422c-18fa-0310-86e9-fd20926502f2