Closes #5924
Changes image cache to make space before cache insertion. Previously it was
possible for the image just added to the cache to be removed (when the
contents of the 20MB cache contain only items added within the same second)
with later references to the removed item (the item being added to the cache)
causing a crash (sometimes).


git-svn-id: http://svn.mythtv.org/svn/trunk@19134 7dbf422c-18fa-0310-86e9-fd20926502f2