Update the program details editor in MythArchive to MythUI.
Editing the description is a bit of pain in a single line edit but it's
all we've got.

There seems to a bug in MythUITextEdit that causes it to draw text out of the
text edit area if the text is too large for it.

It's currently impossible to use a remote to edit text.



git-svn-id: http://svn.mythtv.org/svn/trunk@19181 7dbf422c-18fa-0310-86e9-fd20926502f2