Carries over [19229] from trunk.
Fix a couple issues related to the fsID caching in the Storage Groups
code that tries to determine which Storage Group directories are actually on
the same filesystems.  The first issue was causing some directories to not
be properly detected as on the same filesystem.  The second issue was that
if a pair of directories were ever improperly detected as no being on the 
same filesystem, they could not be properly detected at a later time.

The main issues were in the behind-the-scenes code, not in the display
code in the mythbackend status page.  That page would appear correct while
"-v file,schedule" debugging revealed that filesystems were in fact being
improperly detected.

The bugs were causing Myth to sometimes improperly spread the load out across 
the directories in a Storage Group in an environment where the same directories
were mounted on more than one backend.  This could lead to I/O issues as well
as drives filling up too quickly since more recordings than anticipated were
being sent to the a particular drive.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@19230 7dbf422c-18fa-0310-86e9-fd20926502f2