mpeg: don't try to overwrite any of the incoming packet buffer
In the event the mpeg stream parser fails to process a given packet,
it was overwriting the current packet's sync byte w/a non-sync-byte
value, to signal that we should resync the stream. This works with
many buffers, but not those produced by the new(er) Linux FireWire
stack, as the incoming buffer is write-only for the controller.
Trying to overwrite the buffer results in a segfault. This works
around the issue by simply adding a resync flag, and now I've got
my cable box working pretty darned reliably (about 20 consecutive
livetv start/stops, with multiple channel changes, and I haven't
hit the segfault, which used to trigger roughly 60% of the time).



git-svn-id: http://svn.mythtv.org/svn/trunk@19341 7dbf422c-18fa-0310-86e9-fd20926502f2