Fixes #4792. Implements solution #2.
Solution #3 would be nicer, but require a bit more coding. (See ticket for details.)

(This just rejiggers the video renderer fallback priority in the case of XvMC to help out the folks still using VIA HW, so that when they encounter files the XvMC can't handle opengl is not used as the first fallback).

git-svn-id: http://svn.mythtv.org/svn/trunk@19440 7dbf422c-18fa-0310-86e9-fd20926502f2