Fixes #6041. Don't segfault on old style lirc config files.
The fallback parser in the lirc_client lib relies on some static globals which are only initialized if lirc_init() is called. We stopped calling that in [19499] because of some other bugs in that function. Modern lircrc's were unaffected so this problem didn't turn up in my testing. This pulls in the lirc_client.c and fixes the bugs and gets rid of the static globals so we can call lirc_init() and allow the legacy parser code to work correctly.

git-svn-id: http://svn.mythtv.org/svn/trunk@19519 7dbf422c-18fa-0310-86e9-fd20926502f2