Closes #5884
Adds back the detail screen for gallery as well as adding it to the menu
for all views.

The popup is optional. The current popup is ugly, something nicer will go
in eventually.

Also bets on my wanting to look at python a tiny bit more than scheme.

Fixes view cast dialog as well.



git-svn-id: http://svn.mythtv.org/svn/trunk@19551 7dbf422c-18fa-0310-86e9-fd20926502f2