Ports the Watch Recordings screen (playbackbox) to mythui. This is a large screen and a substantial change (with 1500 fewer lines) and due too the number of popups and possible configurations of this screen there may be some overlooked problems.
The new theme is deliberately styled on the old screen so not to confuse anyone, with minimal additions to demo some of the new features. A more adventurous demo theme will be made available on the -dev list for those that want to see just what is now possible.

Known issues -
Help popup is currently disabled.
Preview video code is removed, to be replaced with a mythui video widget in the near future.

Please report bugs to the dev list for the moment, at least until things settle down.

Refs #12



git-svn-id: http://svn.mythtv.org/svn/trunk@19603 7dbf422c-18fa-0310-86e9-fd20926502f2