Disable VDPAU colorkeying by default.
r19625 breaks playback for people not using the overlay, mainly where the composite extension is in use.
If you think you're using the overlay and suffer from blank screens with vdpau playback, set the environment variable USE_VDPAU_COLORKEY to anything and colorkeying will be enabled. Hopefully a short term fix.

Closes #6102.

N.B. Disabling the composite extension appears to be the way to go with vdpau. Performance should be better and vertical sync will be enabled.


git-svn-id: http://svn.mythtv.org/svn/trunk@19638 7dbf422c-18fa-0310-86e9-fd20926502f2