Carry over [19958] from trunk.
Original commit log:

If MasterBackendOverride is on and MainServer::HandleQueryRecordings()
can't find the slave for a recording, just return a myth:// URL for the master
instead of searching for the file.  This treats files that were recorded on
a slave backend the same way we treat ones recorded on the master if that
slave is unavailable.  Should speed up the loading of PBB for people who
have a slave backend that is down for some reason.



git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@19959 7dbf422c-18fa-0310-86e9-fd20926502f2