Closes #6438 Fixes maxepisodes regression with back-to-back recordings
When there are back-to-back scheduled recordings and maxEpisodes is
set and maxNewest is not, maxEpisodes is not respected.

This change removes the duplicate check added to the maxepisodes query
in [15808]. The maxexpisodes query will now count all recordings that
are not preserved.

Refs #3917

Thanks to David Engel, Chris Pinkham and Sphery.

git-svn-id: http://svn.mythtv.org/svn/trunk@20387 7dbf422c-18fa-0310-86e9-fd20926502f2