Fix focus issue observed with Ubuntu jaunty.
Note: Not everyone sees the focus issue, but a few do.. It appears to be due to show() not always immediately mapping the window in X11. This listenes for the Qt show event and adds a post-show event to the queue when it is seen. Qt appears to reliably not process that post-show event until after the X11 map has really taken place, so we then handle that event by calling raise and activateWindow() to request focus.

It is still possible that we will not be granted focus when we ask for it (due to window manager focus policy), but this change tries to assure that we do ask for focus only after it is possible for the window manager to give us focus.

git-svn-id: http://svn.mythtv.org/svn/trunk@20405 7dbf422c-18fa-0310-86e9-fd20926502f2