Refs #5473. Add code to deal with pulse audio.
By default when compiled on a system with pulse audio we just tell the pulse audio server to leave us alone. For those that want to experiment with pulse audio, we now read the EXPERIMENTALLY_ALLOW_PULSE_AUDIO environment variable and when it is set allow MythTV to run with pulse audio enabled.

git-svn-id: http://svn.mythtv.org/svn/trunk@20421 7dbf422c-18fa-0310-86e9-fd20926502f2