fix and enhance the fuzzy matching of texts in EIT update code
Most importantly fix a min/max mistake introduced in [9668], the fuzzy matching
would always report at least the same score as an exact match even if there
was not a single word match.
Makes the fuzzy matching symmetric (score(a, b) = score(b, a)) and lowers
score for matches of words with 3 or less charecters. Most of the short
words are rather unspecific (articles, conjunctions, prepositions, ...).


git-svn-id: http://svn.mythtv.org/svn/trunk@20485 7dbf422c-18fa-0310-86e9-fd20926502f2