Refs #6478. This merges DBEvent in the EIT program data classes with ProgInfo in the XMLTV program data classes.
The insertion algorithms have not been merged, just the base data representation class. In my runs of mythfilldatabase this reduces the memory footprint of mfd by 15-20% by using the more compact representation developed for EIT for the common data and by compacting the QStrings. The wall clock time is the same before and after.

This also adds some cleanups from #6478 -- mostly to address changes in QSqlQuery fro Qt3 to Qt4 and to workaround the Qt bindings prefix bug in some early Qt4 implementations (bugfixes).

I've tested with DataDirect, XMLTV, EIT and a combined DataDirect/EIT video source, but there may be problems which I did not see with my sources.

NOTE: The plan is to merge the data insertion algorithms, but not until after this data representation changeset has been in the tree a week or two.

git-svn-id: http://svn.mythtv.org/svn/trunk@20656 7dbf422c-18fa-0310-86e9-fd20926502f2