Break out VDPAU playback from VideoOutputXv.
This is largely a straightforward lift and shift and hence there should be no great change in behaviour. The only intentional changes are the removal of the hardware check when checking support (it doesn't really help and just slows everything down) and VDPAU is now 'approved' to display the dummy stream during livetv startup - which means a smoother startup in many cases.

There is also a reasonable amount of code duplication between VideoOutputVDPAU and VideoOutputXv which will be reduced in subsequent commits.


git-svn-id: http://svn.mythtv.org/svn/trunk@20681 7dbf422c-18fa-0310-86e9-fd20926502f2