Closes #6605. Fix 'LiveTV forcing JumpTo 1' endless loop.
If playback is in SwitchToProgram() doing a livetv continuous file transition
and the decoder reaches the of the file after waitingForChange has been set to
true, the decoder will never call the FileChanged() callback to complete the
SwitchToProgram() transition.

However, if there is another file in the tvchain NVP will try to force a
jump to the next file by calling JumpToNext(true, 1). But, since GetDecoder()->
GetWaitForChange() returns true. JumpToProgram() will never be called to do
the transition.

Fix by not testing waitingForChange if a JumpToProgram() is needed.

Thanks to elmojo, and iamlindoro and others for testing.

git-svn-id: http://svn.mythtv.org/svn/trunk@20688 7dbf422c-18fa-0310-86e9-fd20926502f2