Fixes #4688. Update the pause frame after backwards seek.
(2nd attempt, first attempt [20667] was reverted in [20716])

The pause frame is updated before the seek and after the seek
has completed the NVP playback loop sees the pause frame number
that is in the future. That frame number is greater than the
current deleteIter.key() and triggers an unwanted skip forward.

Instead of adding a 5th instance of the refresh the pause frame
idiom, create NuppelVideoPlayer::RefreshPauseFrame(void) and use
that.

git-svn-id: http://svn.mythtv.org/svn/trunk@20722 7dbf422c-18fa-0310-86e9-fd20926502f2