Refs #6612. Fix for NVP video_actually_paused state flutter.
When paused, the video_actually_paused variable was set to false at
the top of NVP::DisplayNormalFrame() then immediately set to true in
DisplayPauseFrame(). Fix by having DisplayNormalFrame() set it based
on paused variable.

Also, do not check PrebufferEnoughFrames() when paused, this could consume
all or most of our prebuffering attempts and then we fail (too) early when
the file becomes available, this is more prevalent in LiveTV.

Thanks to elmojo, stuarta and GreyFoxx for testing.

git-svn-id: http://svn.mythtv.org/svn/trunk@20733 7dbf422c-18fa-0310-86e9-fd20926502f2