Add support for VDPAU inverse telecine, refactor VdpVideoMixer creation and toggling of deinterlacing.
Initial support added for all mixer features with the main video mixer deleted and re-created each time the feature set is changed (features must be enabled when created). This also minimises GPU memory consumption for progressive streams and eliminates the error message currently seen when disabling deinterlacing.

For the time being inverse telecine is enable by adding 'vdpauivtc' to the custom filters section of the display profile (this is temporary). It requires a basic or advanced VDPAU deinterlacer to function and is disabled otherwise. I can't be sure its operating correctly as I don't have any NTSC material that I'm confident is telecined.


git-svn-id: http://svn.mythtv.org/svn/trunk@20741 7dbf422c-18fa-0310-86e9-fd20926502f2