Add support for VDPAU 'skip chroma deinterlace', noise reduction and sharpen/blur.
As for inverse telecine, these are currently enabled via the custom filters setting in the video display profile (yes - this is temporary). This is a comma seperated list of options such as:-

vdpauivtc, vdpauskipchroma, vdpaudenoise=0.5, vdpausharpen=1.0

The valid range for denoise is 0.0 to 1.0 and for sharpen -1.0 to 1.0, with 0.0 disabling both. It should also be noted that, apart from skipchroma, these options will hurt performance - so don't enable them unnecessarily.


git-svn-id: http://svn.mythtv.org/svn/trunk@20742 7dbf422c-18fa-0310-86e9-fd20926502f2