Move 'Watch TV' down in main menu in defaultmenu.
This was discussed on the users mailing list some months ago and in #mythtv-dev. Basically 'Media Library' is almost always what you want when MythTV first starts up so it should be the first thing selected. But moving 'Watch TV' into a sub-menu, as in the DVR menu screens makes it too difficult for people setting up MythTV and for those who actually use Live TV for sports and such to access that functionality. This is a compromise, it is still on the main menu screen, but has been moved to the next to last option (just before 'Utilities / Setup').

git-svn-id: http://svn.mythtv.org/svn/trunk@20761 7dbf422c-18fa-0310-86e9-fd20926502f2