Consolidate video window/display size and positioning calculations into the VideoOutput base class from VideoOutputVDPAU, VideoOutputOpenGL and VideoOutputXv. Also add a new GetDisplayInfo method.
The code will also be used for the windows renderers and possibly directfb (not so sure about ivtv or quartz).

Tested on OS X (OpenGL) and Linux (XVideo, VDPAU and OpenGL) with a fairly comprehensive set of video, window and xrandr settings without any obvious regressions. There are several edge case, pre-existing issues (mostly related to xrandr/DisplayRes) that I'll tidy up in the next week or so.


git-svn-id: http://svn.mythtv.org/svn/trunk@20764 7dbf422c-18fa-0310-86e9-fd20926502f2