Refs #6664. UK EIT fixup. Adds handling of AD, S, SL and W tags in EIT data.
Thanks to Nick for doing a patch. I didn't use it
in the end, rather I fixed the existing m_ukCC
regexp to properly pull out the items of interet
and apply them to the program data.

Been meaning to get this done for ages, and this
ticket spurred me on to sort it out.


git-svn-id: http://svn.mythtv.org/svn/trunk@20768 7dbf422c-18fa-0310-86e9-fd20926502f2