Fixed some misfeatures that crept into ProgLister when the PreviousList functionality was merged in and it was ported to MythUI.
For plPreviouslyRecorded, time ordering is reversed by default.

For plPreviouslyRecorded, call deleteOldRecorded() instead of
EditRecording() when SELECT is pressed.

For plPreviouslyRecorded, use programid as secondary and start time as
tertiary criteria when sorting by title and don't merge entries with
the same title.

Show rsRecorded entires in normal state along with rsWillRecord
entries.

Try to keep the current item at the same relative position when
updating the list.

Do a better job of choosing the new current item after an update,
especially when the old item was deleted.


git-svn-id: http://svn.mythtv.org/svn/trunk@20801 7dbf422c-18fa-0310-86e9-fd20926502f2