On the program details screen save the zoom setting on exit and restore it the next time it's displayed. Closes #6649.
Add a popup menu with options to change the zoom level and switch the details
page shown. This is in addition to the ZOOMIN/ZOOMOUT/INFO keybindings which
have always been there.

Change the background to use a <shape> to make it consistent with changes
elsewhere.

Note: Bumps the binary version for the libmythui API change so plugins need to
be recompiled.



git-svn-id: http://svn.mythtv.org/svn/trunk@20876 7dbf422c-18fa-0310-86e9-fd20926502f2