Fixes #6672. Fixes several HDHomeRun problems in 0.21-fixes.
This is basically just a resync of the hdhomerun library. We now default to using RDP is supported by the firmware, the signal monitoring info is parsed in the library so that the black box model of the HDHomeRun works too, and we automagically reconnect if our socket connection dies for greater reliability.

Note: This is not quite the library on the silicondust website, Mike added raw IP support to avoid breaking existing 0.21 setups and I backported a couple fixes from trunk.

git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@20877 7dbf422c-18fa-0310-86e9-fd20926502f2