Fixes #6240. Look up subtitle files efficiently and allow them to be streamed.
This re-enables external subtitles. Both the frontends and backends need to be upgraded for full functionality, but the protocol is no incremented since it is backward compatible. The subtitle file must be in the same directory as the video file. That directory is checked for subtitle files either by RingBuffer for local files or the backend containing the video file for remotely streamed files.



git-svn-id: http://svn.mythtv.org/svn/trunk@20879 7dbf422c-18fa-0310-86e9-fd20926502f2