Convert MythBrowser to be a regular myth plugin rather than a standalone app.
The main reason for doing this is faster start ups.

I've retained the possibility to use an external browser for now. If you want
to use the built in mythbrowser use 'Internal' for the command. A DB upgrade
will detect if you are currently using the standalone mythbrowser and update
things to use the Internal browser since that's whats 99% of users will want.

Note: I'd recommend doing a 'make distclean' in mythbrowser before doing an
'svn up' to remove any old cruft since a lot of stuff has been renamed and
moved.

Note2: You need to update both MythTV and MythPlugins because the plugin
name change requires changes to the menus.

git-svn-id: http://svn.mythtv.org/svn/trunk@20933 7dbf422c-18fa-0310-86e9-fd20926502f2