Check for recorder errors in TVRec.
In my particular case the dvbstreamhandler was failing due to some device
error and the backend thought the recording was occuring normally.

TVRec will now notice when the recorder is in an error state and will
set the recstatus of the current recording to rsFailed.

Also, failed recordings will automatically be rescheduled if the recording
rule does not check for duplicates in the current recordings.

git-svn-id: http://svn.mythtv.org/svn/trunk@20938 7dbf422c-18fa-0310-86e9-fd20926502f2