Fix incorrect interactive tv video resizing when using full screen OSDs (VDPAU, OpenGL etc) with certain display configurations.
Essentially, the MHEG internals are unaware of the full screen size and only care about the MHEG bounds provided by the VideoOutput method (as any MHEG graphics will occupy the same space as the video). When the display rectangle is a subset of the fullscreen (e.g. 16/9 video on a 16/10 display), video resizing is unnecessarily triggered and the wrong offset is used. The fix is to store the full window size in the interactive tv object and use the correct offset when feeding data back to the VideoOutput and OSD objects.


git-svn-id: http://svn.mythtv.org/svn/trunk@21004 7dbf422c-18fa-0310-86e9-fd20926502f2