Fix segfaults when idle for some Qt/MySql combinations.
Workaround a problem that causes segfaults on both the frontend and backend in
KickDatabase().

Sounds exactly like the problem described in the note for MSqlQuery in
mythdbcon.h.

The solution that seems to work is to ensure that the QSqlQuery is out of
scope when we do the db close() and open().

git-svn-id: http://svn.mythtv.org/svn/trunk@21047 7dbf422c-18fa-0310-86e9-fd20926502f2