Substantial re-write of VideoOutputD3D.
From a user perspective, this adds a full screen OSD, accelerated PiP scaling/colorspace conversion and sharper video rendering (all in line with what is seen with the OpenGL and VDPAU renderers).

Internally, the direct3d specifics have been broken out into a new class (D3D9Context) along with a simple wrapper class (D3D9Video) that acts as the interface between VideoOutputD3D and the context itself. Elsewhere much of the initialisation now uses the base class methods or is consistent with other implementations.

There is no doubt room for improvement, both in terms of features (Color controls, hardware deinterlacing) and performance (I've not profiled it at all!). Tested on XP and Vista with NVidia and ATI cards.


git-svn-id: http://svn.mythtv.org/svn/trunk@21101 7dbf422c-18fa-0310-86e9-fd20926502f2