Allow writing files using myth:// URLs via a writeable RingBuffer and/or RemoteFile.
URL format is as normal, myth://StorageGroup@BackendIP/filename

The StorageGroup is optional and defaults to 'Default'.

The remote file is placed in the directory in the given storage
group with the most free space.  If modifictions are made to support
writing or moving recodings via this, mods will also be made to
the storage group scheduling code to allow this file I/O to be taken
into account.



git-svn-id: http://svn.mythtv.org/svn/trunk@21134 7dbf422c-18fa-0310-86e9-fd20926502f2