Add some more intelligence to the Storage Scheduler.
When using the new "Balance Free Space" method, the Storage Scheduler
will now perform an intermediate step if it can not find a directory
with enough free space to hold the recording.  Previously, the scheduler
would go through the directory list again and put the recording on the
first acceptable directory in the list and hope that something could be
expired from that directory.  The new middle pass will go through the
directory list again looking at what can be expired from each directory to
find out the best place to put the recording based on what recording(s)
may have to be expired to make room for the new recording.

One warning....  The corner case of full drives and multiple recordings
starting at once will still not be handled effectly by this 2nd pass.  The
2nd recording will probably be put on the same filesystem as the first
since "free space" is the deciding factor.  It may be that the "free
space"-only method may work better to use a fuzzy comparison and then
take weight into account which would help balance out multiple recordings
in this scenario.

This patch also does 2 things regarding the default value of the
StorageScheduler setting:

1) The default method on the setup screen is now 'Balanced Free Space'.
2) A DB upgrade inserts a setting into the database to set the method
   to 'Combination' so that users upgrading from previous versions will
   see the same scheduling logic they have been seeing.  Users can
   change the value in mythtv-setup anytime and restart the master
   backend for the new value to take effect.

Closes #5746 using modified patch originally by John R Fulton.



git-svn-id: http://svn.mythtv.org/svn/trunk@21225 7dbf422c-18fa-0310-86e9-fd20926502f2