Allow the PAGELEFT and PAGERIGHT keybindings to cycle through Recording Groups in the Watch Recordings screen.
References #6318.  I'll let Stuart look the UI/theme patch attached to the
ticket since he took this ticket a few months ago.

This patch is by John P Poet, but I added the locking around the Recording
Group cache so that we don't try to access the cache while it is being
rebuilt by PlaybackBox::FillList().



git-svn-id: http://svn.mythtv.org/svn/trunk@21261 7dbf422c-18fa-0310-86e9-fd20926502f2