Rewrite the password handling for entering the setup menu so that we prompt with a non-blocking mythui dialog. In doing this it made sense to remove the hardcoded check on main_settings.xml since that may not exist in all menu themes and instead I've replaced it with a theme defined <password>{pinsettingname}</password>.
As a side benefit this now allows you to define a password for any menu item you want and using any number of different passwords. The location of each password in the settings table is defined by the value of <password> there is no UI for defining the passwords, so it's a case of manually inserting/updating them in the database.



git-svn-id: http://svn.mythtv.org/svn/trunk@21289 7dbf422c-18fa-0310-86e9-fd20926502f2