Only use IDR frames as keyframes for the HD PVR
if I-frames in H264 are used as start points for decoding dangling references
will cause artifacts. They can't be used as keyframes for mythtv's seektables.
IDR-frames or SEI recovery points are valid decoding start points. The last
ffmpeg sync increased the amount of artifacts for the HD PVR. Since we know
the IDR-frame interval of the HD PVR (128 frames), it's safe to use this
standard conforment behaviour.

Ideally we would use it for all H264 source but many DVB H264 transmissions do
not have IDR-frames. The interval of recovery point SEIs has to be tested.



git-svn-id: http://svn.mythtv.org/svn/trunk@21324 7dbf422c-18fa-0310-86e9-fd20926502f2