Allow different but equivalent time zone ID's.
This allows Myth applications to run even if the detected time zone ID (name) differs from the one detected on the master backend, but only if the zones are actually using identical definitions.  Therefore, using a combination of US/Eastern with America/New_York or America/Los_Angeles with US/Pacific or Europe/London with Europe/Belfast for the 2 processes would work.

Different time zones--even if the current time is identical in those differing time zones--are still not allowed as Myth needs to ensure that all local times used by the 2 processes are identical in the past, present, and future.  Therefore, using America/New_York with EST5EDT or US/Pacific with PST8PDT would not work as those combinations specify zones whose times differ in various periods in the past.

Closes #5872


git-svn-id: http://svn.mythtv.org/svn/trunk@21337 7dbf422c-18fa-0310-86e9-fd20926502f2