change data structure holding scanning information from QMap to QList<QPair>
There no real advantage of using a map over a list. In the following commits
the transport_scan_items_t underlaying container will change from QList
to a real double linked list. This is needed for extending the list of
transports to scan during the scan since a reallocation during appending to
a QList invalidates all iterators. We can't use std::list iterators as
keys of the map.

Refs #6756

git-svn-id: http://svn.mythtv.org/svn/trunk@21362 7dbf422c-18fa-0310-86e9-fd20926502f2