Allow for larger NFS blocksizes when comparing total filesystem size to identify filesystems.
This is a much simpler implementation of the fix that went into trunk in [21297] and [21298] , here always allowing up to 1MiB total filesystem size differences, where 1MiB is the current maximum blocksize used for NFS within the Linux kernel.  The more robust fix in trunk could not be backported to -fixes because of the protocol version change.

Refs #6846


git-svn-id: http://svn.mythtv.org/svn/branches/release-0-21-fixes@21368 7dbf422c-18fa-0310-86e9-fd20926502f2