Adds MythVideo metadata image download directly to image storage groups.
If a MythVideo item exists in a storage group, then fanart/banner/screenshot/cover
image downloads will go to those Storage Groups on the host that contains the
video.  If an item is local, then the image downloads will go to the local directories
set for those image types.  This means that if one defines a video storage group on 
a host and intends to download images for videos hosted within, they will need to 
configure image storage groups on that host as well.


git-svn-id: http://svn.mythtv.org/svn/trunk@21391 7dbf422c-18fa-0310-86e9-fd20926502f2