Refs #6862. Refactor ProgramInfo so that it can be used in plugins and split remoteutil.{cpp,h} for the same purpose.
This creates a new class RecordingInfo which is just the ProgramInfo with the ScheduledRecording pointer and those functions which are particular to scheduling recordings as opposed to mythvideo videos and the like.

ProgramInfo still contains the same fields and serializes the same way.

As noted in the ticket, ProgramList/RecordingList are non-ideal as is. But changing those requires code changes. This changeset mostly just moves things around to avoid libmythtv linkage in the plugins and the associated ABI problems.

This changeset changes the binary ABI for plugins so a distclean is recommended and full rebuild of the plugins is required.



git-svn-id: http://svn.mythtv.org/svn/trunk@21408 7dbf422c-18fa-0310-86e9-fd20926502f2